remove and rename a bunch of files
