--- components/d2l-end-of-lesson-image.js
@@ -1,478 +1,0 @@
-import '@polymer/polymer/polymer-legacy.js';
-import { html } from '@polymer/polymer/lib/utils/html-tag.js';
-import { PolymerElement } from '@polymer/polymer/polymer-element.js';
-/*
-@memberOf window.D2L.Polymer.Mixins;
-*/
-class D2LEndOfLessonImage extends PolymerElement {
-	static get template() {
-		return html`
-		<style>
-			.completed .st0{fill:#B9C2D1;}
-			.completed .st1{fill:#D3D9E3;}
-			.completed .st2{fill:#F2F3F5;}
-			.completed .st3{fill:#F9FAFB;}
-			.completed .st4{fill:#E6EAF0;}
-			.completed .st5{fill:#B9C2CF;}
-			.completed .st6{opacity:0.24;fill:none;stroke:#000000;stroke-width:2;stroke-linecap:round;stroke-linejoin:round;stroke-miterlimit:10;}
-			.completed .st7{fill:none;stroke:#565A5C;stroke-width:2;stroke-linecap:round;stroke-linejoin:round;stroke-miterlimit:10;}
-			.completed .st8{opacity:0.24;}
-			.completed .st9{fill:#565A5C;}
-			.completed .st10{fill:#7C8695;}
-			.completed .st11{fill:#7C8694;}
-			.completed .st12{opacity:0.24;stroke:#000000;stroke-width:3;stroke-linejoin:round;stroke-miterlimit:10;}
-			.completed .st13{fill:#B9C2CF;stroke:#565A5C;stroke-width:2;stroke-miterlimit:10;}
-			.completed .st14{fill:#7C8694;stroke:#565A5C;stroke-width:2;stroke-linejoin:round;stroke-miterlimit:10;}
-			.completed .st15{fill:none;}
-			.completed .st16{fill:none;stroke:#565A5C;stroke-width:2;stroke-miterlimit:10;}
-			.completed .st17{fill:#72777A;}
-			.completed .st18{display:none;}
-			.completed .st19{display:inline;fill:#72777A;}
-			.completed .st20{fill:#B9C2CF;stroke:#565A5C;stroke-width:2;stroke-linejoin:round;stroke-miterlimit:10;}
-			.completed .st21{fill:#7C8694;stroke:#565A5C;stroke-width:3;stroke-linejoin:round;stroke-miterlimit:10;}
-			.completed .st22{fill:#E6EAF0;stroke:#565A5C;stroke-width:2;stroke-miterlimit:10;}
-			.completed .st23{fill:none;stroke:#565A5C;stroke-width:2;stroke-linecap:round;stroke-miterlimit:10;}
-			.completed .st24{opacity:0.13;}
-			.completed .st25{fill:#FFFFFF;}
-			.completed .st26{opacity:0.13;fill:#F9FAFB;}
-			.completed .st27{opacity:0.3;}
-			.completed .st28{opacity:0.3;fill:#F9FAFB;}
-			.completed .st29{fill:#D3D9E9;}
-			.completed .st30{opacity:0.47;}
-			.completed .st31{opacity:0.47;fill:#F9FAFB;}
-			.completed .st32{fill:none;stroke:#B9C2D0;stroke-width:2;stroke-linecap:round;stroke-miterlimit:10;}
-			.completed .st33{opacity:0.2;}
-			.completed .st34{opacity:0.35;}
-
-			.incomplete .st0{fill:#72777A;}
-			.incomplete .st1{fill:#F1F2F4;}
-			.incomplete .st2{fill:#D2D8E2;}
-			.incomplete .st3{fill:#D2D8E9;}
-			.incomplete .st4{fill:#B8C1CF;}
-			.incomplete .st5{opacity:0.24;fill:none;stroke:#000000;stroke-width:2;stroke-linecap:round;stroke-linejoin:round;stroke-miterlimit:10;}
-			.incomplete .st6{fill:none;stroke:#565A5C;stroke-width:2;stroke-linecap:round;stroke-linejoin:round;stroke-miterlimit:10;}
-			.incomplete .st7{fill:#B8C1D1;}
-			.incomplete .st8{opacity:0.24;}
-			.incomplete .st9{opacity:0.24;stroke:#000000;stroke-width:3;stroke-linejoin:round;stroke-miterlimit:10;}
-			.incomplete .st10{fill:#E5E9EF;}
-			.incomplete .st11{fill:#565A5C;}
-			.incomplete .st12{fill:#7C8595;}
-			.incomplete .st13{fill:#B8C1CF;stroke:#565A5C;stroke-width:2;stroke-miterlimit:10;}
-			.incomplete .st14{fill:#7C8594;stroke:#565A5C;stroke-width:2;stroke-linejoin:round;stroke-miterlimit:10;}
-			.incomplete .st15{fill:none;}
-			.incomplete .st16{fill:none;stroke:#B8C1D0;stroke-width:2;stroke-linecap:round;stroke-miterlimit:10;}
-			.incomplete .st17{fill:none;stroke:#565A5C;stroke-width:2;stroke-miterlimit:10;}
-			.incomplete .st18{fill:#7C8594;}
-			.incomplete .st19{opacity:0.13;}
-			.incomplete .st20{fill:#F8F9FA;}
-			.incomplete .st21{fill:#FFFFFF;}
-			.incomplete .st22{opacity:0.2;}
-			.incomplete .st23{opacity:0.13;fill:#F8F9FA;}
-			@media(max-width:514px) {
-				svg { width: 100%; }
-			}
-		</style>
-		<template is="dom-if" if="[[completed]]">
-			<?xml version="1.0" encoding="UTF-8"?>
-			<svg width="450px" height="320px" viewBox="0 0 450 320" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
-				<!-- Generator: Sketch 55.1 (78136) - https://sketchapp.com -->
-				<title>Illustration - Moose Puzzle - Complete</title>
-				<desc>Created with Sketch.</desc>
-				<defs>
-					<rect id="path-1" x="11.875" y="13.1387435" width="361.25" height="213.973822" rx="13.75"></rect>
-					<filter x="-2.5%" y="-1.4%" width="103.3%" height="105.6%" filterUnits="objectBoundingBox" id="filter-2">
-						<feOffset dx="-3" dy="3" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feGaussianBlur stdDeviation="1.5" in="shadowOffsetOuter1" result="shadowBlurOuter1"></feGaussianBlur>
-						<feColorMatrix values="0 0 0 0 0.803921569   0 0 0 0 0.835294118   0 0 0 0 0.862745098  0 0 0 1 0" type="matrix" in="shadowBlurOuter1"></feColorMatrix>
-					</filter>
-					<polygon id="path-3" points="211.692105 7.10542736e-15 197.8774 63 0 63 0 2.84217094e-14 211.692105 2.84217094e-14"></polygon>
-					<path d="M195.893688,196.99356 C197.318688,196.99356 198.681188,196.730785 199.931188,196.242775 C200.324938,196.092618 200.706188,195.917435 201.074938,195.723482 L201.074937,211.471204 L202.324938,211.471204 L202.324938,195.710969 L202.318688,195.704712 C202.306188,195.029005 201.756188,194.478429 201.068688,194.478429 C200.868688,194.478429 200.687438,194.522225 200.518688,194.609817 C200.506188,194.616073 200.487438,194.62233 200.474938,194.628586 C199.099938,195.335576 197.543688,195.742251 195.893688,195.742251 C190.381188,195.742251 185.893688,191.250052 185.893688,185.73178 C185.893688,180.213508 190.381188,175.721309 195.893688,175.721309 C197.531188,175.721309 199.081188,176.115471 200.443688,176.822461 C200.462438,176.834974 200.474938,176.84123 200.493688,176.847487 C200.662438,176.941335 200.862438,176.991387 201.068688,176.991387 C201.743688,176.991387 202.293688,176.459581 202.318688,175.790131 L202.324938,175.783874 L202.324938,142.029817 C202.674938,142.255052 203.031188,142.461518 203.406188,142.642958 C204.906188,143.393743 206.599938,143.812932 208.393688,143.812932 C214.599938,143.812932 219.643688,138.763901 219.643688,132.551152 C219.643688,126.34466 214.599938,121.289372 208.393688,121.289372 C206.606188,121.289372 204.912438,121.70856 203.412438,122.459346 C203.037438,122.640785 202.674938,122.847251 202.324938,123.072487 L202.324938,88.9993455 C202.306188,88.3236387 201.756188,87.7855759 201.074938,87.7855759 C200.906188,87.7855759 200.749938,87.8168586 200.599938,87.8794241 C200.593688,87.8794241 200.587438,87.8856806 200.581188,87.8856806 C200.574938,87.8919372 200.568688,87.8919372 200.556188,87.8981937 C199.318688,88.4487696 197.949938,88.7553403 196.518688,88.7553403 C191.006188,88.7553403 186.518688,84.2631414 186.518688,78.7448691 C186.518688,73.2265969 191.006188,68.7343979 196.518688,68.7343979 C197.949938,68.7343979 199.318688,69.0409686 200.556188,69.5915445 C200.706188,69.6541099 200.899938,69.6979058 201.068688,69.6979058 C201.756188,69.6979058 202.324938,69.1348168 202.324938,68.4403403 L202.324938,34.4860733 C202.699938,34.6675131 203.087438,34.8239267 203.481188,34.9615707 C204.624938,35.3619895 205.862438,35.574712 207.143688,35.574712 C213.349938,35.574712 218.393688,30.5256806 218.393688,24.3129319 C218.393688,18.1064398 213.349938,13.0511518 207.143688,13.0511518 C205.862438,13.0511518 204.624938,13.2638743 203.481188,13.6642932 C203.087438,13.8019372 202.699938,13.9583508 202.324938,14.1397906 L202.324938,0 L201.074938,0 L201.074938,14.064712 L201.074938,14.1397906 C201.074938,14.8280105 201.637438,15.3910995 202.324938,15.3910995 C202.524938,15.3910995 202.706188,15.3473037 202.868688,15.2659686 C202.887438,15.259712 202.906188,15.247199 202.924938,15.2409424 C204.212438,14.6340576 205.637438,14.3024607 207.143688,14.3024607 C212.656188,14.3024607 217.143688,18.7946597 217.143688,24.3129319 C217.143688,29.8312042 212.656188,34.3234031 207.143688,34.3234031 C205.637438,34.3234031 204.206188,33.9855497 202.924938,33.3849215 C202.749938,33.2848168 202.537438,33.2347644 202.324938,33.2347644 C201.631188,33.2347644 201.074938,33.7915969 201.074938,34.4860733 L201.074938,34.4985864 L201.074938,68.4528534 C200.693688,68.2839267 200.299938,68.1337696 199.899938,68.0086387 C198.831188,67.6645288 197.699938,67.483089 196.518688,67.483089 C190.312438,67.483089 185.268688,72.538377 185.268688,78.7448691 C185.268688,84.9576178 190.312438,90.0066492 196.518688,90.0066492 C197.699938,90.0066492 198.837438,89.8252094 199.906188,89.4810995 C200.306188,89.3559686 200.693688,89.2058115 201.074938,89.0368848 L201.074938,122.959869 C201.068688,122.991152 201.068688,123.028691 201.068688,123.06623 C201.068688,123.760707 201.631188,124.317539 202.318688,124.317539 C202.568688,124.317539 202.799938,124.248717 202.993688,124.11733 C203.006188,124.11733 203.018688,124.11733 203.024938,124.111073 C204.581188,123.116283 206.418688,122.540681 208.393688,122.540681 C213.906188,122.540681 218.393688,127.03288 218.393688,132.551152 C218.393688,138.069424 213.906188,142.561623 208.393688,142.561623 C206.468688,142.561623 204.681188,142.017304 203.156188,141.072565 C202.937438,140.878613 202.649938,140.759738 202.324938,140.759738 C201.637438,140.759738 201.074938,141.316571 201.074938,142.011047 L201.074938,142.02356 L201.074938,175.740079 C200.706188,175.546126 200.324938,175.377199 199.937438,175.220785 C198.681188,174.732775 197.318688,174.47 195.893688,174.47 C189.687438,174.47 184.643688,179.525288 184.643688,185.73178 C184.643688,191.944529 189.687438,196.99356 195.893688,196.99356 Z" id="path-5"></path>
-					<filter x="-1.4%" y="-0.2%" width="105.7%" height="100.5%" filterUnits="objectBoundingBox" id="filter-6">
-						<feOffset dx="1" dy="0" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.6 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-					<path d="M259.643688,196.99356 C258.218688,196.99356 256.856188,196.730785 255.606188,196.242775 C255.212438,196.092618 254.831188,195.917435 254.462438,195.723482 L254.462438,211.471204 L253.212438,211.471204 L253.212438,195.710969 L253.218688,195.704712 C253.231188,195.029005 253.781188,194.478429 254.468688,194.478429 C254.668688,194.478429 254.849938,194.522225 255.018688,194.609817 C255.031188,194.616073 255.049938,194.62233 255.062438,194.628586 C256.437438,195.335576 257.993688,195.742251 259.643688,195.742251 C265.156188,195.742251 269.643688,191.250052 269.643688,185.73178 C269.643688,180.213508 265.156188,175.721309 259.643688,175.721309 C258.006188,175.721309 256.456188,176.115471 255.093688,176.822461 C255.074938,176.834974 255.062438,176.84123 255.043688,176.847487 C254.874938,176.941335 254.674938,176.991387 254.468688,176.991387 C253.793688,176.991387 253.243688,176.459581 253.218688,175.790131 L253.212438,175.783874 L253.212438,142.029817 C252.862438,142.255052 252.506188,142.461518 252.131188,142.642958 C250.631188,143.393743 248.937438,143.812932 247.143688,143.812932 C240.937438,143.812932 235.893688,138.763901 235.893688,132.551152 C235.893688,126.34466 240.937438,121.289372 247.143688,121.289372 C248.931188,121.289372 250.624938,121.70856 252.124938,122.459346 C252.499938,122.640785 252.862438,122.847251 253.212438,123.072487 L253.212438,88.9993455 C253.231188,88.3236387 253.781188,87.7855759 254.462438,87.7855759 C254.631188,87.7855759 254.787438,87.8168586 254.937438,87.8794241 C254.943688,87.8794241 254.949938,87.8856806 254.956188,87.8856806 C254.962438,87.8919372 254.968688,87.8919372 254.981188,87.8981937 C256.218688,88.4487696 257.587438,88.7553403 259.018688,88.7553403 C264.531188,88.7553403 269.018688,84.2631414 269.018688,78.7448691 C269.018688,73.2265969 264.531188,68.7343979 259.018688,68.7343979 C257.587438,68.7343979 256.218688,69.0409686 254.981188,69.5915445 C254.831188,69.6541099 254.637438,69.6979058 254.468688,69.6979058 C253.781188,69.6979058 253.212438,69.1348168 253.212438,68.4403403 L253.212438,34.4860733 C252.837438,34.6675131 252.449938,34.8239267 252.056188,34.9615707 C250.912438,35.3619895 249.674938,35.574712 248.393688,35.574712 C242.187438,35.574712 237.143688,30.5256806 237.143688,24.3129319 C237.143688,18.1064398 242.187438,13.0511518 248.393688,13.0511518 C249.674938,13.0511518 250.912438,13.2638743 252.056188,13.6642932 C252.449938,13.8019372 252.837438,13.9583508 253.212438,14.1397906 L253.212438,0 L254.462438,0 L254.462438,14.064712 L254.462438,14.1397906 C254.462438,14.8280105 253.899938,15.3910995 253.212438,15.3910995 C253.012438,15.3910995 252.831188,15.3473037 252.668688,15.2659686 C252.649938,15.259712 252.631188,15.247199 252.612438,15.2409424 C251.324938,14.6340576 249.899938,14.3024607 248.393688,14.3024607 C242.881188,14.3024607 238.393688,18.7946597 238.393688,24.3129319 C238.393688,29.8312042 242.881188,34.3234031 248.393688,34.3234031 C249.899938,34.3234031 251.331188,33.9855497 252.612438,33.3849215 C252.787438,33.2848168 252.999938,33.2347644 253.212438,33.2347644 C253.906188,33.2347644 254.462438,33.7915969 254.462438,34.4860733 L254.462438,34.4985864 L254.462438,68.4528534 C254.843688,68.2839267 255.237438,68.1337696 255.637438,68.0086387 C256.706188,67.6645288 257.837438,67.483089 259.018688,67.483089 C265.224938,67.483089 270.268688,72.538377 270.268688,78.7448691 C270.268688,84.9576178 265.224938,90.0066492 259.018688,90.0066492 C257.837438,90.0066492 256.699938,89.8252094 255.631188,89.4810995 C255.231188,89.3559686 254.843688,89.2058115 254.462438,89.0368848 L254.462438,122.959869 C254.468688,122.991152 254.468688,123.028691 254.468688,123.06623 C254.468688,123.760707 253.906188,124.317539 253.218688,124.317539 C252.968688,124.317539 252.737438,124.248717 252.543688,124.11733 C252.531188,124.11733 252.518688,124.11733 252.512438,124.111073 C250.956188,123.116283 249.118688,122.540681 247.143688,122.540681 C241.631188,122.540681 237.143688,127.03288 237.143688,132.551152 C237.143688,138.069424 241.631188,142.561623 247.143688,142.561623 C249.068688,142.561623 250.856188,142.017304 252.381188,141.072565 C252.599938,140.878613 252.887438,140.759738 253.212438,140.759738 C253.899938,140.759738 254.462438,141.316571 254.462438,142.011047 L254.462438,142.02356 L254.462438,175.740079 C254.831188,175.546126 255.212438,175.377199 255.599938,175.220785 C256.856188,174.732775 258.218688,174.47 259.643688,174.47 C265.849938,174.47 270.893688,179.525288 270.893688,185.73178 C270.893688,191.944529 265.849938,196.99356 259.643688,196.99356 Z" id="path-7"></path>
-					<filter x="-1.4%" y="-0.2%" width="105.7%" height="100.5%" filterUnits="objectBoundingBox" id="filter-8">
-						<feOffset dx="1" dy="0" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.6 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-					<path d="M299.643688,196.99356 C301.068688,196.99356 302.431188,196.730785 303.681188,196.242775 C304.074938,196.092618 304.456188,195.917435 304.824938,195.723482 L304.824938,211.471204 L306.074938,211.471204 L306.074938,195.710969 L306.068688,195.704712 C306.056188,195.029005 305.506188,194.478429 304.818688,194.478429 C304.618688,194.478429 304.437438,194.522225 304.268688,194.609817 C304.256188,194.616073 304.237438,194.62233 304.224938,194.628586 C302.849938,195.335576 301.293688,195.742251 299.643688,195.742251 C294.131188,195.742251 289.643688,191.250052 289.643688,185.73178 C289.643688,180.213508 294.131188,175.721309 299.643688,175.721309 C301.281188,175.721309 302.831188,176.115471 304.193688,176.822461 C304.212438,176.834974 304.224938,176.84123 304.243688,176.847487 C304.412438,176.941335 304.612438,176.991387 304.818688,176.991387 C305.493688,176.991387 306.043688,176.459581 306.068688,175.790131 L306.074938,175.783874 L306.074938,142.029817 C306.424938,142.255052 306.781188,142.461518 307.156188,142.642958 C308.656188,143.393743 310.349938,143.812932 312.143688,143.812932 C318.349938,143.812932 323.393688,138.763901 323.393688,132.551152 C323.393688,126.34466 318.349938,121.289372 312.143688,121.289372 C310.356188,121.289372 308.662438,121.70856 307.162438,122.459346 C306.787438,122.640785 306.424938,122.847251 306.074938,123.072487 L306.074938,88.9993455 C306.056188,88.3236387 305.506188,87.7855759 304.824938,87.7855759 C304.656188,87.7855759 304.499938,87.8168586 304.349938,87.8794241 C304.343688,87.8794241 304.337438,87.8856806 304.331188,87.8856806 C304.324938,87.8919372 304.318688,87.8919372 304.306188,87.8981937 C303.068688,88.4487696 301.699938,88.7553403 300.268688,88.7553403 C294.756188,88.7553403 290.268688,84.2631414 290.268688,78.7448691 C290.268688,73.2265969 294.756188,68.7343979 300.268688,68.7343979 C301.699938,68.7343979 303.068688,69.0409686 304.306188,69.5915445 C304.456188,69.6541099 304.649938,69.6979058 304.818688,69.6979058 C305.506188,69.6979058 306.074938,69.1348168 306.074938,68.4403403 L306.074938,34.4860733 C306.449938,34.6675131 306.837438,34.8239267 307.231188,34.9615707 C308.374938,35.3619895 309.612438,35.574712 310.893688,35.574712 C317.099938,35.574712 322.143688,30.5256806 322.143688,24.3129319 C322.143688,18.1064398 317.099938,13.0511518 310.893688,13.0511518 C309.612438,13.0511518 308.374938,13.2638743 307.231188,13.6642932 C306.837438,13.8019372 306.449938,13.9583508 306.074938,14.1397906 L306.074938,0 L304.824938,0 L304.824938,14.064712 L304.824938,14.1397906 C304.824938,14.8280105 305.387438,15.3910995 306.074938,15.3910995 C306.274938,15.3910995 306.456188,15.3473037 306.618688,15.2659686 C306.637438,15.259712 306.656188,15.247199 306.674938,15.2409424 C307.962438,14.6340576 309.387438,14.3024607 310.893688,14.3024607 C316.406188,14.3024607 320.893688,18.7946597 320.893688,24.3129319 C320.893688,29.8312042 316.406188,34.3234031 310.893688,34.3234031 C309.387438,34.3234031 307.956188,33.9855497 306.674938,33.3849215 C306.499938,33.2848168 306.287438,33.2347644 306.074938,33.2347644 C305.381188,33.2347644 304.824938,33.7915969 304.824938,34.4860733 L304.824938,34.4985864 L304.824938,68.4528534 C304.443688,68.2839267 304.049938,68.1337696 303.649938,68.0086387 C302.581188,67.6645288 301.449938,67.483089 300.268688,67.483089 C294.062438,67.483089 289.018688,72.538377 289.018688,78.7448691 C289.018688,84.9576178 294.062438,90.0066492 300.268688,90.0066492 C301.449938,90.0066492 302.587438,89.8252094 303.656188,89.4810995 C304.056188,89.3559686 304.443688,89.2058115 304.824938,89.0368848 L304.824938,122.959869 C304.818688,122.991152 304.818688,123.028691 304.818688,123.06623 C304.818688,123.760707 305.381188,124.317539 306.068688,124.317539 C306.318688,124.317539 306.549938,124.248717 306.743688,124.11733 C306.756188,124.11733 306.768688,124.11733 306.774938,124.111073 C308.331188,123.116283 310.168688,122.540681 312.143688,122.540681 C317.656188,122.540681 322.143688,127.03288 322.143688,132.551152 C322.143688,138.069424 317.656188,142.561623 312.143688,142.561623 C310.218688,142.561623 308.431188,142.017304 306.906188,141.072565 C306.687438,140.878613 306.399938,140.759738 306.074938,140.759738 C305.387438,140.759738 304.824938,141.316571 304.824938,142.011047 L304.824938,142.02356 L304.824938,175.740079 C304.456188,175.546126 304.074938,175.377199 303.687438,175.220785 C302.431188,174.732775 301.068688,174.47 299.643688,174.47 C293.437438,174.47 288.393688,179.525288 288.393688,185.73178 C288.393688,191.944529 293.437438,196.99356 299.643688,196.99356 Z" id="path-9"></path>
-					<filter x="-1.4%" y="-0.2%" width="105.7%" height="100.5%" filterUnits="objectBoundingBox" id="filter-10">
-						<feOffset dx="1" dy="0" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.6 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-					<path d="M157.143688,196.99356 C155.718688,196.99356 154.356188,196.730785 153.106188,196.242775 C152.712438,196.092618 152.331188,195.917435 151.962438,195.723482 L151.962438,211.471204 L150.712438,211.471204 L150.712438,195.710969 L150.718688,195.704712 C150.731188,195.029005 151.281188,194.478429 151.968688,194.478429 C152.168688,194.478429 152.349938,194.522225 152.518688,194.609817 C152.531188,194.616073 152.549938,194.62233 152.562438,194.628586 C153.937438,195.335576 155.493688,195.742251 157.143688,195.742251 C162.656188,195.742251 167.143688,191.250052 167.143688,185.73178 C167.143688,180.213508 162.656188,175.721309 157.143688,175.721309 C155.506188,175.721309 153.956188,176.115471 152.593688,176.822461 C152.574938,176.834974 152.562438,176.84123 152.543688,176.847487 C152.374938,176.941335 152.174938,176.991387 151.968688,176.991387 C151.293688,176.991387 150.743688,176.459581 150.718688,175.790131 L150.712438,175.783874 L150.712438,142.029817 C150.362438,142.255052 150.006188,142.461518 149.631188,142.642958 C148.131188,143.393743 146.437438,143.812932 144.643688,143.812932 C138.437438,143.812932 133.393688,138.763901 133.393688,132.551152 C133.393688,126.34466 138.437438,121.289372 144.643688,121.289372 C146.431188,121.289372 148.124938,121.70856 149.624938,122.459346 C149.999938,122.640785 150.362438,122.847251 150.712438,123.072487 L150.712438,88.9993455 C150.731188,88.3236387 151.281188,87.7855759 151.962438,87.7855759 C152.131188,87.7855759 152.287438,87.8168586 152.437438,87.8794241 C152.443688,87.8794241 152.449938,87.8856806 152.456188,87.8856806 C152.462438,87.8919372 152.468688,87.8919372 152.481188,87.8981937 C153.718688,88.4487696 155.087438,88.7553403 156.518688,88.7553403 C162.031188,88.7553403 166.518688,84.2631414 166.518688,78.7448691 C166.518688,73.2265969 162.031188,68.7343979 156.518688,68.7343979 C155.087438,68.7343979 153.718688,69.0409686 152.481188,69.5915445 C152.331188,69.6541099 152.137438,69.6979058 151.968688,69.6979058 C151.281188,69.6979058 150.712438,69.1348168 150.712438,68.4403403 L150.712438,34.4860733 C150.337438,34.6675131 149.949938,34.8239267 149.556188,34.9615707 C148.412438,35.3619895 147.174938,35.574712 145.893688,35.574712 C139.687438,35.574712 134.643688,30.5256806 134.643688,24.3129319 C134.643688,18.1064398 139.687438,13.0511518 145.893688,13.0511518 C147.174938,13.0511518 148.412438,13.2638743 149.556188,13.6642932 C149.949938,13.8019372 150.337438,13.9583508 150.712438,14.1397906 L150.712438,0 L151.962438,0 L151.962438,14.064712 L151.962438,14.1397906 C151.962438,14.8280105 151.399938,15.3910995 150.712438,15.3910995 C150.512438,15.3910995 150.331188,15.3473037 150.168688,15.2659686 C150.149938,15.259712 150.131188,15.247199 150.112438,15.2409424 C148.824938,14.6340576 147.399938,14.3024607 145.893688,14.3024607 C140.381188,14.3024607 135.893688,18.7946597 135.893688,24.3129319 C135.893688,29.8312042 140.381188,34.3234031 145.893688,34.3234031 C147.399938,34.3234031 148.831188,33.9855497 150.112438,33.3849215 C150.287438,33.2848168 150.499938,33.2347644 150.712438,33.2347644 C151.406188,33.2347644 151.962438,33.7915969 151.962438,34.4860733 L151.962438,34.4985864 L151.962438,68.4528534 C152.343688,68.2839267 152.737438,68.1337696 153.137438,68.0086387 C154.206188,67.6645288 155.337438,67.483089 156.518688,67.483089 C162.724938,67.483089 167.768688,72.538377 167.768688,78.7448691 C167.768688,84.9576178 162.724938,90.0066492 156.518688,90.0066492 C155.337438,90.0066492 154.199938,89.8252094 153.131188,89.4810995 C152.731188,89.3559686 152.343688,89.2058115 151.962438,89.0368848 L151.962438,122.959869 C151.968688,122.991152 151.968688,123.028691 151.968688,123.06623 C151.968688,123.760707 151.406188,124.317539 150.718688,124.317539 C150.468688,124.317539 150.237438,124.248717 150.043688,124.11733 C150.031188,124.11733 150.018688,124.11733 150.012438,124.111073 C148.456188,123.116283 146.618688,122.540681 144.643688,122.540681 C139.131188,122.540681 134.643688,127.03288 134.643688,132.551152 C134.643688,138.069424 139.131188,142.561623 144.643688,142.561623 C146.568688,142.561623 148.356188,142.017304 149.881188,141.072565 C150.099938,140.878613 150.387438,140.759738 150.712438,140.759738 C151.399938,140.759738 151.962438,141.316571 151.962438,142.011047 L151.962438,142.02356 L151.962438,175.740079 C152.331188,175.546126 152.712438,175.377199 153.099938,175.220785 C154.356188,174.732775 155.718688,174.47 157.143688,174.47 C163.349938,174.47 168.393688,179.525288 168.393688,185.73178 C168.393688,191.944529 163.349938,196.99356 157.143688,196.99356 Z" id="path-11"></path>
-					<filter x="-1.4%" y="-0.2%" width="105.7%" height="100.5%" filterUnits="objectBoundingBox" id="filter-12">
-						<feOffset dx="1" dy="0" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.6 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-					<path d="M93.3936875,196.99356 C94.8186875,196.99356 96.1811875,196.730785 97.4311875,196.242775 C97.8249375,196.092618 98.2061875,195.917435 98.5749375,195.723482 L98.5724375,211.471204 L99.8249375,211.471204 L99.8249375,195.710969 L99.8186875,195.704712 C99.8061875,195.029005 99.2561875,194.478429 98.5686875,194.478429 C98.3686875,194.478429 98.1874375,194.522225 98.0186875,194.609817 C98.0061875,194.616073 97.9874375,194.62233 97.9749375,194.628586 C96.5999375,195.335576 95.0436875,195.742251 93.3936875,195.742251 C87.8811875,195.742251 83.3936875,191.250052 83.3936875,185.73178 C83.3936875,180.213508 87.8811875,175.721309 93.3936875,175.721309 C95.0311875,175.721309 96.5811875,176.115471 97.9436875,176.822461 C97.9624375,176.834974 97.9749375,176.84123 97.9936875,176.847487 C98.1624375,176.941335 98.3624375,176.991387 98.5686875,176.991387 C99.2436875,176.991387 99.7936875,176.459581 99.8186875,175.790131 L99.8249375,175.783874 L99.8249375,142.029817 C100.174938,142.255052 100.531188,142.461518 100.906188,142.642958 C102.406188,143.393743 104.099938,143.812932 105.893688,143.812932 C112.099938,143.812932 117.143688,138.763901 117.143688,132.551152 C117.143688,126.34466 112.099938,121.289372 105.893688,121.289372 C104.106188,121.289372 102.412438,121.70856 100.912438,122.459346 C100.537438,122.640785 100.174938,122.847251 99.8249375,123.072487 L99.8249375,88.9993455 C99.8061875,88.3236387 99.2561875,87.7855759 98.5749375,87.7855759 C98.4061875,87.7855759 98.2499375,87.8168586 98.0999375,87.8794241 C98.0936875,87.8794241 98.0874375,87.8856806 98.0811875,87.8856806 C98.0749375,87.8919372 98.0686875,87.8919372 98.0561875,87.8981937 C96.8186875,88.4487696 95.4499375,88.7553403 94.0186875,88.7553403 C88.5061875,88.7553403 84.0186875,84.2631414 84.0186875,78.7448691 C84.0186875,73.2265969 88.5061875,68.7343979 94.0186875,68.7343979 C95.4499375,68.7343979 96.8186875,69.0409686 98.0561875,69.5915445 C98.2061875,69.6541099 98.3999375,69.6979058 98.5686875,69.6979058 C99.2561875,69.6979058 99.8249375,69.1348168 99.8249375,68.4403403 L99.8249375,34.4860733 C100.199938,34.6675131 100.587438,34.8239267 100.981188,34.9615707 C102.124938,35.3619895 103.362438,35.574712 104.643688,35.574712 C110.849938,35.574712 115.893688,30.5256806 115.893688,24.3129319 C115.893688,18.1064398 110.849938,13.0511518 104.643688,13.0511518 C103.362438,13.0511518 102.124938,13.2638743 100.981188,13.6642932 C100.587438,13.8019372 100.199938,13.9583508 99.8249375,14.1397906 L99.8249375,0 L98.5749375,0 L98.5749375,14.064712 L98.5749375,14.1397906 C98.5749375,14.8280105 99.1374375,15.3910995 99.8249375,15.3910995 C100.024938,15.3910995 100.206188,15.3473037 100.368688,15.2659686 C100.387438,15.259712 100.406188,15.247199 100.424938,15.2409424 C101.712438,14.6340576 103.137438,14.3024607 104.643688,14.3024607 C110.156188,14.3024607 114.643688,18.7946597 114.643688,24.3129319 C114.643688,29.8312042 110.156188,34.3234031 104.643688,34.3234031 C103.137438,34.3234031 101.706188,33.9855497 100.424938,33.3849215 C100.249938,33.2848168 100.037438,33.2347644 99.8249375,33.2347644 C99.1311875,33.2347644 98.5749375,33.7915969 98.5749375,34.4860733 L98.5749375,34.4985864 L98.5749375,68.4528534 C98.1936875,68.2839267 97.7999375,68.1337696 97.3999375,68.0086387 C96.3311875,67.6645288 95.1999375,67.483089 94.0186875,67.483089 C87.8124375,67.483089 82.7686875,72.538377 82.7686875,78.7448691 C82.7686875,84.9576178 87.8124375,90.0066492 94.0186875,90.0066492 C95.1999375,90.0066492 96.3374375,89.8252094 97.4061875,89.4810995 C97.8061875,89.3559686 98.1936875,89.2058115 98.5749375,89.0368848 L98.5749375,122.959869 C98.5686875,122.991152 98.5686875,123.028691 98.5686875,123.06623 C98.5686875,123.760707 99.1311875,124.317539 99.8186875,124.317539 C100.068688,124.317539 100.299938,124.248717 100.493688,124.11733 C100.506188,124.11733 100.518688,124.11733 100.524938,124.111073 C102.081188,123.116283 103.918688,122.540681 105.893688,122.540681 C111.406188,122.540681 115.893688,127.03288 115.893688,132.551152 C115.893688,138.069424 111.406188,142.561623 105.893688,142.561623 C103.968688,142.561623 102.181188,142.017304 100.656188,141.072565 C100.437438,140.878613 100.149938,140.759738 99.8249375,140.759738 C99.1374375,140.759738 98.5749375,141.316571 98.5749375,142.011047 L98.5749375,142.02356 L98.5749375,175.740079 C98.2061875,175.546126 97.8249375,175.377199 97.4374375,175.220785 C96.1811875,174.732775 94.8186875,174.47 93.3936875,174.47 C87.1874375,174.47 82.1436875,179.525288 82.1436875,185.73178 C82.1436875,191.944529 87.1874375,196.99356 93.3936875,196.99356 Z" id="path-13"></path>
-					<filter x="-1.4%" y="-0.2%" width="105.7%" height="100.5%" filterUnits="objectBoundingBox" id="filter-14">
-						<feOffset dx="1" dy="0" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.6 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-					<path d="M55.2686875,196.99356 C53.8436875,196.99356 52.4811875,196.730785 51.2311875,196.242775 C50.8374375,196.092618 50.4561875,195.917435 50.0874375,195.723482 L50.0874375,211.471204 L48.8374375,211.471204 L48.8374375,195.710969 L48.8436875,195.704712 C48.8561875,195.029005 49.4061875,194.478429 50.0936875,194.478429 C50.2936875,194.478429 50.4749375,194.522225 50.6436875,194.609817 C50.6561875,194.616073 50.6749375,194.62233 50.6874375,194.628586 C52.0624375,195.335576 53.6186875,195.742251 55.2686875,195.742251 C60.7811875,195.742251 65.2686875,191.250052 65.2686875,185.73178 C65.2686875,180.213508 60.7811875,175.721309 55.2686875,175.721309 C53.6311875,175.721309 52.0811875,176.115471 50.7186875,176.822461 C50.6999375,176.834974 50.6874375,176.84123 50.6686875,176.847487 C50.4999375,176.941335 50.2999375,176.991387 50.0936875,176.991387 C49.4186875,176.991387 48.8686875,176.459581 48.8436875,175.790131 L48.8374375,175.783874 L48.8374375,142.029817 C48.4874375,142.255052 48.1311875,142.461518 47.7561875,142.642958 C46.2561875,143.393743 44.5624375,143.812932 42.7686875,143.812932 C36.5624375,143.812932 31.5186875,138.763901 31.5186875,132.551152 C31.5186875,126.34466 36.5624375,121.289372 42.7686875,121.289372 C44.5561875,121.289372 46.2499375,121.70856 47.7499375,122.459346 C48.1249375,122.640785 48.4874375,122.847251 48.8374375,123.072487 L48.8374375,88.9993455 C48.8561875,88.3236387 49.4061875,87.7855759 50.0874375,87.7855759 C50.2561875,87.7855759 50.4124375,87.8168586 50.5624375,87.8794241 C50.5686875,87.8794241 50.5749375,87.8856806 50.5811875,87.8856806 C50.5874375,87.8919372 50.5936875,87.8919372 50.6061875,87.8981937 C51.8436875,88.4487696 53.2124375,88.7553403 54.6436875,88.7553403 C60.1561875,88.7553403 64.6436875,84.2631414 64.6436875,78.7448691 C64.6436875,73.2265969 60.1561875,68.7343979 54.6436875,68.7343979 C53.2124375,68.7343979 51.8436875,69.0409686 50.6061875,69.5915445 C50.4561875,69.6541099 50.2624375,69.6979058 50.0936875,69.6979058 C49.4061875,69.6979058 48.8374375,69.1348168 48.8374375,68.4403403 L48.8374375,34.4860733 C48.4624375,34.6675131 48.0749375,34.8239267 47.6811875,34.9615707 C46.5374375,35.3619895 45.2999375,35.574712 44.0186875,35.574712 C37.8124375,35.574712 32.7686875,30.5256806 32.7686875,24.3129319 C32.7686875,18.1064398 37.8124375,13.0511518 44.0186875,13.0511518 C45.2999375,13.0511518 46.5374375,13.2638743 47.6811875,13.6642932 C48.0749375,13.8019372 48.4624375,13.9583508 48.8374375,14.1397906 L48.8374375,0 L50.0874375,0 L50.0874375,14.064712 L50.0874375,14.1397906 C50.0874375,14.8280105 49.5249375,15.3910995 48.8374375,15.3910995 C48.6374375,15.3910995 48.4561875,15.3473037 48.2936875,15.2659686 C48.2749375,15.259712 48.2561875,15.247199 48.2374375,15.2409424 C46.9499375,14.6340576 45.5249375,14.3024607 44.0186875,14.3024607 C38.5061875,14.3024607 34.0186875,18.7946597 34.0186875,24.3129319 C34.0186875,29.8312042 38.5061875,34.3234031 44.0186875,34.3234031 C45.5249375,34.3234031 46.9561875,33.9855497 48.2374375,33.3849215 C48.4124375,33.2848168 48.6249375,33.2347644 48.8374375,33.2347644 C49.5311875,33.2347644 50.0874375,33.7915969 50.0874375,34.4860733 L50.0874375,34.4985864 L50.0874375,68.4528534 C50.4686875,68.2839267 50.8624375,68.1337696 51.2624375,68.0086387 C52.3311875,67.6645288 53.4624375,67.483089 54.6436875,67.483089 C60.8499375,67.483089 65.8936875,72.538377 65.8936875,78.7448691 C65.8936875,84.9576178 60.8499375,90.0066492 54.6436875,90.0066492 C53.4624375,90.0066492 52.3249375,89.8252094 51.2561875,89.4810995 C50.8561875,89.3559686 50.4686875,89.2058115 50.0874375,89.0368848 L50.0874375,122.959869 C50.0936875,122.991152 50.0936875,123.028691 50.0936875,123.06623 C50.0936875,123.760707 49.5311875,124.317539 48.8436875,124.317539 C48.5936875,124.317539 48.3624375,124.248717 48.1686875,124.11733 C48.1561875,124.11733 48.1436875,124.11733 48.1374375,124.111073 C46.5811875,123.116283 44.7436875,122.540681 42.7686875,122.540681 C37.2561875,122.540681 32.7686875,127.03288 32.7686875,132.551152 C32.7686875,138.069424 37.2561875,142.561623 42.7686875,142.561623 C44.6936875,142.561623 46.4811875,142.017304 48.0061875,141.072565 C48.2249375,140.878613 48.5124375,140.759738 48.8374375,140.759738 C49.5249375,140.759738 50.0874375,141.316571 50.0874375,142.011047 L50.0874375,142.02356 L50.0874375,175.740079 C50.4561875,175.546126 50.8374375,175.377199 51.2249375,175.220785 C52.4811875,174.732775 53.8436875,174.47 55.2686875,174.47 C61.4749375,174.47 66.5186875,179.525288 66.5186875,185.73178 C66.5186875,191.944529 61.4749375,196.99356 55.2686875,196.99356 Z" id="path-15"></path>
-					<filter x="-1.4%" y="-0.2%" width="105.7%" height="100.5%" filterUnits="objectBoundingBox" id="filter-16">
-						<feOffset dx="1" dy="0" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.6 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-					<path d="M358.9375,159.378839 L340.7375,159.378839 C340.9625,159.729206 341.16875,160.085829 341.35,160.461221 C342.1,161.962792 342.51875,163.658316 342.51875,165.453944 C342.51875,171.666693 337.475,176.715724 331.26875,176.715724 C325.06875,176.715724 320.01875,171.666693 320.01875,165.453944 C320.01875,163.664572 320.4375,161.969049 321.1875,160.467478 C321.36875,160.092085 321.575,159.729206 321.8,159.378839 L289.3625,159.378839 C288.68125,159.36007 288.13125,158.809494 288.13125,158.121274 C288.13125,157.921064 288.175,157.739625 288.2625,157.570698 C288.26875,157.558185 288.275,157.539415 288.28125,157.526902 C288.9875,156.150462 289.39375,154.592583 289.39375,152.940855 C289.39375,147.422583 284.90625,142.930384 279.39375,142.930384 C273.88125,142.930384 269.39375,147.422583 269.39375,152.940855 C269.39375,154.58007 269.7875,156.131693 270.49375,157.495619 C270.50625,157.514389 270.5125,157.526902 270.51875,157.545672 C270.6125,157.714598 270.6625,157.914808 270.6625,158.121274 C270.6625,158.796981 270.13125,159.347557 269.4625,159.372583 L237.6125,159.378839 C237.8375,159.729206 238.04375,160.085829 238.225,160.461221 C238.975,161.962792 239.39375,163.658316 239.39375,165.453944 C239.39375,171.666693 234.35,176.715724 228.14375,176.715724 C221.94375,176.715724 216.89375,171.666693 216.89375,165.453944 C216.89375,163.664572 217.3125,161.969049 218.0625,160.467478 C218.24375,160.092085 218.45,159.729206 218.675,159.378839 L186.8625,159.378839 C186.18125,159.36007 185.63125,158.809494 185.63125,158.121274 C185.63125,157.921064 185.675,157.739625 185.7625,157.570698 C185.76875,157.558185 185.775,157.539415 185.78125,157.526902 C186.4875,156.150462 186.89375,154.592583 186.89375,152.940855 C186.89375,147.422583 182.40625,142.930384 176.89375,142.930384 C171.38125,142.930384 166.89375,147.422583 166.89375,152.940855 C166.89375,154.58007 167.2875,156.131693 167.99375,157.495619 C168.00625,157.514389 168.0125,157.526902 168.01875,157.545672 C168.1125,157.714598 168.1625,157.914808 168.1625,158.121274 C168.1625,158.796981 167.63125,159.347557 166.9625,159.372583 L135.7375,159.378839 C135.9625,159.729206 136.16875,160.085829 136.35,160.461221 C137.1,161.962792 137.51875,163.658316 137.51875,165.453944 C137.51875,171.666693 132.475,176.715724 126.26875,176.715724 C120.06875,176.715724 115.01875,171.666693 115.01875,165.453944 C115.01875,163.664572 115.4375,161.969049 116.1875,160.467478 C116.36875,160.092085 116.575,159.729206 116.8,159.378839 L84.0125,159.378839 C83.3375,159.36007 82.8,158.809494 82.8,158.12753 C82.8,157.958604 82.83125,157.80219 82.89375,157.652033 C82.89375,157.645776 82.9,157.63952 82.9,157.633263 C82.90625,157.627007 82.90625,157.62075 82.9125,157.608237 C83.4625,156.369441 83.76875,154.999258 83.76875,153.566509 C83.76875,148.048237 79.28125,143.556038 73.76875,143.556038 C68.25625,143.556038 63.76875,148.048237 63.76875,153.566509 C63.76875,154.999258 64.075,156.369441 64.625,157.608237 C64.6875,157.758394 64.73125,157.952347 64.73125,158.121274 C64.73125,158.809494 64.16875,159.378839 63.475,159.378839 L33.93125,159.378839 C34.1125,159.754232 34.26875,160.142138 34.40625,160.5363 C34.80625,161.681248 35.01875,162.920043 35.01875,164.202635 C35.01875,170.415384 29.975,175.464415 23.76875,175.464415 C17.56875,175.464415 12.51875,170.415384 12.51875,164.202635 C12.51875,162.920043 12.73125,161.681248 13.13125,160.5363 C13.26875,160.142138 13.425,159.754232 13.60625,159.378839 L0.206875,159.378839 L0.206875,158.12753 L13.53125,158.12753 L13.60625,158.12753 C14.29375,158.12753 14.85625,158.690619 14.85625,159.378839 C14.85625,159.579049 14.8125,159.760488 14.73125,159.923159 C14.725,159.941928 14.7125,159.960698 14.70625,159.979468 C14.1,161.268316 13.76875,162.694808 13.76875,164.202635 C13.76875,169.720907 18.25625,174.213106 23.76875,174.213106 C29.28125,174.213106 33.76875,169.720907 33.76875,164.202635 C33.76875,162.694808 33.43125,161.262059 32.83125,159.979468 C32.73125,159.804284 32.68125,159.591562 32.68125,159.378839 C32.68125,158.684363 33.2375,158.12753 33.93125,158.12753 L33.94375,158.12753 L63.4875,158.12753 C63.31875,157.745881 63.16875,157.351719 63.04375,156.9513 C62.7,155.881431 62.51875,154.748996 62.51875,153.566509 C62.51875,147.353761 67.56875,142.304729 73.76875,142.304729 C79.975,142.304729 85.01875,147.353761 85.01875,153.566509 C85.01875,154.748996 84.8375,155.887687 84.49375,156.957557 C84.36875,157.357975 84.21875,157.745881 84.05,158.12753 L116.6875,158.12753 C116.71875,158.121274 116.75625,158.121274 116.79375,158.121274 C117.4875,158.121274 118.04375,158.684363 118.04375,159.372583 C118.04375,159.622845 117.975,159.854337 117.84375,160.04829 C117.84375,160.060803 117.84375,160.073316 117.8375,160.079572 C116.84375,161.637452 116.26875,163.476876 116.26875,165.453944 C116.26875,170.972216 120.75625,175.464415 126.26875,175.464415 C131.78125,175.464415 136.26875,170.972216 136.26875,165.453944 C136.26875,163.526928 135.725,161.737557 134.78125,160.21096 C134.5875,159.991981 134.46875,159.70418 134.46875,159.378839 C134.46875,158.690619 135.025,158.12753 135.71875,158.12753 L135.73125,158.12753 L166.9125,158.12753 C166.71875,157.758394 166.55,157.376745 166.39375,156.988839 C165.90625,155.731274 165.64375,154.367347 165.64375,152.940855 C165.64375,146.728106 170.69375,141.679075 176.89375,141.679075 C183.1,141.679075 188.14375,146.728106 188.14375,152.940855 C188.14375,154.367347 187.88125,155.731274 187.39375,156.982583 C187.24375,157.376745 187.06875,157.758394 186.875,158.12753 L218.5625,158.12753 C218.59375,158.121274 218.63125,158.121274 218.66875,158.121274 C219.3625,158.121274 219.91875,158.684363 219.91875,159.372583 C219.91875,159.622845 219.85,159.854337 219.71875,160.04829 C219.71875,160.060803 219.71875,160.073316 219.7125,160.079572 C218.71875,161.637452 218.14375,163.476876 218.14375,165.453944 C218.14375,170.972216 222.63125,175.464415 228.14375,175.464415 C233.65625,175.464415 238.14375,170.972216 238.14375,165.453944 C238.14375,163.526928 237.6,161.737557 236.65625,160.21096 C236.4625,159.991981 236.34375,159.70418 236.34375,159.378839 C236.34375,158.690619 236.9,158.12753 237.59375,158.12753 L237.60625,158.12753 L269.4125,158.12753 C269.21875,157.758394 269.05,157.376745 268.89375,156.988839 C268.40625,155.731274 268.14375,154.367347 268.14375,152.940855 C268.14375,146.728106 273.19375,141.679075 279.39375,141.679075 C285.6,141.679075 290.64375,146.728106 290.64375,152.940855 C290.64375,154.367347 290.38125,155.731274 289.89375,156.982583 C289.74375,157.376745 289.56875,157.758394 289.375,158.12753 L321.6875,158.12753 C321.71875,158.121274 321.75625,158.121274 321.79375,158.121274 C322.4875,158.121274 323.04375,158.684363 323.04375,159.372583 C323.04375,159.622845 322.975,159.854337 322.84375,160.04829 C322.84375,160.060803 322.84375,160.073316 322.8375,160.079572 C321.84375,161.637452 321.26875,163.476876 321.26875,165.453944 C321.26875,170.972216 325.75625,175.464415 331.26875,175.464415 C336.78125,175.464415 341.26875,170.972216 341.26875,165.453944 C341.26875,163.526928 340.725,161.737557 339.78125,160.21096 C339.5875,159.991981 339.46875,159.70418 339.46875,159.378839 C339.46875,158.690619 340.025,158.12753 340.71875,158.12753 L340.73125,158.12753 L358.9375,158.12753 L358.9375,159.378839 Z" id="path-17"></path>
-					<filter x="-0.1%" y="-4.3%" width="100.6%" height="105.7%" filterUnits="objectBoundingBox" id="filter-18">
-						<feOffset dx="1" dy="-1" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.6 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-					<path d="M358.9375,105.891453 L341.503125,105.891453 C341.728125,105.541086 341.934375,105.184463 342.115625,104.80907 C342.865625,103.3075 343.284375,101.611976 343.284375,99.8163479 C343.284375,93.6035992 338.240625,88.5545678 332.034375,88.5545678 C325.834375,88.5545678 320.784375,93.6035992 320.784375,99.8163479 C320.784375,101.60572 321.203125,103.301243 321.953125,104.802814 C322.134375,105.178207 322.340625,105.541086 322.565625,105.891453 L290.128125,105.891453 C289.446875,105.910222 288.896875,106.460798 288.896875,107.149018 C288.896875,107.349227 288.940625,107.530667 289.028125,107.699594 C289.034375,107.712107 289.040625,107.730877 289.046875,107.74339 C289.753125,109.11983 290.159375,110.677709 290.159375,112.329437 C290.159375,117.847709 285.671875,122.339908 280.159375,122.339908 C274.646875,122.339908 270.159375,117.847709 270.159375,112.329437 C270.159375,110.690222 270.553125,109.138599 271.259375,107.774673 C271.271875,107.755903 271.278125,107.74339 271.284375,107.72462 C271.378125,107.555693 271.428125,107.355484 271.428125,107.149018 C271.428125,106.473311 270.896875,105.922735 270.228125,105.897709 L238.378125,105.891453 C238.603125,105.541086 238.809375,105.184463 238.990625,104.80907 C239.740625,103.3075 240.159375,101.611976 240.159375,99.8163479 C240.159375,93.6035992 235.115625,88.5545678 228.909375,88.5545678 C222.709375,88.5545678 217.659375,93.6035992 217.659375,99.8163479 C217.659375,101.60572 218.078125,103.301243 218.828125,104.802814 C219.009375,105.178207 219.215625,105.541086 219.440625,105.891453 L187.628125,105.891453 C186.946875,105.910222 186.396875,106.460798 186.396875,107.149018 C186.396875,107.349227 186.440625,107.530667 186.528125,107.699594 C186.534375,107.712107 186.540625,107.730877 186.546875,107.74339 C187.253125,109.11983 187.659375,110.677709 187.659375,112.329437 C187.659375,117.847709 183.171875,122.339908 177.659375,122.339908 C172.146875,122.339908 167.659375,117.847709 167.659375,112.329437 C167.659375,110.690222 168.053125,109.138599 168.759375,107.774673 C168.771875,107.755903 168.778125,107.74339 168.784375,107.72462 C168.878125,107.555693 168.928125,107.355484 168.928125,107.149018 C168.928125,106.473311 168.396875,105.922735 167.728125,105.897709 L136.503125,105.891453 C136.728125,105.541086 136.934375,105.184463 137.115625,104.80907 C137.865625,103.3075 138.284375,101.611976 138.284375,99.8163479 C138.284375,93.6035992 133.240625,88.5545678 127.034375,88.5545678 C120.834375,88.5545678 115.784375,93.6035992 115.784375,99.8163479 C115.784375,101.60572 116.203125,103.301243 116.953125,104.802814 C117.134375,105.178207 117.340625,105.541086 117.565625,105.891453 L84.778125,105.891453 C84.103125,105.910222 83.565625,106.460798 83.565625,107.142762 C83.565625,107.311688 83.596875,107.468102 83.659375,107.618259 C83.659375,107.624515 83.665625,107.630772 83.665625,107.637029 C83.671875,107.643285 83.671875,107.649542 83.678125,107.662055 C84.228125,108.900851 84.534375,110.271034 84.534375,111.703782 C84.534375,117.222055 80.046875,121.714254 74.534375,121.714254 C69.021875,121.714254 64.534375,117.222055 64.534375,111.703782 C64.534375,110.271034 64.840625,108.900851 65.390625,107.662055 C65.453125,107.511898 65.496875,107.317945 65.496875,107.149018 C65.496875,106.460798 64.934375,105.891453 64.240625,105.891453 L34.696875,105.891453 C34.878125,105.51606 35.034375,105.128154 35.171875,104.733992 C35.571875,103.589044 35.784375,102.350248 35.784375,101.067657 C35.784375,94.8549081 30.740625,89.8058767 24.534375,89.8058767 C18.334375,89.8058767 13.284375,94.8549081 13.284375,101.067657 C13.284375,102.350248 13.496875,103.589044 13.896875,104.733992 C14.034375,105.128154 14.190625,105.51606 14.371875,105.891453 L0.189925641,105.891453 L0.189925641,107.142762 L14.296875,107.142762 L14.371875,107.142762 C15.059375,107.142762 15.621875,106.579673 15.621875,105.891453 C15.621875,105.691243 15.578125,105.509803 15.496875,105.347133 C15.490625,105.328364 15.478125,105.309594 15.471875,105.290824 C14.865625,104.001976 14.534375,102.575484 14.534375,101.067657 C14.534375,95.5493846 19.021875,91.0571856 24.534375,91.0571856 C30.046875,91.0571856 34.534375,95.5493846 34.534375,101.067657 C34.534375,102.575484 34.196875,104.008233 33.596875,105.290824 C33.496875,105.466008 33.446875,105.67873 33.446875,105.891453 C33.446875,106.585929 34.003125,107.142762 34.696875,107.142762 L34.709375,107.142762 L64.253125,107.142762 C64.084375,107.524411 63.934375,107.918573 63.809375,108.318992 C63.465625,109.388861 63.284375,110.521296 63.284375,111.703782 C63.284375,117.916531 68.334375,122.965563 74.534375,122.965563 C80.740625,122.965563 85.784375,117.916531 85.784375,111.703782 C85.784375,110.521296 85.603125,109.382604 85.259375,108.312735 C85.134375,107.912316 84.984375,107.524411 84.815625,107.142762 L117.453125,107.142762 C117.484375,107.149018 117.521875,107.149018 117.559375,107.149018 C118.253125,107.149018 118.809375,106.585929 118.809375,105.897709 C118.809375,105.647447 118.740625,105.415955 118.609375,105.222002 C118.609375,105.209489 118.609375,105.196976 118.603125,105.19072 C117.609375,103.63284 117.034375,101.793416 117.034375,99.8163479 C117.034375,94.2980757 121.521875,89.8058767 127.034375,89.8058767 C132.546875,89.8058767 137.034375,94.2980757 137.034375,99.8163479 C137.034375,101.743364 136.490625,103.532735 135.546875,105.059332 C135.353125,105.278311 135.234375,105.566112 135.234375,105.891453 C135.234375,106.579673 135.790625,107.142762 136.484375,107.142762 L136.496875,107.142762 L167.678125,107.142762 C167.484375,107.511898 167.315625,107.893547 167.159375,108.281453 C166.671875,109.539018 166.409375,110.902945 166.409375,112.329437 C166.409375,118.542186 171.459375,123.591217 177.659375,123.591217 C183.865625,123.591217 188.909375,118.542186 188.909375,112.329437 C188.909375,110.902945 188.646875,109.539018 188.159375,108.287709 C188.009375,107.893547 187.834375,107.511898 187.640625,107.142762 L219.328125,107.142762 C219.359375,107.149018 219.396875,107.149018 219.434375,107.149018 C220.128125,107.149018 220.684375,106.585929 220.684375,105.897709 C220.684375,105.647447 220.615625,105.415955 220.484375,105.222002 C220.484375,105.209489 220.484375,105.196976 220.478125,105.19072 C219.484375,103.63284 218.909375,101.793416 218.909375,99.8163479 C218.909375,94.2980757 223.396875,89.8058767 228.909375,89.8058767 C234.421875,89.8058767 238.909375,94.2980757 238.909375,99.8163479 C238.909375,101.743364 238.365625,103.532735 237.421875,105.059332 C237.228125,105.278311 237.109375,105.566112 237.109375,105.891453 C237.109375,106.579673 237.665625,107.142762 238.359375,107.142762 L238.371875,107.142762 L270.178125,107.142762 C269.984375,107.511898 269.815625,107.893547 269.659375,108.281453 C269.171875,109.539018 268.909375,110.902945 268.909375,112.329437 C268.909375,118.542186 273.959375,123.591217 280.159375,123.591217 C286.365625,123.591217 291.409375,118.542186 291.409375,112.329437 C291.409375,110.902945 291.146875,109.539018 290.659375,108.287709 C290.509375,107.893547 290.334375,107.511898 290.140625,107.142762 L322.453125,107.142762 C322.484375,107.149018 322.521875,107.149018 322.559375,107.149018 C323.253125,107.149018 323.809375,106.585929 323.809375,105.897709 C323.809375,105.647447 323.740625,105.415955 323.609375,105.222002 C323.609375,105.209489 323.609375,105.196976 323.603125,105.19072 C322.609375,103.63284 322.034375,101.793416 322.034375,99.8163479 C322.034375,94.2980757 326.521875,89.8058767 332.034375,89.8058767 C337.546875,89.8058767 342.034375,94.2980757 342.034375,99.8163479 C342.034375,101.743364 341.490625,103.532735 340.546875,105.059332 C340.353125,105.278311 340.234375,105.566112 340.234375,105.891453 C340.234375,106.579673 340.790625,107.142762 341.484375,107.142762 L341.496875,107.142762 L358.9375,107.142762 L358.9375,105.891453 Z" id="path-19"></path>
-					<filter x="-0.1%" y="-4.3%" width="100.6%" height="105.7%" filterUnits="objectBoundingBox" id="filter-20">
-						<feOffset dx="1" dy="-1" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.6 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-					<path d="M358.9375,53.0175827 L341.3625,53.0175827 C341.5875,53.3679492 341.79375,53.7245723 341.975,54.0999649 C342.725,55.6015356 343.14375,57.2970592 343.14375,59.0926874 C343.14375,65.3054361 338.1,70.3544675 331.89375,70.3544675 C325.69375,70.3544675 320.64375,65.3054361 320.64375,59.0926874 C320.64375,57.3033157 321.0625,55.6077921 321.8125,54.1062215 C321.99375,53.7308288 322.2,53.3679492 322.425,53.0175827 L289.9875,53.0175827 L289.98125,53.0113262 C289.30625,52.9988131 288.75625,52.4482372 288.75625,51.7600173 C288.75625,51.5598079 288.8,51.3783681 288.8875,51.2094414 C288.89375,51.1969283 288.9,51.1781586 288.90625,51.1656455 C289.6125,49.7892058 290.01875,48.2313262 290.01875,46.5795984 C290.01875,41.0613262 285.53125,36.5691272 280.01875,36.5691272 C274.50625,36.5691272 270.01875,41.0613262 270.01875,46.5795984 C270.01875,48.2188131 270.4125,49.7704361 271.11875,51.1343628 C271.13125,51.1531325 271.1375,51.1656455 271.14375,51.1844152 C271.2375,51.3533419 271.2875,51.5535513 271.2875,51.7600173 C271.2875,52.4357241 270.75625,52.9863 270.0875,53.0113262 L270.08125,53.0175827 L238.2375,53.0175827 C238.4625,53.3679492 238.66875,53.7245723 238.85,54.0999649 C239.6,55.6015356 240.01875,57.2970592 240.01875,59.0926874 C240.01875,65.3054361 234.975,70.3544675 228.76875,70.3544675 C222.56875,70.3544675 217.51875,65.3054361 217.51875,59.0926874 C217.51875,57.3033157 217.9375,55.6077921 218.6875,54.1062215 C218.86875,53.7308288 219.075,53.3679492 219.3,53.0175827 L187.4875,53.0175827 L187.48125,53.0113262 C186.80625,52.9988131 186.25625,52.4482372 186.25625,51.7600173 C186.25625,51.5598079 186.3,51.3783681 186.3875,51.2094414 C186.39375,51.1969283 186.4,51.1781586 186.40625,51.1656455 C187.1125,49.7892058 187.51875,48.2313262 187.51875,46.5795984 C187.51875,41.0613262 183.03125,36.5691272 177.51875,36.5691272 C172.00625,36.5691272 167.51875,41.0613262 167.51875,46.5795984 C167.51875,48.2188131 167.9125,49.7704361 168.61875,51.1343628 C168.63125,51.1531325 168.6375,51.1656455 168.64375,51.1844152 C168.7375,51.3533419 168.7875,51.5535513 168.7875,51.7600173 C168.7875,52.4357241 168.25625,52.9863 167.5875,53.0113262 L167.58125,53.0175827 L136.3625,53.0175827 C136.5875,53.3679492 136.79375,53.7245723 136.975,54.0999649 C137.725,55.6015356 138.14375,57.2970592 138.14375,59.0926874 C138.14375,65.3054361 133.1,70.3544675 126.89375,70.3544675 C120.69375,70.3544675 115.64375,65.3054361 115.64375,59.0926874 C115.64375,57.3033157 116.0625,55.6077921 116.8125,54.1062215 C116.99375,53.7308288 117.2,53.3679492 117.425,53.0175827 L84.6375,53.0175827 C83.9625,52.9988131 83.425,52.4482372 83.425,51.7662738 C83.425,51.5973471 83.45625,51.4409335 83.51875,51.2907764 C83.51875,51.2845199 83.525,51.2782634 83.525,51.2720068 C83.53125,51.2657503 83.53125,51.2594937 83.5375,51.2469806 C84.0875,50.0081848 84.39375,48.6380016 84.39375,47.2052529 C84.39375,41.6869806 79.90625,37.1947817 74.39375,37.1947817 C68.88125,37.1947817 64.39375,41.6869806 64.39375,47.2052529 C64.39375,48.6380016 64.7,50.0081848 65.25,51.2469806 C65.3125,51.3971377 65.35625,51.5910906 65.35625,51.7600173 C65.35625,52.4482372 64.79375,53.0175827 64.1,53.0175827 L34.55625,53.0175827 C34.7375,53.3929754 34.89375,53.7808812 35.03125,54.1750435 C35.43125,55.3199911 35.64375,56.5587869 35.64375,57.8413785 C35.64375,64.0541272 30.6,69.1031586 24.39375,69.1031586 C18.19375,69.1031586 13.14375,64.0541272 13.14375,57.8413785 C13.14375,56.5587869 13.35625,55.3199911 13.75625,54.1750435 C13.89375,53.7808812 14.05,53.3929754 14.23125,53.0175827 L0.1875,53.0175827 L0.1875,51.7662738 L14.15625,51.7662738 L14.23125,51.7662738 C14.91875,51.7662738 15.48125,52.3293628 15.48125,53.0175827 C15.48125,53.2177921 15.4375,53.3992319 15.35625,53.5619021 C15.35,53.5806717 15.3375,53.5994414 15.33125,53.618211 C14.725,54.9070592 14.39375,56.3335513 14.39375,57.8413785 C14.39375,63.3596508 18.88125,67.8518497 24.39375,67.8518497 C29.90625,67.8518497 34.39375,63.3596508 34.39375,57.8413785 C34.39375,56.3335513 34.05625,54.9008026 33.45625,53.618211 C33.35625,53.4430277 33.30625,53.2303052 33.30625,53.0175827 C33.30625,52.3231063 33.8625,51.7662738 34.55625,51.7662738 L34.56875,51.7662738 L64.1125,51.7662738 C63.94375,51.3846246 63.79375,50.9904623 63.66875,50.5900435 C63.325,49.5201743 63.14375,48.3877398 63.14375,47.2052529 C63.14375,40.9925042 68.19375,35.9434728 74.39375,35.9434728 C80.6,35.9434728 85.64375,40.9925042 85.64375,47.2052529 C85.64375,48.3877398 85.4625,49.5264309 85.11875,50.5963 C84.99375,50.9967188 84.84375,51.3846246 84.675,51.7662738 L117.3125,51.7662738 C117.34375,51.7600173 117.38125,51.7600173 117.41875,51.7600173 C118.1125,51.7600173 118.66875,52.3231063 118.66875,53.0113262 C118.66875,53.261588 118.6,53.4930801 118.46875,53.687033 C118.46875,53.6995461 118.46875,53.7120592 118.4625,53.7183157 C117.46875,55.2761953 116.89375,57.1156194 116.89375,59.0926874 C116.89375,64.6109597 121.38125,69.1031586 126.89375,69.1031586 C132.40625,69.1031586 136.89375,64.6109597 136.89375,59.0926874 C136.89375,57.1656717 136.35,55.3763 135.40625,53.8497031 C135.2125,53.6307241 135.09375,53.342923 135.09375,53.0175827 C135.09375,52.3293628 135.65,51.7662738 136.34375,51.7662738 L136.35625,51.7662738 L167.5375,51.7662738 C167.34375,51.3971377 167.175,51.0154885 167.01875,50.6275827 C166.53125,49.3700173 166.26875,48.0060906 166.26875,46.5795984 C166.26875,40.3668497 171.31875,35.3178183 177.51875,35.3178183 C183.725,35.3178183 188.76875,40.3668497 188.76875,46.5795984 C188.76875,48.0060906 188.50625,49.3700173 188.01875,50.6213262 C187.86875,51.0154885 187.69375,51.3971377 187.5,51.7662738 L219.1875,51.7662738 C219.21875,51.7600173 219.25625,51.7600173 219.29375,51.7600173 C219.9875,51.7600173 220.54375,52.3231063 220.54375,53.0113262 C220.54375,53.261588 220.475,53.4930801 220.34375,53.687033 C220.34375,53.6995461 220.34375,53.7120592 220.3375,53.7183157 C219.34375,55.2761953 218.76875,57.1156194 218.76875,59.0926874 C218.76875,64.6109597 223.25625,69.1031586 228.76875,69.1031586 C234.28125,69.1031586 238.76875,64.6109597 238.76875,59.0926874 C238.76875,57.1656717 238.225,55.3763 237.28125,53.8497031 C237.0875,53.6307241 236.96875,53.342923 236.96875,53.0175827 C236.96875,52.3293628 237.525,51.7662738 238.21875,51.7662738 L238.23125,51.7662738 L270.0375,51.7662738 C269.84375,51.3971377 269.675,51.0154885 269.51875,50.6275827 C269.03125,49.3700173 268.76875,48.0060906 268.76875,46.5795984 C268.76875,40.3668497 273.81875,35.3178183 280.01875,35.3178183 C286.225,35.3178183 291.26875,40.3668497 291.26875,46.5795984 C291.26875,48.0060906 291.00625,49.3700173 290.51875,50.6213262 C290.36875,51.0154885 290.19375,51.3971377 290,51.7662738 L322.3125,51.7662738 C322.34375,51.7600173 322.38125,51.7600173 322.41875,51.7600173 C323.1125,51.7600173 323.66875,52.3231063 323.66875,53.0113262 C323.66875,53.261588 323.6,53.4930801 323.46875,53.687033 C323.46875,53.6995461 323.46875,53.7120592 323.4625,53.7183157 C322.46875,55.2761953 321.89375,57.1156194 321.89375,59.0926874 C321.89375,64.6109597 326.38125,69.1031586 331.89375,69.1031586 C337.40625,69.1031586 341.89375,64.6109597 341.89375,59.0926874 C341.89375,57.1656717 341.35,55.3763 340.40625,53.8497031 C340.2125,53.6307241 340.09375,53.342923 340.09375,53.0175827 C340.09375,52.3293628 340.65,51.7662738 341.34375,51.7662738 L341.35625,51.7662738 L358.9375,51.7662738 L358.9375,53.0175827 Z" id="path-21"></path>
-					<filter x="-0.1%" y="-4.3%" width="100.6%" height="105.7%" filterUnits="objectBoundingBox" id="filter-22">
-						<feOffset dx="1" dy="-1" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.6 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-					<path d="M38.4908512,3.3508804 C38.4173404,3.62471732 38.3399363,3.8975746 38.2585,4.16936126 C38.2129376,4.32138781 38.1658012,4.47077516 38.117157,4.61767193 C36.9208274,3.80961789 35.8872081,2.77754239 35.0769886,1.58190671 C35.405465,1.36436281 35.7634163,1.14992006 36.1534376,0.940361886 C36.7924082,1.86876559 37.5837067,2.68444502 38.4908512,3.3508804 Z M47.8377825,25.2325066 C48.4218197,25.5788073 48.800276,25.6981002 49.0378962,25.6787616 C53.0068522,27.4196958 55.784375,31.3892235 55.784375,35.999594 C55.784375,37.7952223 55.365625,39.4907458 54.615625,40.9923165 C54.434375,41.3677092 54.228125,41.7243322 54.003125,42.0746987 L68.2124375,42.0775458 L68.2124375,25.1825916 C68.2311875,24.5068848 68.7811875,23.968822 69.4624375,23.968822 C69.6311875,23.968822 69.7874375,24.0001047 69.9374375,24.0626702 C69.9436875,24.0626702 69.9499375,24.0689267 69.9561875,24.0689267 C69.9624375,24.0751832 69.9686875,24.0751832 69.9811875,24.0814398 C71.2186875,24.6320157 72.5874375,24.9385864 74.0186875,24.9385864 C78.6583595,24.9385864 82.5719235,21.7563214 83.6949227,17.4584101 C84.0725526,17.654765 84.4523163,17.8467201 84.8341176,18.0342223 C83.4848916,22.7397932 79.1489691,26.1898953 74.0186875,26.1898953 C72.8374375,26.1898953 71.6999375,26.0084555 70.6311875,25.6643455 C70.2311875,25.5392147 69.8436875,25.3890576 69.4624375,25.2201309 L69.4624375,42.0777963 L85.228125,42.0809552 C85.896875,42.1059814 86.428125,42.6565573 86.428125,43.3322641 C86.428125,43.5387301 86.378125,43.7389395 86.284375,43.9078662 C86.278125,43.9266359 86.271875,43.939149 86.259375,43.9579186 C85.553125,45.3218453 85.159375,46.8734683 85.159375,48.512683 C85.159375,54.0309552 89.646875,58.5231542 95.159375,58.5231542 C100.671875,58.5231542 105.159375,54.0309552 105.159375,48.512683 C105.159375,46.8609552 104.753125,45.3030757 104.046875,43.9266359 C104.040625,43.9141228 104.034375,43.8953531 104.028125,43.8828401 C103.940625,43.7139134 103.896875,43.5324736 103.896875,43.3322641 C103.896875,42.6440442 104.446875,42.0934683 105.128125,42.0746987 L118.574938,42.0746987 L118.574938,25.2201309 C118.193688,25.3890576 117.806188,25.5392147 117.406188,25.6643455 C116.337438,26.0084555 115.199938,26.1898953 114.018688,26.1898953 C111.248282,26.1898953 108.70952,25.1838066 106.747637,23.5172249 C107.465339,23.535806 108.183942,23.5388882 108.90298,23.5262128 C110.400746,24.422851 112.150878,24.9385864 114.018688,24.9385864 C115.449938,24.9385864 116.818688,24.6320157 118.056188,24.0814398 C118.068688,24.0751832 118.074938,24.0751832 118.081188,24.0689267 C118.087438,24.0689267 118.093688,24.0626702 118.099938,24.0626702 C118.249938,24.0001047 118.406188,23.968822 118.574938,23.968822 C119.256188,23.968822 119.806188,24.5068848 119.824938,25.1825916 L119.824938,42.0746987 L136.940625,42.0746987 C136.715625,41.7243322 136.509375,41.3614526 136.328125,40.9860599 C135.578125,39.4844893 135.159375,37.7889657 135.159375,35.999594 C135.159375,29.9966294 139.874089,25.0801192 145.785398,24.7549349 C146.823412,25.2276851 147.844219,25.7399889 148.84546,26.2897186 C148.065392,26.093408 147.249281,25.9891228 146.409375,25.9891228 C140.896875,25.9891228 136.409375,30.4813217 136.409375,35.999594 C136.409375,37.976662 136.984375,39.8160861 137.978125,41.3739657 C137.984375,41.3802223 137.984375,41.3927353 137.984375,41.4052484 C138.115625,41.5992013 138.184375,41.8306935 138.184375,42.0809552 C138.184375,42.7691751 137.628125,43.3322641 136.934375,43.3322641 C136.896875,43.3322641 136.859375,43.3322641 136.828125,43.3260076 L119.824938,43.3260076 L119.824938,59.255733 C120.174938,59.0304974 120.537438,58.8240314 120.912438,58.6425916 C122.412438,57.8918063 124.106188,57.4726178 125.893688,57.4726178 C132.099938,57.4726178 137.143688,62.5279058 137.143688,68.7343979 C137.143688,74.7551156 132.406915,79.6829433 126.465731,79.9818215 C125.992581,79.5893866 125.527593,79.1622472 125.074089,78.7116884 C125.344213,78.7336673 125.617535,78.7448691 125.893688,78.7448691 C131.406188,78.7448691 135.893688,74.2526702 135.893688,68.7343979 C135.893688,63.2161257 131.406188,58.7239267 125.893688,58.7239267 C123.918688,58.7239267 122.081187,59.2995288 120.524938,60.2943194 C120.518687,60.3005759 120.506188,60.3005759 120.493688,60.3005759 C120.299938,60.4319634 120.068688,60.5007853 119.818688,60.5007853 C119.131188,60.5007853 118.568688,59.9439529 118.568688,59.2494764 C118.568688,59.2119372 118.568688,59.1743979 118.574938,59.1431152 L118.574938,43.3260076 L105.140625,43.3260076 C105.334375,43.6951437 105.509375,44.0767929 105.659375,44.4709552 C106.146875,45.7222641 106.409375,47.0861908 106.409375,48.512683 C106.409375,54.7254317 101.365625,59.7744631 95.159375,59.7744631 C88.959375,59.7744631 83.909375,54.7254317 83.909375,48.512683 C83.909375,47.0861908 84.171875,45.7222641 84.659375,44.4646987 C84.815625,44.0767929 84.984375,43.6951437 85.178125,43.3260076 L69.4624375,43.3260076 L69.4624375,59.1431152 C69.4686875,59.1743979 69.4686875,59.2119372 69.4686875,59.2494764 C69.4686875,59.9439529 68.9061875,60.5007853 68.2186875,60.5007853 C67.9686875,60.5007853 67.7374375,60.4319634 67.5436875,60.3005759 C67.5311875,60.3005759 67.5186875,60.3005759 67.5124375,60.2943194 C65.9561875,59.2995288 64.1186875,58.7239267 62.1436875,58.7239267 C56.6311875,58.7239267 52.1436875,63.2161257 52.1436875,68.7343979 C52.1436875,74.2526702 56.6311875,78.7448691 62.1436875,78.7448691 C64.0686875,78.7448691 65.8561875,78.2005497 67.3811875,77.2558115 C67.5999375,77.0618586 67.8874375,76.9429843 68.2124375,76.9429843 C68.8999375,76.9429843 69.4624375,77.4998168 69.4624375,78.1942932 L69.4624375,78.2068063 L69.4624375,94.3107764 L74.8576419,94.3107764 C74.8660776,94.7091883 74.8728046,95.126509 74.8770043,95.5577494 L69.4624375,95.5588344 L69.4624375,111.923325 C69.8311875,111.729372 70.2124375,111.560445 70.5999375,111.404031 C71.5188776,111.047054 72.4946705,110.810597 73.5101918,110.709354 C73.4940747,111.114331 73.4808977,111.536432 73.4703999,111.9728 C72.2661775,112.113897 71.1261045,112.469995 70.0936875,113.005707 C70.0749375,113.01822 70.0624375,113.024476 70.0436875,113.030733 C69.8749375,113.124581 69.6749375,113.174634 69.4686875,113.174634 C68.7936875,113.174634 68.2436875,112.642827 68.2186875,111.973377 L68.2124375,111.96712 L68.2124375,95.5590848 L63.8433162,95.5599603 C63.8253333,95.1092969 63.8253175,94.6942326 63.8386213,94.3107764 L68.2124375,94.3107764 L68.2124375,78.2130628 C67.8624375,78.4382984 67.5061875,78.6447644 67.1311875,78.8262042 C65.6311875,79.5769895 63.9374375,79.996178 62.1436875,79.996178 C55.9374375,79.996178 50.8936875,74.9471466 50.8936875,68.7343979 C50.8936875,62.5279058 55.9374375,57.4726178 62.1436875,57.4726178 C63.9311875,57.4726178 65.6249375,57.8918063 67.1249375,58.6425916 C67.4999375,58.8240314 67.8624375,59.0304974 68.2124375,59.255733 L68.2124375,43.3260076 L53.996875,43.3260076 L53.984375,43.3260076 C53.290625,43.3260076 52.734375,42.7629186 52.734375,42.0746987 C52.734375,41.7493584 52.853125,41.4615573 53.046875,41.2425783 C53.990625,39.7159814 54.534375,37.9266097 54.534375,35.999594 C54.534375,30.4813217 50.046875,25.9891228 44.534375,25.9891228 C39.021875,25.9891228 34.534375,30.4813217 34.534375,35.999594 C34.534375,37.976662 35.109375,39.8160861 36.103125,41.3739657 C36.109375,41.3802223 36.109375,41.3927353 36.109375,41.4052484 C36.240625,41.5992013 36.309375,41.8306935 36.309375,42.0809552 C36.309375,42.7691751 35.753125,43.3322641 35.059375,43.3322641 C35.021875,43.3322641 34.984375,43.3322641 34.953125,43.3260076 L34.7910412,43.3260076 C34.7265208,42.909263 34.6649728,42.4921593 34.6063373,42.0746987 L35.065625,42.0746987 C34.8445405,41.7304293 34.641559,41.3740785 34.4626101,41.0056464 C34.1990981,38.954542 34.0044606,36.8948975 33.871625,34.8269924 L33.875375,34.9602568 C33.8685,34.9077018 33.859125,34.8538955 33.846625,34.8007149 C33.810375,34.6417987 33.751,34.4847594 33.65975,34.3577516 C33.454125,34.5585866 33.46475,34.8901835 33.4335,35.1629688 C33.3944029,35.501244 33.3447088,35.8287046 33.2853063,36.1456791 C33.284686,36.0970551 33.284375,36.0483594 33.284375,35.999594 C33.284375,29.7868453 38.334375,24.7378139 44.534375,24.7378139 C45.6835845,24.7378139 46.7929345,24.9109342 47.8377825,25.2325066 Z M154.306886,29.8666105 C155.299948,30.6506393 156.204834,31.4560689 157.028686,32.277371 C157.437183,33.4428973 157.659375,34.695613 157.659375,35.999594 C157.659375,37.7952223 157.240625,39.4907458 156.490625,40.9923165 C156.309375,41.3677092 156.103125,41.7243322 155.878125,42.0746987 L163.320542,42.0761607 C163.47039,42.4942207 163.607591,42.9110466 163.732958,43.3260076 L155.871875,43.3260076 L155.859375,43.3260076 C155.165625,43.3260076 154.609375,42.7629186 154.609375,42.0746987 C154.609375,41.7493584 154.728125,41.4615573 154.921875,41.2425783 C155.865625,39.7159814 156.409375,37.9266097 156.409375,35.999594 C156.409375,33.6911158 155.624049,31.5622034 154.306886,29.8666105 Z M164.360443,57.4761343 C164.278277,57.9317301 164.193087,58.3536077 164.109692,58.7380367 C163.22578,58.7848575 162.371874,58.9472085 161.562811,59.2102393 C161.3875,58.7309308 161.245841,58.2950151 161.161072,58.0247029 C162.172483,57.6943335 163.246553,57.5038196 164.360443,57.4761343 Z M160.993507,58.080911 C160.869167,58.5822513 160.720086,59.0906008 160.54764,59.6031005 C157.070745,61.1728084 154.643688,64.676128 154.643688,68.7343979 C154.643688,69.1126245 154.664769,69.4860307 154.705817,69.8535006 C154.348276,70.3009235 153.993515,70.723469 153.646564,71.1163485 C153.480922,70.3483839 153.393688,69.5514545 153.393688,68.7343979 C153.393688,63.8051944 156.575061,59.6021193 160.993507,58.080911 Z M142.910788,111.267433 C142.785218,111.674709 142.650707,112.070836 142.506655,112.453398 C138.906132,111.405114 136.043733,108.599669 134.914836,105.028692 C135.571749,102.150177 135.682706,99.3851596 135.469527,96.8414301 C135.499926,96.7775225 135.530919,96.7139514 135.5625,96.6507241 C135.74375,96.2753314 135.95,95.9124518 136.175,95.5620853 L135.333566,95.5620853 C135.278977,95.1376975 135.2159,94.7204112 135.145469,94.3107764 L136.0625,94.3107764 C136.09375,94.3045199 136.13125,94.3045199 136.16875,94.3045199 C136.8625,94.3045199 137.41875,94.8676089 137.41875,95.5558288 C137.41875,95.8060906 137.35,96.0375827 137.21875,96.2315356 C137.21875,96.2440487 137.21875,96.2565618 137.2125,96.2628183 C136.21875,97.8206979 135.64375,99.660122 135.64375,101.63719 C135.64375,106.207374 138.721728,110.073773 142.910788,111.267433 Z M119.824938,137.028172 C119.340777,136.94529 118.926615,136.83634 118.574937,136.705456 L118.574938,131.906728 C118.28596,132.058722 117.989307,132.19919 117.684976,132.325119 C118.059751,131.727191 118.508157,131.166852 118.911935,130.709149 C119.430814,130.856438 119.808356,131.329454 119.818688,131.887958 L119.824938,131.894215 L119.824938,137.028172 Z M73.9403707,131.900756 C73.9688262,132.458784 73.986125,132.771382 73.986125,132.771382 C73.8137543,132.889733 73.6433535,133.004956 73.4749011,133.117115 C72.4724594,133.013938 71.509921,132.778846 70.6061875,132.426021 C70.2124375,132.275864 69.8311875,132.100681 69.4624375,131.906728 L69.4624375,135.411183 C69.0280047,135.613121 68.6115755,135.789445 68.2124375,135.942248 L68.2124375,131.894215 L68.2186875,131.887958 C68.2311875,131.212251 68.7811875,130.661675 69.4686875,130.661675 C69.6686875,130.661675 69.8499375,130.705471 70.0186875,130.793063 C70.0311875,130.799319 70.0499375,130.805576 70.0624375,130.811832 C71.2403958,131.417508 72.5513789,131.802775 73.9403707,131.900756 Z M30.2926012,42.0746987 C29.8287905,42.5546229 29.3454748,42.9690334 28.8644943,43.3260076 L17.3249375,43.3260076 L17.3249375,56.2360745 C16.8969861,56.6509119 16.479627,57.0089709 16.0749375,57.317418 L16.0749375,43.3260076 L2.315625,43.3260076 C2.484375,43.7076568 2.634375,44.0955626 2.759375,44.4959814 C3.103125,45.5658505 3.284375,46.7045416 3.284375,47.8870285 C3.284375,50.7083062 2.24426893,53.2896057 0.527517733,55.2670857 C0.356014636,54.7741436 0.207777285,54.292805 0.0802835261,53.824095 C1.30782047,52.1617395 2.034375,50.1074479 2.034375,47.8870285 C2.034375,46.4542798 1.728125,45.0840966 1.178125,43.8453008 C1.171875,43.8327877 1.171875,43.8265312 1.165625,43.8202746 C1.165625,43.8140181 1.159375,43.8077615 1.159375,43.801505 C1.096875,43.6513479 1.065625,43.4949343 1.065625,43.3260076 C1.065625,42.6440442 1.603125,42.0934683 2.278125,42.0746987 L16.0749375,42.0746987 L16.0749375,25.2201309 C15.6936875,25.3890576 15.3061875,25.5392147 14.9061875,25.6643455 C13.8374375,26.0084555 12.6999375,26.1898953 11.5186875,26.1898953 C9.91572499,26.1898953 8.39031237,25.8530765 7.00932311,25.2463825 C6.913741,24.7242862 6.86372721,24.2362378 6.85299765,23.7800651 C8.24682251,24.5194891 9.83494159,24.9385864 11.5186875,24.9385864 C12.9499375,24.9385864 14.3186875,24.6320157 15.5561875,24.0814398 C15.5686875,24.0751832 15.5749375,24.0751832 15.5811875,24.0689267 C15.5874375,24.0689267 15.5936875,24.0626702 15.5999375,24.0626702 C15.7499375,24.0001047 15.9061875,23.968822 16.0749375,23.968822 C16.7561875,23.968822 17.3061875,24.5068848 17.3249375,25.1825916 L17.3249375,42.0746987 L30.2926012,42.0746987 Z M6.96057894,6.019929 C6.69712957,5.65189123 6.47023782,5.29913154 6.27494448,4.96614282 C7.84134301,4.13651723 9.62595436,3.66633508 11.5186875,3.66633508 C12.6999375,3.66633508 13.8311875,3.84777487 14.8999375,4.19188482 C15.2999375,4.31701571 15.6936875,4.46717277 16.0749375,4.63609948 L16.0749375,2.26409421 C16.2243303,2.29862406 16.3725061,2.33291784 16.51925,2.36685079 C16.8035969,2.43261179 17.0718582,2.50911792 17.3249375,2.59480405 L17.3249375,4.62358639 C17.3249375,5.31806283 16.7561875,5.88115183 16.0686875,5.88115183 C15.8999375,5.88115183 15.7061875,5.83735602 15.5561875,5.77479058 C14.3186875,5.22421466 12.9499375,4.91764398 11.5186875,4.91764398 C9.87810754,4.91764398 8.32831411,5.31552808 6.96057894,6.019929 Z" id="path-23"></path>
-					<filter x="-0.3%" y="-1.1%" width="101.2%" height="101.5%" filterUnits="objectBoundingBox" id="filter-24">
-						<feOffset dx="1" dy="-1" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feComposite in="shadowOffsetOuter1" in2="SourceAlpha" operator="out" result="shadowOffsetOuter1"></feComposite>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.2 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-					<path d="M30.5031553,57.4213507 C30.1484354,57.9423573 29.934875,58.2456139 29.934875,58.2456139 C29.934875,58.2456139 29.8837922,57.8535091 29.7258291,57.3113461 C29.9817143,57.357935 30.2409573,57.3947376 30.5031553,57.4213507 Z M24.5192743,54.6039415 C24.164071,54.7667517 23.7861778,54.975501 23.3841988,55.2362436 C21.4547151,53.2138804 20.2686875,50.4748364 20.2686875,47.4621466 C20.2686875,42.1154073 24.011854,37.6230222 29.0167685,36.4810082 C29.14345,36.7077141 29.2748617,36.9327541 29.411125,37.1560537 C29.445152,37.2117654 29.5775774,37.3702818 29.7848375,37.6023876 C25.0953153,38.426964 21.5186875,42.5356548 21.5186875,47.4621466 C21.5186875,50.2555415 22.6685949,52.7860084 24.5192743,54.6039415 Z M2.62939916,21.1955283 C2.27834623,20.7135996 1.84365535,20.2802841 1.35704673,19.8567529 C1.28937195,19.675349 1.22640197,19.4920004 1.16875,19.3073209 C0.825,18.2374518 0.64375,17.1050173 0.64375,15.9225304 C0.64375,13.6711926 1.30688983,11.5726682 2.44758777,9.81183996 C2.98496379,9.91183157 3.47257581,9.88063119 3.879875,9.65416099 C4.39970463,9.36501936 4.64296589,8.47118058 4.71971894,7.25408476 C6.05642572,6.14367127 7.65296584,5.33658358 9.40271703,4.9390425 C9.61224139,5.37027452 9.80683736,5.754789 9.98797276,6.09456646 C5.38324024,6.98818811 1.89375,11.0564208 1.89375,15.9225304 C1.89375,17.3552791 2.2,18.7254623 2.75,19.9642581 C2.8125,20.1144152 2.85625,20.3083681 2.85625,20.4772948 C2.85625,20.7430855 2.77235276,20.991146 2.62939916,21.1955283 Z M12.658679,5.94107326 C12.7305705,5.5691191 12.8018254,5.1552123 12.8752662,4.70316124 C15.45648,4.92721044 17.7901952,6.02645153 19.5781579,7.70231817 C19.8445351,8.57986508 20.2326986,9.27554425 20.7835901,9.02570926 C22.2625522,10.932182 23.14375,13.3256991 23.14375,15.9225304 C23.14375,17.1050173 22.9625,18.2437084 22.61875,19.3135775 C22.49375,19.7139963 22.34375,20.1019021 22.175,20.4835513 L26.1857695,20.4835513 C26.1547107,20.9009513 26.1290423,21.3181528 26.1092444,21.7348602 L22.1375,21.7348602 C21.4625,21.7160906 20.925,21.1655147 20.925,20.4835513 C20.925,20.3146246 20.95625,20.158211 21.01875,20.0080539 C21.01875,20.0017974 21.025,19.9955408 21.025,19.9892843 C21.03125,19.9830277 21.03125,19.9767712 21.0375,19.9642581 C21.5875,18.7254623 21.89375,17.3552791 21.89375,15.9225304 C21.89375,10.6618549 17.8154305,6.33369296 12.658679,5.94107326 Z M145.074089,111.24572 C145.527593,111.696279 145.992581,112.123418 146.465731,112.515853 C146.276237,112.525386 146.085518,112.530209 145.893688,112.530209 C144.099938,112.530209 142.406188,112.111021 140.906188,111.360236 C140.531188,111.178796 140.174938,110.97233 139.824938,110.747094 L139.824938,111.101578 C139.357379,110.940164 138.932903,110.792406 138.576372,110.667717 C138.607968,110.001734 139.157557,109.477016 139.824938,109.477016 C140.149938,109.477016 140.437438,109.59589 140.656188,109.789843 C141.962418,110.599053 143.461235,111.114489 145.074089,111.24572 Z M119.271012,111.5095 C118.77259,111.774165 118.27299,112.015947 117.773405,112.236402 C116.717209,111.855398 115.579179,111.647661 114.39375,111.647661 C111.80269,111.647661 109.438084,112.64013 107.659498,114.265127 C107.079643,114.262086 106.514307,114.241061 105.966207,114.205603 C108.029514,111.870398 111.043763,110.396352 114.39375,110.396352 C116.140422,110.396352 117.795016,110.796272 119.271012,111.5095 Z M54.4626101,73.5396778 C54.4594408,73.5331528 54.4562791,73.526624 54.453125,73.5200914 C53.7235646,72.0594427 53.3074453,70.4152687 53.2853063,68.6797105 C53.3447088,68.362736 53.3944029,68.0352754 53.4335,67.6970003 C53.46475,67.4242149 53.454125,67.0926181 53.65975,66.891783 C53.751,67.0187908 53.810375,67.1758301 53.846625,67.3347463 C53.859125,67.387927 53.8685,67.4417332 53.875375,67.4942882 C53.874125,67.4498668 53.872875,67.4054453 53.871625,67.3610238 C54.0044606,69.4289289 54.1990981,71.4885734 54.4626101,73.5396778 Z M54.6063373,74.6087301 C54.6649728,75.0261907 54.7265208,75.4432945 54.7910411,75.860039 L48.8644943,75.860039 C49.3454748,75.5030648 49.8287905,75.0886543 50.2926012,74.6087301 L54.6063375,74.6087301 Z M51.5790861,93.5504758 C51.5582912,94.1599692 51.5089518,94.7523002 51.4230792,95.3090598 C49.5987538,92.8527103 46.6781606,91.2579581 43.3936875,91.2579581 C42.8176577,91.2579581 42.2533244,91.3069223 41.7042542,91.4009698 L40.9531649,90.2737139 C41.7392341,90.098674 42.5561703,90.0066492 43.3936875,90.0066492 C46.6169587,90.0066492 49.5266645,91.3702306 51.5790861,93.5504758 Z M58.117157,37.1517033 C58.1658012,37.0048066 58.2129376,36.8554192 58.2585,36.7033927 C58.3399363,36.431606 58.4173404,36.1587487 58.4908512,35.8849118 C60.146518,37.1012504 62.1880845,37.8204361 64.39375,37.8204361 C69.90625,37.8204361 74.39375,33.3282372 74.39375,27.8099649 C74.39375,25.8829492 73.85,24.0935775 72.90625,22.5669806 C72.7125,22.3480016 72.59375,22.0602005 72.59375,21.7348602 C72.59375,21.0466403 73.15,20.4835513 73.84375,20.4835513 L73.85625,20.4835513 L84.1771759,20.4835513 C84.1399295,20.6955441 84.1111714,20.918096 84.091625,21.1527513 C84.0739507,21.365096 84.0803256,21.5586867 84.1084108,21.7348602 L73.8625,21.7348602 C74.0875,22.0852267 74.29375,22.4418497 74.475,22.8172424 C75.225,24.3188131 75.64375,26.0143366 75.64375,27.8099649 C75.64375,34.0227136 70.6,39.071745 64.39375,39.071745 C62.072023,39.071745 59.9115597,38.3637226 58.117157,37.1517033 Z M76.1869172,1.67351483 C76.3584325,1.30434739 76.5387389,0.908671604 76.7284371,0.485677119 C78.4170275,2.00012584 80.6209115,2.94984752 83.0400634,3.0345022 C83.1217616,3.50869303 83.2231923,3.93414877 83.3505169,4.29190807 C80.6270463,4.28162909 78.1295653,3.29910214 76.1869172,1.67351483 Z M88.2124375,4.48956904 C88.6585663,4.23883943 89.081677,3.99416498 89.4624375,3.81258243 L89.4624375,12.2176118 C89.0679297,13.0566784 88.64361,13.756358 88.2124375,14.36565 L88.2124375,4.48956904 Z M88.2124375,23.2604124 C88.6290796,23.2141682 89.0504566,23.1486039 89.4624375,23.0717966 L89.4624375,27.6407276 C89.0842897,27.8949044 88.6688245,28.1524283 88.2124375,28.4120571 L88.2124375,23.2604124 Z M69.0378962,58.212793 C68.6488103,58.0421252 68.2482743,57.892875 67.8377825,57.7665381 C68.4218197,58.1128387 68.800276,58.2321317 69.0378962,58.212793 Z" id="path-25"></path>
-					<filter x="-0.3%" y="-1.3%" width="101.4%" height="101.8%" filterUnits="objectBoundingBox" id="filter-26">
-						<feOffset dx="1" dy="-1" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feComposite in="shadowOffsetOuter1" in2="SourceAlpha" operator="out" result="shadowOffsetOuter1"></feComposite>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.1 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-				</defs>
-				<g id="Illustration---Moose-Puzzle---Complete" stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
-					<rect fill="#FFFFFF" x="0" y="0" width="450" height="320"></rect>
-					<g id="Moose" transform="translate(33.000000, 41.000000)">
-						<g id="Border-Shadow" fill="black" fill-opacity="1">
-							<use filter="url(#filter-2)" xlink:href="#path-1"></use>
-						</g>
-						<rect id="Border" stroke="#B5BDC2" stroke-width="1.25" fill="#F1F5FB" x="12.5" y="13.7637435" width="360" height="212.723822" rx="13.75"></rect>
-						<g id="Background" transform="translate(20.220000, 23.723941)">
-							<g id="Far-Trees" opacity="0.4" transform="translate(0.994677, 104.790189)">
-								<mask id="mask-4" fill="white">
-									<use xlink:href="#path-3"></use>
-								</mask>
-								<g id="Path"></g>
-								<path d="M111.029948,81.8669969 C110.833698,81.859489 110.652448,81.7600099 110.539948,81.5985911 C110.428124,81.4371723 110.398073,81.2319576 110.459323,81.0455126 L124.237448,39.2755702 C124.322448,39.0184262 124.562448,38.8463712 124.830573,38.8463712 C124.841198,38.8463712 124.852448,38.8463712 124.863073,38.8469969 C125.144323,38.8613869 125.381198,39.062222 125.441198,39.3375099 L128.683698,54.1223503 L133.659323,33.2548974 C133.726198,32.9727272 133.978073,32.7743948 134.266823,32.7743948 L134.272448,32.7743948 C134.564323,32.7768974 134.814948,32.9808607 134.877448,33.2655335 L138.798073,51.146112 L142.420573,35.953345 C142.488073,35.6711749 142.739323,35.4728424 143.028698,35.4728424 L143.033698,35.4728424 C143.324323,35.4747194 143.575573,35.6780571 143.638698,35.9627298 L146.336823,48.1110623 L147.246823,44.2545283 C147.313073,43.9723581 147.565573,43.7727743 147.854948,43.7727743 L147.860573,43.7727743 C148.152448,43.7759026 148.403698,43.9804916 148.465573,44.2657901 L150.553073,53.9033712 L153.364323,42.1116618 C153.431823,41.8294916 153.683698,41.6305335 153.972448,41.6305335 L153.978073,41.6305335 C154.269323,41.6336618 154.520573,41.8376251 154.583073,42.1222979 L157.824948,56.9071382 L162.800573,36.0390597 C162.867448,35.7575152 163.119323,35.5585571 163.408698,35.5585571 L163.414323,35.5585571 C163.705573,35.5616853 163.956823,35.765023 164.018698,36.0503215 L167.939948,53.9309 L171.562448,38.738133 C171.629323,38.4559628 171.881198,38.2576304 172.169948,38.2576304 L172.174948,38.2576304 C172.466198,38.2595073 172.716823,38.462845 172.779948,38.7475178 L175.478073,50.8964759 L176.388073,47.0393162 C176.454948,46.7565204 176.706823,46.5575623 176.996198,46.5575623 L177.002448,46.5575623 C177.294323,46.5600649 177.545573,46.7652796 177.606823,47.050578 L178.389948,50.6656094 L182.323073,34.1708555 C182.389948,33.8886853 182.641823,33.6903529 182.931198,33.6903529 L182.936823,33.6903529 C183.228073,33.6928555 183.478698,33.8968188 183.541198,34.1814916 L186.783698,48.9663319 L191.759323,28.0982534 C191.826198,27.8160832 192.078073,27.6177508 192.366823,27.6177508 L192.372448,27.6177508 C192.664323,27.6202534 192.914948,27.8242168 192.977448,28.1095152 L196.898073,45.9900937 L200.520573,30.7973267 C200.587448,30.5151565 200.839323,30.3161984 201.128698,30.3161984 L201.133698,30.3161984 C201.424323,30.318701 201.675573,30.5220387 201.738698,30.8067115 L204.038073,41.1619183 L211.604323,20.1224105 C211.694323,19.8727743 211.929948,19.7082272 212.191823,19.7082272 C212.205573,19.7082272 212.219323,19.7088529 212.232448,19.7094785 C212.510573,19.7276225 212.743073,19.9278319 212.802448,20.1999916 L216.044948,34.9842063 L221.019948,14.1167534 C221.087448,13.8345832 221.339323,13.6362508 221.628073,13.6362508 L221.633698,13.6362508 C221.925573,13.6387534 222.176198,13.8427168 222.238698,14.1273895 L226.159323,32.0085937 L229.781823,16.815201 C229.848698,16.5336565 230.100573,16.3346984 230.389948,16.3346984 L230.394323,16.3346984 C230.685573,16.337201 230.936823,16.5399131 230.999948,16.8252115 L233.698073,28.973544 L234.608073,25.1170099 C234.673698,24.8373424 234.921198,24.6390099 235.208073,24.635256 L235.216198,24.635256 C235.499948,24.635256 235.748073,24.8267063 235.820573,25.1019942 L247.588698,69.9101147 C247.638073,70.1003136 247.596198,70.3030257 247.474323,70.4569366 L239.778073,80.2133921 C239.670573,80.3497848 239.510573,80.4348738 239.338073,80.4492639 C239.338073,80.4492639 186.973073,84.6987089 186.956198,84.6987089 C186.948698,84.6987089 111.029948,81.8669969 111.029948,81.8669969" id="Fill-8" fill="#8CCD5A" mask="url(#mask-4)"></path>
-								<path d="M21.2347604,78.0910471 C21.0753854,78.0847906 20.9241354,78.0178455 20.8128854,77.9039764 L3.17851044,59.8707382 C3.02476044,59.713699 2.96476044,59.487212 3.02038544,59.2738639 L14.7885104,14.4657435 C14.8603854,14.1904555 15.1091354,13.9990052 15.3928854,13.9990052 L15.4010104,13.9990052 C15.6878854,14.0027592 15.9353854,14.2010916 16.0010104,14.4807592 L16.9116354,18.3379188 L19.6091354,6.18896073 C19.6722604,5.90428796 19.9235104,5.70095026 20.2147604,5.69844764 L20.2191354,5.69844764 C20.5085104,5.69844764 20.7603854,5.89740576 20.8272604,6.17957592 L24.4497604,21.3723429 L28.3703854,3.4917644 C28.4328854,3.20646597 28.6835104,3.00250262 28.9753854,3 L28.9810104,3 C29.2697604,3 29.5216354,3.19833246 29.5891354,3.48050262 L34.5641354,24.3485812 L37.8066354,9.56374084 C37.8660104,9.29158115 38.0985104,9.09137173 38.3766354,9.07322775 C38.3897604,9.07260209 38.4035104,9.07260209 38.4172604,9.07260209 C38.6791354,9.07260209 38.9153854,9.23652356 39.0047604,9.48615969 L46.5710104,30.5256675 L48.8703854,20.1704607 C48.9335104,19.885788 49.1847604,19.6824503 49.4760104,19.6805733 L49.4803854,19.6805733 C49.7697604,19.6805733 50.0216354,19.8789058 50.0885104,20.1610759 L53.7110104,35.3538429 L57.6316354,17.4732644 C57.6941354,17.187966 57.9447604,16.9846283 58.2366354,16.9815 L58.2422604,16.9815 C58.5310104,16.9815 58.7828854,17.1804581 58.8503854,17.4620026 L63.8260104,38.3300812 L67.0678854,23.5452408 C67.1297604,23.2605681 67.3803854,23.0566047 67.6722604,23.0541021 L67.6778854,23.0541021 C67.9666354,23.0541021 68.2185104,23.2524346 68.2860104,23.5346047 L72.2185104,40.0299843 L73.0022604,36.4143272 C73.0635104,36.1290288 73.3147604,35.9244398 73.6066354,35.9213115 L73.6128854,35.9213115 C73.9022604,35.9213115 74.1541354,36.1202696 74.2210104,36.4030654 L75.1310104,40.2602251 L77.8291354,28.111267 C77.8922604,27.8265942 78.1428854,27.6232565 78.4341354,27.6213796 L78.4391354,27.6213796 C78.7278854,27.6213796 78.9797604,27.819712 79.0472604,28.1018822 L82.6691354,43.2946492 L86.5903854,25.4140707 C86.6522604,25.1293979 86.9035104,24.9254346 87.1947604,24.9229319 L87.2003854,24.9229319 C87.4897604,24.9229319 87.7416354,25.1212644 87.8085104,25.4034346 L92.7841354,46.2708874 L96.0260104,31.4860471 C96.0885104,31.2013743 96.3391354,30.997411 96.6303854,30.9949084 L96.6366354,30.9949084 C96.9253854,30.9949084 97.1772604,31.1932408 97.2441354,31.475411 L100.05601,43.2677461 L102.14351,33.6301649 C102.205385,33.3442408 102.456635,33.1396518 102.74851,33.1365236 L102.754135,33.1365236 C103.044135,33.1365236 103.29601,33.3361073 103.36226,33.6182775 L104.27226,37.4754372 L106.970385,25.3264791 C107.03351,25.0418063 107.28476,24.8384686 107.57601,24.8365916 L107.580385,24.8365916 C107.86976,24.8365916 108.121635,25.0349241 108.18851,25.3170942 L111.81101,40.5104869 L115.731635,22.6292827 C115.794135,22.3446099 116.04476,22.1406466 116.336635,22.138144 L116.34226,22.138144 C116.63101,22.138144 116.882885,22.3364764 116.950385,22.6186466 L121.925385,43.4860995 L125.167885,28.7018848 C125.22976,28.4197147 125.47601,28.2170026 125.764135,28.2101204 C125.76851,28.2101204 125.77351,28.2101204 125.77851,28.2101204 C126.06101,28.2101204 126.309135,28.4003194 126.38226,28.674356 L139.462885,77.6612225 C139.51226,77.8470419 139.474135,78.046 139.357885,78.1992853 C139.24226,78.3531963 139.06226,78.4451675 138.86976,78.4482958 C138.86976,78.4482958 63.3922604,79.7646728 63.3885104,79.7646728 C63.3803854,79.7646728 21.2347604,78.0910471 21.2347604,78.0910471" id="Fill-12" fill="#8CCD5A" mask="url(#mask-4)"></path>
-							</g>
-							<g id="Close-Trees" opacity="0.8" transform="translate(196.269936, -0.000000)" fill="#8CCD5A">
-								<path d="M117.335346,173.822636 C117.299108,173.891786 117.255493,173.957801 117.204752,174.019529 C116.977877,174.296068 116.634752,174.456861 116.271627,174.456861 L116.259127,174.456861 L63.5552819,173.901271 L65.0635019,180.227085 C65.1297519,180.506752 65.0641269,180.801435 64.8847519,181.026671 C64.7078769,181.248778 64.4391269,181.377663 64.1560019,181.377663 L64.1460019,181.377663 L0.922876943,180.691946 C0.642251943,180.688818 0.377251943,180.559307 0.202876943,180.339702 C0.027876943,180.119472 -0.038373057,179.832922 0.021626943,179.55826 L30.2328769,41.7841448 C30.3266269,41.3586997 30.7010019,41.0546317 31.1360019,41.0508777 L31.1441269,41.0508777 C31.5753769,41.0508777 31.9516269,41.3468123 32.0516269,41.7678777 L48.2830453,109.846086 L72.8672519,0.912204188 C72.9866269,0.383526178 73.4660019,0.0050052356 74.0222519,1.42108547e-14 L74.0335019,1.42108547e-14 C74.5853769,1.42108547e-14 75.0660019,0.368510471 75.1941269,0.891557592 L111.826337,150.17781 L122.271002,102.546454 C122.318502,102.328726 122.510377,102.172938 122.733502,102.171061 L122.737877,102.171061 C122.958502,102.171061 123.151002,102.322469 123.202252,102.53832 L140.115377,173.475022 C140.149752,173.618297 140.116002,173.769079 140.024127,173.8842 C139.933502,173.998069 139.796002,174.064388 139.650377,174.064388 L139.645377,174.064388 L117.335346,173.822636 Z" id="Path"></path>
-							</g>
-							<path d="M2.2874375,175.237553 C26.2074375,199.09626 72.2366756,189.729069 97.046,189.488898 C121.915203,189.248146 119.448063,199.964042 150.599313,195.160268 C173.809938,191.581524 202.298688,193.702493 227.333063,194.324393 C297.141188,196.059959 345.607438,189.032391 345.607438,176.653409 C345.607438,174.326356 342.22861,169.284795 319.724995,169.284795 C213.386709,169.284795 -11.0543714,161.930102 2.2874375,175.237553 Z" id="Fill-16" fill="#9EA5A9"></path>
-							<path d="M329.903437,173.211678 C329.903437,177.244568 302.381079,177.641954 282.285726,177.083294 C286.909877,176.072964 292.429008,176.167303 292.429008,174.535596 C292.429008,169.318454 287.420308,164.632039 277.403437,161.820822 C282.352119,160.909575 287.971637,160.392024 293.927369,160.392024 C313.800704,160.392024 329.903437,166.130933 329.903437,173.211678" id="Fill-57" fill="#E3E9F1"></path>
-							<path d="M297.158937,176.43992 C297.158937,181.558869 268.296845,182.063481 247.223395,181.35393 C252.072822,180.071454 257.860687,180.191324 257.860687,178.120673 C257.860687,171.498842 252.607881,165.550473 242.103938,161.982096 C247.293402,160.825934 253.186279,160.169229 259.431416,160.169229 C280.272619,160.169229 297.158937,167.45292 297.158937,176.43992" id="Fill-45" fill="#F9FBFF"></path>
-							<path d="M132.753437,166.538453 C132.753437,171.325961 104.490312,171.797705 83.8540625,171.134511 C88.6028125,169.935131 94.2709375,170.047124 94.2709375,168.110097 C94.2709375,161.916744 89.1271875,156.353425 78.8409375,153.016184 C83.9228125,151.934427 89.6934375,151.320035 95.8090625,151.320035 C116.217187,151.320035 132.753437,158.132786 132.753437,166.538453" id="Fill-47" fill="#CDD5DC"></path>
-							<path d="M111.57025,171.743836 C111.57025,173.418713 104.758375,174.218925 100.62775,174.700679 C86.5265,176.344273 69.162125,171.658747 54.75525,169.495234 C39.73025,171.769488 14.957125,171.099412 14.957125,165.35215 C14.957125,157.121666 24.69025,161.932323 38.78775,161.932323 C44.154625,161.932323 51.020875,157.858061 55.135875,159.516045 C60.984,157.249925 68.482125,155.889752 76.65775,155.889752 C83.959,155.889752 88.5465,157.002791 94.204625,158.520629 C101.487125,160.474548 111.57025,166.7386 111.57025,171.743836" id="Fill-49" fill="#F9FBFF"></path>
-							<path d="M105.4875,178.237441 C105.4875,186.094409 59.999375,187.132996 47.97,181.3532 C43.4525,181.742983 38.6325,181.909407 33.860625,181.852472 C16.639375,181.637873 0,178.478318 0,172.369428 C0,159.71369 15.995625,149.454834 35.73,149.454834 C37.21,149.454834 38.6675,149.511768 40.103125,149.626263 C45.155,147.873179 50.938125,146.878388 57.084375,146.878388 C73.999375,146.878388 88.17,154.415648 91.86875,164.538737 C100.0725,167.466174 105.4875,172.509574 105.4875,178.237441" id="Fill-51" fill="#B5BDC2"></path>
-							<path d="M282.82025,182.730015 C282.82025,184.81845 276.008375,185.816369 271.87775,186.416997 C257.7765,188.467266 240.412125,182.62428 226.00525,179.925832 C210.98025,182.761924 186.207125,181.926675 186.207125,174.758552 C186.207125,164.494065 195.94025,170.494091 210.03775,170.494091 C215.404625,170.494091 222.270875,165.412526 226.385875,167.479688 C232.234,164.654232 239.732125,162.957458 247.90775,162.957458 C255.209,162.957458 259.7965,164.345785 265.454625,166.239015 C272.737125,168.675314 282.82025,176.487861 282.82025,182.730015" id="Fill-53" fill="#868C8F"></path>
-							<path d="M255.754187,187.331391 C242.746062,190.295742 220.102937,190.517223 205.716687,187.963302 C189.426687,190.873846 162.467312,188.655276 162.467312,181.315723 C162.467312,170.803477 174.856062,162.280812 190.144812,162.280812 C193.986687,162.280812 197.651062,162.818875 200.978562,163.806784 C207.232937,158.093933 217.182312,153.895791 228.359812,154.403823 C281.061062,156.797577 274.646687,182.595187 255.754187,187.331391" id="Fill-55" fill="#B5BDC2"></path>
-						</g>
-						<g id="Background-Puzzle-Lines" transform="translate(12.937500, 14.390052)">
-							<g id="Fill-433">
-								<use fill="black" fill-opacity="1" filter="url(#filter-6)" xlink:href="#path-5"></use>
-								<use fill="#B5BDC2" fill-rule="evenodd" xlink:href="#path-5"></use>
-							</g>
-							<g id="Fill-435">
-								<use fill="black" fill-opacity="1" filter="url(#filter-8)" xlink:href="#path-7"></use>
-								<use fill="#B5BDC2" fill-rule="evenodd" xlink:href="#path-7"></use>
-							</g>
-							<g id="Fill-437">
-								<use fill="black" fill-opacity="1" filter="url(#filter-10)" xlink:href="#path-9"></use>
-								<use fill="#B5BDC2" fill-rule="evenodd" xlink:href="#path-9"></use>
-							</g>
-							<g id="Fill-439">
-								<use fill="black" fill-opacity="1" filter="url(#filter-12)" xlink:href="#path-11"></use>
-								<use fill="#B5BDC2" fill-rule="evenodd" xlink:href="#path-11"></use>
-							</g>
-							<g id="Fill-441">
-								<use fill="black" fill-opacity="1" filter="url(#filter-14)" xlink:href="#path-13"></use>
-								<use fill="#B5BDC2" fill-rule="evenodd" xlink:href="#path-13"></use>
-							</g>
-							<g id="Fill-443">
-								<use fill="black" fill-opacity="1" filter="url(#filter-16)" xlink:href="#path-15"></use>
-								<use fill="#B5BDC2" fill-rule="evenodd" xlink:href="#path-15"></use>
-							</g>
-							<g id="Fill-445">
-								<use fill="black" fill-opacity="1" filter="url(#filter-18)" xlink:href="#path-17"></use>
-								<use fill="#B5BDC2" fill-rule="evenodd" xlink:href="#path-17"></use>
-							</g>
-							<g id="Fill-447">
-								<use fill="black" fill-opacity="1" filter="url(#filter-20)" xlink:href="#path-19"></use>
-								<use fill="#B5BDC2" fill-rule="evenodd" xlink:href="#path-19"></use>
-							</g>
-							<g id="Fill-449">
-								<use fill="black" fill-opacity="1" filter="url(#filter-22)" xlink:href="#path-21"></use>
-								<use fill="#B5BDC2" fill-rule="evenodd" xlink:href="#path-21"></use>
-							</g>
-						</g>
-						<g transform="translate(69.429627, 35.910622)">
-							<path d="M42.937123,55.8757015 C43.335873,54.7595339 42.917748,53.4900811 41.886498,52.9069711 C39.753373,51.7007093 35.633373,47.2691989 35.418998,46.9182067 C29.550873,37.3018978 32.680248,24.4578376 35.123373,14.3472617 C35.647748,12.1768664 35.430248,9.60480096 32.878998,10.4569423 C30.591498,11.2202407 28.709623,15.7725025 27.733998,17.6294449 C26.210873,20.5255994 25.502123,17.5318428 25.223373,15.8175496 C24.648373,12.280725 24.865248,8.62565174 24.878373,5.06067269 C24.885248,3.30821457 24.127123,1.65961509 22.257123,2.51050515 C20.607748,3.26066483 20.148998,6.89384023 19.881498,8.33972766 C18.097123,17.9866936 18.810873,22.6315523 13.813998,11.1983428 C5.75024801,-7.25158124 13.100248,17.6294449 9.88774801,19.416314 C6.67524801,21.2025575 -1.53350199,10.8410942 0.250873011,19.7729371 C2.03587301,28.7054057 9.11399801,28.3531622 9.53087301,33.3496387 C9.85399801,37.2286962 5.62462301,35.1840575 3.69712301,35.3166962 C1.31462301,35.481869 1.34649801,37.9725994 2.27837301,39.4234921 C3.68649801,41.6145339 7.62587301,42.6393559 9.88774801,43.354479 C15.920248,45.2614737 17.181498,49.1261413 13.853998,54.297801 C13.000248,55.624814 10.068373,57.9691413 11.575248,59.4982407 C12.947748,60.8903219 17.512748,58.836924 18.810873,58.360801 C29.518373,54.4304397 22.023373,68.7222643 22.023373,68.7222643 C22.023373,68.7222643 19.881498,73.3671229 27.377123,66.5781465 C34.872123,59.7897957 35.942748,68.0077669 35.942748,68.0077669 C35.942748,68.0077669 41.342748,60.339746 42.937123,55.8757015" id="Fill-59" fill="#9EA5A9"></path>
-							<path d="M189.576498,103.254323 C188.864623,106.111687 187.082123,100.037208 187.082123,100.037208 C186.877748,100.937525 186.591498,101.862868 186.231498,102.796344 C184.023373,108.54298 179.182123,114.576792 176.728373,116.115902 C175.191498,117.081912 175.690248,119.840423 174.863998,121.289438 C172.427123,125.570792 171.232748,129.918465 170.513373,134.813585 C170.513373,134.813585 171.608998,152.079771 166.702123,158.072289 C161.803373,164.072941 160.307123,174.003328 160.307123,174.003328 C160.307123,174.003328 153.167748,180.430051 146.036498,179.357054 C138.897123,178.284682 146.036498,171.857959 146.036498,171.857959 L155.310248,158.637255 C168.517748,142.198184 156.741498,125.047119 156.741498,125.047119 L152.455873,122.550132 C152.455873,122.550132 151.932748,122.460038 151.033373,122.321143 C146.862748,121.674216 133.850873,114.265842 130.310248,117.573051 C119.801498,127.372051 105.342748,121.117383 103.919623,121.47776 C102.488373,121.838137 100.705873,132.554347 100.705873,132.554347 C100.705873,132.554347 101.433373,143.794229 99.993998,148.63304 C98.758998,152.750472 99.993998,175.067567 99.993998,175.067567 C87.498373,183.647166 85.355248,175.787695 85.355248,175.787695 C85.355248,175.787695 92.183998,150.949839 90.335873,141.674512 C88.479623,132.399184 92.690873,132.570614 88.822748,130.041093 C84.946498,127.511572 84.946498,124.310098 84.946498,124.310098 L77.864623,120.405389 C72.867748,118.260645 70.724623,113.258538 70.724623,113.258538 C70.724623,113.258538 66.799623,111.113794 66.079623,104.687072 C65.368373,98.2522159 56.495248,93.978996 56.495248,93.978996 C56.495248,93.978996 57.182123,98.4242709 57.312748,101.396129 C57.410873,103.762355 57.787123,103.90125 56.960873,106.856216 C56.527748,108.412218 55.742748,106.545266 54.875873,108.019308 C53.510248,110.336106 53.018998,108.076242 51.857748,108.019308 C50.001498,107.937347 50.165248,109.009718 48.954623,106.856216 C48.341498,105.783844 47.049623,107.184059 46.632748,106.512732 C45.928998,105.391559 47.957123,104.622004 48.022748,101.633253 C48.063373,99.8651531 46.672748,100.438253 46.542123,99.1281321 C46.329623,97.0084148 45.692123,96.0173782 43.958373,97.8924646 C39.312748,102.893946 35.747748,101.10958 35.747748,101.10958 C35.747748,101.10958 29.170873,104.729616 26.668373,97.9350091 C24.173998,91.1485353 26.105248,86.4636347 26.105248,86.4636347 L33.604623,69.6729462 C30.390873,60.7411033 38.601498,59.6687316 38.601498,59.6687316 L40.743998,54.6666243 C31.822123,50.0167604 31.102123,44.3026583 31.102123,44.3026583 C31.822123,42.1579148 37.889623,43.5906635 42.527123,44.6630353 C47.163998,45.735407 47.523373,49.6651426 47.523373,49.6651426 L57.517123,50.0167604 C57.517123,50.0167604 56.330248,44.1731478 66.975248,41.433407 L64.159623,48.3518939 C70.432123,46.4930745 76.212123,44.9001583 82.664623,45.3093363 C90.867123,45.8336347 96.182748,50.6636871 102.496498,55.2810169 C114.215873,63.8524829 129.132748,67.7571923 143.477123,64.907962 C148.849623,63.8443494 153.298998,62.8126452 158.859623,63.4345457 C165.197748,64.1471661 171.290248,66.3494698 176.728373,69.6729462 C196.004623,81.4615274 190.287748,100.397585 189.576498,103.254323" id="Fill-61" fill="#A87B3B"></path>
-							<path d="M46.2763105,83.3754669 C46.7994355,84.3708832 46.2631855,86.4374198 45.2063105,88.6359695 C43.9944355,91.1586083 42.0975605,93.8558046 40.3319355,95.3110769 C39.9788105,95.6013805 39.6313105,95.8428832 39.2950605,96.022446" id="Smile" stroke="#B9C2D1" stroke-width="1.25" stroke-linecap="round"></path>
-							<path d="M144.07923,97.2479779 C140.01394,110.237585 145.716897,117.352869 150.821799,121.648501" id="Flank" stroke="#B5BDC2" stroke-width="1.25" stroke-linecap="round"></path>
-							<path d="M110.66003,109.027299 C111.160655,110.012177 110.647469,112.056836 109.636049,114.232111 C108.476296,116.728043 107.04497,119.449791 105.355283,120.889656 C105.017346,121.176887 104.920314,121.47084 104.598526,121.648501" id="Flank" stroke="#B5BDC2" stroke-width="1.25" stroke-linecap="round"></path>
-							<path d="M48.409498,65.936225 C48.409498,67.4202774 49.611373,68.6234109 51.093248,68.6234109 C52.576373,68.6234109 53.777623,67.4202774 53.777623,65.936225 C53.777623,64.4521727 52.576373,63.2496648 51.093248,63.2496648 C49.611373,63.2496648 48.409498,64.4521727 48.409498,65.936225" id="Fill-65" fill="#494C4E"></path>
-							<path d="M109.592297,113.66735 L103.337265,122.089109 C103.337265,122.089109 118.407615,128.609661 130.257796,117.733168 C130.257796,117.733168 133.212515,114.663046 143.457919,120.033357 C143.457919,120.033357 155.046376,124.155445 153.432207,123.014958 C148.419391,119.472721 144.047735,111.371963 144.040848,111.36167 C141.627107,107.729539 135.236174,109.242641 131.922682,110.166285 C130.774981,110.486061 129.646691,110.859362 128.530297,111.25119 C122.263338,113.452565 116.382703,116.25369 109.592297,113.66735 Z" id="Belly" fill="#B5BDC2"></path>
-							<path d="M59.854498,77.0968994 C59.866998,77.15008 59.876373,77.2038863 59.883248,77.2564413 C59.881998,77.2120198 59.880748,77.1675983 59.879498,77.1231769 C60.130123,81.024758 60.600748,84.8969334 61.338873,88.7378261 C62.077623,92.5824727 62.979498,96.8262868 64.404498,100.476355 C64.404498,100.476355 58.710123,94.6027109 56.563248,94.2861298 C56.563248,94.2861298 58.598248,102.860098 56.962623,106.859907 C56.962623,106.859907 55.350748,108.972742 54.259498,109.256164 C53.359498,109.489533 53.146373,109.54459 52.036998,108.123104 C52.036998,108.123104 49.005748,109.418208 48.520748,106.626538 C48.520748,106.626538 47.357623,106.996926 46.661373,106.707248 C45.965123,106.41757 46.297623,105.00797 46.984498,104.522462 C47.671373,104.036954 48.023248,101.630062 48.023248,101.630062 L46.338248,99.1011664 L45.336373,96.9107502 C49.833248,93.5121952 50.508873,92.6175093 50.508873,87.6886036 C50.508873,87.6886036 58.441373,86.1113287 59.441373,77.4591533 C59.472623,77.186368 59.461998,76.8547711 59.667623,76.653936 C59.758873,76.7809439 59.818248,76.9379832 59.854498,77.0968994" id="Beard" fill="#B5BDC2"></path>
-							<path d="M58.045748,56.5004175 C60.377623,52.9986295 62.964498,50.8094646 64.266373,46.4655457 C65.683248,41.7368494 65.879498,36.684064 65.588248,31.786441 C65.343873,27.6733887 64.408248,23.6591897 63.510123,19.6518729 C63.009498,17.4170353 60.351373,10.243907 62.215748,8.44139651 C64.952623,5.79425253 68.601373,13.7919934 69.411998,15.3336059 C70.972623,18.302962 71.673873,15.0877237 71.938873,13.3915745 C72.466998,10.0155431 72.416373,6.41365043 72.210123,3.02573159 C72.093248,1.10872635 73.090748,-0.771990927 75.028873,0.32415567 C76.572623,1.19694363 76.990748,4.56922112 77.263873,6.04388865 C79.048248,15.6908546 78.334498,20.3357133 83.331373,8.90250384 C91.394498,-9.5480459 86.886998,13.3421478 90.099498,15.1283913 C93.311998,16.9146347 98.678873,8.54525515 96.893873,17.4770981 C95.109498,26.4095667 90.515748,25.9178023 90.099498,30.9149044 C89.742623,35.2025143 99.190748,31.8440012 99.190748,31.8440012 C99.190748,31.8440012 100.693873,36.8429803 87.130748,41.1305902 C73.567623,45.4182002 89.171998,55.9892578 89.171998,55.9892578 C89.171998,55.9892578 89.041998,59.9946976 78.334498,56.064962 C67.626998,52.1346007 75.121998,66.4264253 75.121998,66.4264253 C75.121998,66.4264253 77.263873,71.0712839 69.768248,64.2823075 C62.273248,57.4939567 61.656373,63.7792813 61.656373,63.7792813 C61.656373,63.7792813 55.886373,59.7438101 58.045748,56.5004175" id="Fill-73" fill="#9EA5A9"></path>
-						</g>
-						<g id="Body-Puzzle-Lines" transform="translate(95.437500, 78.206806)">
-							<g id="Combined-Shape">
-								<use fill="black" fill-opacity="1" filter="url(#filter-24)" xlink:href="#path-23"></use>
-								<use fill-opacity="0.7" fill="#6E7376" fill-rule="evenodd" xlink:href="#path-23"></use>
-							</g>
-						</g>
-						<g id="Antler-Puzzle-Lines" transform="translate(75.437500, 45.672775)">
-							<g id="Combined-Shape">
-								<use fill="black" fill-opacity="1" filter="url(#filter-26)" xlink:href="#path-25"></use>
-								<use fill-opacity="0.4" fill="#494C4E" fill-rule="evenodd" xlink:href="#path-25"></use>
-							</g>
-						</g>
-					</g>
-				</g>
-			</svg>
-		</template>
-		<template is="dom-if" if="[[!completed]]">
-			<?xml version="1.0" encoding="UTF-8"?>
-			<svg width="450px" height="320px" viewBox="0 0 450 320" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
-				<!-- Generator: Sketch 55.1 (78136) - https://sketchapp.com -->
-				<title>Illustration - Moose Puzzle - Incomplete</title>
-				<desc>Created with Sketch.</desc>
-				<defs>
-					<path d="M61.875,227.112565 L61.875,209.479628 C62.0921859,209.114576 62.4903634,208.868482 62.95,208.868482 C63.15,208.868482 63.33125,208.912277 63.5,208.999869 C63.5125,209.006126 63.53125,209.012382 63.54375,209.018639 C64.91875,209.725628 66.475,210.132304 68.125,210.132304 C73.6375,210.132304 78.125,205.640105 78.125,200.121832 C78.125,194.60356 73.6375,190.111361 68.125,190.111361 C66.4875,190.111361 64.9375,190.505524 63.575,191.212513 C63.55625,191.225026 63.54375,191.231283 63.525,191.237539 C63.35625,191.331387 63.15625,191.38144 62.95,191.38144 C62.4929804,191.38144 62.0932632,191.13765 61.875,190.77241 L61.875,173.681675 L46.7261011,173.681675 L46.86875,173.768892 C47.05,174.144284 47.20625,174.53219 47.34375,174.926352 C47.74375,176.0713 47.95625,177.310096 47.95625,178.592687 C47.95625,184.805436 42.9125,189.854468 36.70625,189.854468 C30.50625,189.854468 25.45625,184.805436 25.45625,178.592687 C25.45625,177.310096 25.66875,176.0713 26.06875,174.926352 C26.20625,174.53219 26.3625,174.144284 26.54375,173.768892 L26.7108906,173.681675 L11.875,173.681675 L11.875,193.52212 L11.875,26.8887435 C11.875,19.2948281 18.0310847,13.1387435 25.625,13.1387435 L359.375,13.1387435 C366.968915,13.1387435 373.125,19.2948281 373.125,26.8887435 L373.125,213.362565 C373.125,220.956481 366.968915,227.112565 359.375,227.112565 L61.875,227.112565 Z" id="path-1"></path>
-					<filter x="-2.5%" y="-1.4%" width="103.3%" height="105.6%" filterUnits="objectBoundingBox" id="filter-2">
-						<feOffset dx="-3" dy="3" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feGaussianBlur stdDeviation="1.5" in="shadowOffsetOuter1" result="shadowBlurOuter1"></feGaussianBlur>
-						<feColorMatrix values="0 0 0 0 0.803921569   0 0 0 0 0.835294118   0 0 0 0 0.862745098  0 0 0 1 0" type="matrix" in="shadowBlurOuter1"></feColorMatrix>
-					</filter>
-					<path d="M195.893687,196.99356 C197.318688,196.99356 198.681187,196.730785 199.931187,196.242775 C200.324937,196.092618 200.706187,195.917435 201.074938,195.723482 L201.074937,211.471204 L202.324938,211.471204 L202.324938,195.710969 L202.318688,195.704712 C202.306187,195.029005 201.756188,194.478429 201.068688,194.478429 C200.868687,194.478429 200.687438,194.522225 200.518687,194.609817 C200.506188,194.616073 200.487437,194.62233 200.474937,194.628586 C199.099938,195.335576 197.543687,195.742251 195.893687,195.742251 C190.381188,195.742251 185.893687,191.250052 185.893687,185.73178 C185.893687,180.213508 190.381188,175.721309 195.893687,175.721309 C197.531187,175.721309 199.081187,176.115471 200.443688,176.822461 C200.462437,176.834974 200.474937,176.84123 200.493688,176.847487 C200.662438,176.941335 200.862437,176.991387 201.068688,176.991387 C201.743688,176.991387 202.293688,176.459581 202.318688,175.790131 L202.324938,175.783874 L202.324938,142.029817 C202.674937,142.255052 203.031188,142.461518 203.406187,142.642958 C204.906188,143.393743 206.599938,143.812932 208.393687,143.812932 C214.599938,143.812932 219.643687,138.763901 219.643687,132.551152 C219.643687,126.34466 214.599938,121.289372 208.393687,121.289372 C206.606187,121.289372 204.912438,121.70856 203.412438,122.459346 C203.037438,122.640785 202.674937,122.847251 202.324938,123.072487 L202.324938,88.9993455 C202.306187,88.3236387 201.756188,87.7855759 201.074938,87.7855759 C200.906187,87.7855759 200.749937,87.8168586 200.599938,87.8794241 C200.593687,87.8794241 200.587437,87.8856806 200.581187,87.8856806 C200.574938,87.8919372 200.568688,87.8919372 200.556187,87.8981937 C199.318688,88.4487696 197.949938,88.7553403 196.518687,88.7553403 C191.006188,88.7553403 186.518687,84.2631414 186.518687,78.7448691 C186.518687,73.2265969 191.006188,68.7343979 196.518687,68.7343979 C197.949938,68.7343979 199.318688,69.0409686 200.556187,69.5915445 C200.706187,69.6541099 200.899937,69.6979058 201.068688,69.6979058 C201.756188,69.6979058 202.324938,69.1348168 202.324938,68.4403403 L202.324938,34.4860733 C202.699937,34.6675131 203.087437,34.8239267 203.481187,34.9615707 C204.624938,35.3619895 205.862437,35.574712 207.143687,35.574712 C213.349938,35.574712 218.393687,30.5256806 218.393687,24.3129319 C218.393687,18.1064398 213.349938,13.0511518 207.143687,13.0511518 C205.862437,13.0511518 204.624938,13.2638743 203.481187,13.6642932 C203.087437,13.8019372 202.699937,13.9583508 202.324938,14.1397906 L202.324938,0 L201.074938,0 L201.074938,14.064712 L201.074938,14.1397906 C201.074938,14.8280105 201.637438,15.3910995 202.324938,15.3910995 C202.524938,15.3910995 202.706187,15.3473037 202.868688,15.2659686 C202.887438,15.259712 202.906187,15.247199 202.924937,15.2409424 C204.212437,14.6340576 205.637437,14.3024607 207.143687,14.3024607 C212.656187,14.3024607 217.143687,18.7946597 217.143687,24.3129319 C217.143687,29.8312042 212.656187,34.3234031 207.143687,34.3234031 C205.637437,34.3234031 204.206187,33.9855497 202.924937,33.3849215 C202.749937,33.2848168 202.537438,33.2347644 202.324938,33.2347644 C201.631188,33.2347644 201.074938,33.7915969 201.074938,34.4860733 L201.074938,34.4985864 L201.074938,68.4528534 C200.693688,68.2839267 200.299937,68.1337696 199.899938,68.0086387 C198.831187,67.6645288 197.699937,67.483089 196.518687,67.483089 C190.312437,67.483089 185.268687,72.538377 185.268687,78.7448691 C185.268687,84.9576178 190.312437,90.0066492 196.518687,90.0066492 C197.699937,90.0066492 198.837437,89.8252094 199.906188,89.4810995 C200.306187,89.3559686 200.693688,89.2058115 201.074938,89.0368848 L201.074938,122.959869 C201.068688,122.991152 201.068688,123.028691 201.068688,123.06623 C201.068688,123.760707 201.631188,124.317539 202.318688,124.317539 C202.568688,124.317539 202.799937,124.248717 202.993688,124.11733 C203.006188,124.11733 203.018687,124.11733 203.024938,124.111073 C204.581187,123.116283 206.418687,122.540681 208.393687,122.540681 C213.906187,122.540681 218.393687,127.03288 218.393687,132.551152 C218.393687,138.069424 213.906187,142.561623 208.393687,142.561623 C206.468688,142.561623 204.681187,142.017304 203.156187,141.072565 C202.937437,140.878613 202.649937,140.759738 202.324938,140.759738 C201.637438,140.759738 201.074938,141.316571 201.074938,142.011047 L201.074938,142.02356 L201.074938,175.740079 C200.706187,175.546126 200.324937,175.377199 199.937437,175.220785 C198.681187,174.732775 197.318688,174.47 195.893687,174.47 C189.687437,174.47 184.643687,179.525288 184.643687,185.73178 C184.643687,191.944529 189.687437,196.99356 195.893687,196.99356 Z" id="path-3"></path>
-					<filter x="-1.4%" y="-0.2%" width="105.7%" height="100.5%" filterUnits="objectBoundingBox" id="filter-4">
-						<feOffset dx="1" dy="0" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.6 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-					<path d="M259.643687,196.99356 C258.218687,196.99356 256.856188,196.730785 255.606188,196.242775 C255.212438,196.092618 254.831187,195.917435 254.462437,195.723482 L254.462437,211.471204 L253.212437,211.471204 L253.212437,195.710969 L253.218687,195.704712 C253.231188,195.029005 253.781187,194.478429 254.468687,194.478429 C254.668688,194.478429 254.849937,194.522225 255.018687,194.609817 C255.031187,194.616073 255.049937,194.62233 255.062438,194.628586 C256.437437,195.335576 257.993688,195.742251 259.643687,195.742251 C265.156187,195.742251 269.643687,191.250052 269.643687,185.73178 C269.643687,180.213508 265.156187,175.721309 259.643687,175.721309 C258.006188,175.721309 256.456187,176.115471 255.093687,176.822461 C255.074938,176.834974 255.062438,176.84123 255.043687,176.847487 C254.874937,176.941335 254.674937,176.991387 254.468687,176.991387 C253.793687,176.991387 253.243687,176.459581 253.218687,175.790131 L253.212437,175.783874 L253.212437,142.029817 C252.862437,142.255052 252.506187,142.461518 252.131188,142.642958 C250.631187,143.393743 248.937437,143.812932 247.143687,143.812932 C240.937437,143.812932 235.893687,138.763901 235.893687,132.551152 C235.893687,126.34466 240.937437,121.289372 247.143687,121.289372 C248.931188,121.289372 250.624937,121.70856 252.124937,122.459346 C252.499937,122.640785 252.862437,122.847251 253.212437,123.072487 L253.212437,88.9993455 C253.231188,88.3236387 253.781187,87.7855759 254.462437,87.7855759 C254.631188,87.7855759 254.787438,87.8168586 254.937437,87.8794241 C254.943688,87.8794241 254.949938,87.8856806 254.956187,87.8856806 C254.962437,87.8919372 254.968687,87.8919372 254.981188,87.8981937 C256.218687,88.4487696 257.587437,88.7553403 259.018687,88.7553403 C264.531187,88.7553403 269.018687,84.2631414 269.018687,78.7448691 C269.018687,73.2265969 264.531187,68.7343979 259.018687,68.7343979 C257.587437,68.7343979 256.218687,69.0409686 254.981188,69.5915445 C254.831187,69.6541099 254.637438,69.6979058 254.468687,69.6979058 C253.781187,69.6979058 253.212437,69.1348168 253.212437,68.4403403 L253.212437,34.4860733 C252.837438,34.6675131 252.449938,34.8239267 252.056188,34.9615707 C250.912437,35.3619895 249.674937,35.574712 248.393687,35.574712 C242.187437,35.574712 237.143687,30.5256806 237.143687,24.3129319 C237.143687,18.1064398 242.187437,13.0511518 248.393687,13.0511518 C249.674937,13.0511518 250.912437,13.2638743 252.056188,13.6642932 C252.449938,13.8019372 252.837438,13.9583508 253.212437,14.1397906 L253.212437,0 L254.462437,0 L254.462437,14.064712 L254.462437,14.1397906 C254.462437,14.8280105 253.899937,15.3910995 253.212437,15.3910995 C253.012437,15.3910995 252.831187,15.3473037 252.668687,15.2659686 C252.649937,15.259712 252.631188,15.247199 252.612437,15.2409424 C251.324938,14.6340576 249.899938,14.3024607 248.393687,14.3024607 C242.881188,14.3024607 238.393687,18.7946597 238.393687,24.3129319 C238.393687,29.8312042 242.881188,34.3234031 248.393687,34.3234031 C249.899938,34.3234031 251.331187,33.9855497 252.612437,33.3849215 C252.787438,33.2848168 252.999937,33.2347644 253.212437,33.2347644 C253.906187,33.2347644 254.462437,33.7915969 254.462437,34.4860733 L254.462437,34.4985864 L254.462437,68.4528534 C254.843687,68.2839267 255.237437,68.1337696 255.637437,68.0086387 C256.706187,67.6645288 257.837438,67.483089 259.018687,67.483089 C265.224938,67.483089 270.268687,72.538377 270.268687,78.7448691 C270.268687,84.9576178 265.224938,90.0066492 259.018687,90.0066492 C257.837438,90.0066492 256.699938,89.8252094 255.631187,89.4810995 C255.231188,89.3559686 254.843687,89.2058115 254.462437,89.0368848 L254.462437,122.959869 C254.468687,122.991152 254.468687,123.028691 254.468687,123.06623 C254.468687,123.760707 253.906187,124.317539 253.218687,124.317539 C252.968687,124.317539 252.737437,124.248717 252.543687,124.11733 C252.531187,124.11733 252.518687,124.11733 252.512437,124.111073 C250.956187,123.116283 249.118688,122.540681 247.143687,122.540681 C241.631188,122.540681 237.143687,127.03288 237.143687,132.551152 C237.143687,138.069424 241.631188,142.561623 247.143687,142.561623 C249.068687,142.561623 250.856188,142.017304 252.381188,141.072565 C252.599938,140.878613 252.887438,140.759738 253.212437,140.759738 C253.899937,140.759738 254.462437,141.316571 254.462437,142.011047 L254.462437,142.02356 L254.462437,175.740079 C254.831187,175.546126 255.212438,175.377199 255.599938,175.220785 C256.856188,174.732775 258.218687,174.47 259.643687,174.47 C265.849938,174.47 270.893687,179.525288 270.893687,185.73178 C270.893687,191.944529 265.849938,196.99356 259.643687,196.99356 Z" id="path-5"></path>
-					<filter x="-1.4%" y="-0.2%" width="105.7%" height="100.5%" filterUnits="objectBoundingBox" id="filter-6">
-						<feOffset dx="1" dy="0" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.6 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-					<path d="M299.643687,196.99356 C301.068688,196.99356 302.431187,196.730785 303.681187,196.242775 C304.074937,196.092618 304.456187,195.917435 304.824938,195.723482 L304.824938,211.471204 L306.074938,211.471204 L306.074938,195.710969 L306.068688,195.704712 C306.056188,195.029005 305.506188,194.478429 304.818688,194.478429 C304.618687,194.478429 304.437437,194.522225 304.268687,194.609817 C304.256188,194.616073 304.237437,194.62233 304.224938,194.628586 C302.849938,195.335576 301.293687,195.742251 299.643687,195.742251 C294.131188,195.742251 289.643687,191.250052 289.643687,185.73178 C289.643687,180.213508 294.131188,175.721309 299.643687,175.721309 C301.281187,175.721309 302.831187,176.115471 304.193688,176.822461 C304.212438,176.834974 304.224938,176.84123 304.243688,176.847487 C304.412438,176.941335 304.612437,176.991387 304.818688,176.991387 C305.493688,176.991387 306.043688,176.459581 306.068688,175.790131 L306.074938,175.783874 L306.074938,142.029817 C306.424937,142.255052 306.781187,142.461518 307.156187,142.642958 C308.656187,143.393743 310.349938,143.812932 312.143688,143.812932 C318.349938,143.812932 323.393687,138.763901 323.393687,132.551152 C323.393687,126.34466 318.349938,121.289372 312.143688,121.289372 C310.356188,121.289372 308.662438,121.70856 307.162438,122.459346 C306.787438,122.640785 306.424937,122.847251 306.074938,123.072487 L306.074938,88.9993455 C306.056188,88.3236387 305.506188,87.7855759 304.824938,87.7855759 C304.656187,87.7855759 304.499937,87.8168586 304.349938,87.8794241 C304.343687,87.8794241 304.337438,87.8856806 304.331187,87.8856806 C304.324938,87.8919372 304.318688,87.8919372 304.306187,87.8981937 C303.068688,88.4487696 301.699938,88.7553403 300.268687,88.7553403 C294.756188,88.7553403 290.268687,84.2631414 290.268687,78.7448691 C290.268687,73.2265969 294.756188,68.7343979 300.268687,68.7343979 C301.699938,68.7343979 303.068688,69.0409686 304.306187,69.5915445 C304.456187,69.6541099 304.649938,69.6979058 304.818688,69.6979058 C305.506188,69.6979058 306.074938,69.1348168 306.074938,68.4403403 L306.074938,34.4860733 C306.449937,34.6675131 306.837438,34.8239267 307.231187,34.9615707 C308.374937,35.3619895 309.612437,35.574712 310.893687,35.574712 C317.099938,35.574712 322.143688,30.5256806 322.143688,24.3129319 C322.143688,18.1064398 317.099938,13.0511518 310.893687,13.0511518 C309.612437,13.0511518 308.374937,13.2638743 307.231187,13.6642932 C306.837438,13.8019372 306.449937,13.9583508 306.074938,14.1397906 L306.074938,0 L304.824938,0 L304.824938,14.064712 L304.824938,14.1397906 C304.824938,14.8280105 305.387437,15.3910995 306.074938,15.3910995 C306.274938,15.3910995 306.456187,15.3473037 306.618688,15.2659686 C306.637437,15.259712 306.656187,15.247199 306.674937,15.2409424 C307.962438,14.6340576 309.387437,14.3024607 310.893687,14.3024607 C316.406187,14.3024607 320.893687,18.7946597 320.893687,24.3129319 C320.893687,29.8312042 316.406187,34.3234031 310.893687,34.3234031 C309.387437,34.3234031 307.956187,33.9855497 306.674937,33.3849215 C306.499937,33.2848168 306.287438,33.2347644 306.074938,33.2347644 C305.381188,33.2347644 304.824938,33.7915969 304.824938,34.4860733 L304.824938,34.4985864 L304.824938,68.4528534 C304.443688,68.2839267 304.049937,68.1337696 303.649938,68.0086387 C302.581187,67.6645288 301.449937,67.483089 300.268687,67.483089 C294.062437,67.483089 289.018687,72.538377 289.018687,78.7448691 C289.018687,84.9576178 294.062437,90.0066492 300.268687,90.0066492 C301.449937,90.0066492 302.587437,89.8252094 303.656187,89.4810995 C304.056188,89.3559686 304.443688,89.2058115 304.824938,89.0368848 L304.824938,122.959869 C304.818688,122.991152 304.818688,123.028691 304.818688,123.06623 C304.818688,123.760707 305.381188,124.317539 306.068688,124.317539 C306.318688,124.317539 306.549937,124.248717 306.743688,124.11733 C306.756188,124.11733 306.768687,124.11733 306.774938,124.111073 C308.331188,123.116283 310.168687,122.540681 312.143688,122.540681 C317.656187,122.540681 322.143688,127.03288 322.143688,132.551152 C322.143688,138.069424 317.656187,142.561623 312.143688,142.561623 C310.218687,142.561623 308.431188,142.017304 306.906187,141.072565 C306.687437,140.878613 306.399938,140.759738 306.074938,140.759738 C305.387437,140.759738 304.824938,141.316571 304.824938,142.011047 L304.824938,142.02356 L304.824938,175.740079 C304.456187,175.546126 304.074937,175.377199 303.687437,175.220785 C302.431187,174.732775 301.068688,174.47 299.643687,174.47 C293.437437,174.47 288.393687,179.525288 288.393687,185.73178 C288.393687,191.944529 293.437437,196.99356 299.643687,196.99356 Z" id="path-7"></path>
-					<filter x="-1.4%" y="-0.2%" width="105.7%" height="100.5%" filterUnits="objectBoundingBox" id="filter-8">
-						<feOffset dx="1" dy="0" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.6 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-					<path d="M157.143687,196.99356 C155.718687,196.99356 154.356188,196.730785 153.106188,196.242775 C152.712438,196.092618 152.331187,195.917435 151.962437,195.723482 L151.962437,211.471204 L150.712437,211.471204 L150.712437,195.710969 L150.718687,195.704712 C150.731188,195.029005 151.281187,194.478429 151.968687,194.478429 C152.168688,194.478429 152.349937,194.522225 152.518687,194.609817 C152.531187,194.616073 152.549937,194.62233 152.562438,194.628586 C153.937437,195.335576 155.493688,195.742251 157.143687,195.742251 C162.656187,195.742251 167.143687,191.250052 167.143687,185.73178 C167.143687,180.213508 162.656187,175.721309 157.143687,175.721309 C155.506188,175.721309 153.956187,176.115471 152.593687,176.822461 C152.574938,176.834974 152.562438,176.84123 152.543687,176.847487 C152.374937,176.941335 152.174937,176.991387 151.968687,176.991387 C151.293687,176.991387 150.743687,176.459581 150.718687,175.790131 L150.712437,175.783874 L150.712437,142.029817 C150.362437,142.255052 150.006187,142.461518 149.631188,142.642958 C148.131187,143.393743 146.437437,143.812932 144.643687,143.812932 C138.437437,143.812932 133.393687,138.763901 133.393687,132.551152 C133.393687,126.34466 138.437437,121.289372 144.643687,121.289372 C146.431188,121.289372 148.124937,121.70856 149.624937,122.459346 C149.999937,122.640785 150.362437,122.847251 150.712437,123.072487 L150.712437,88.9993455 C150.731188,88.3236387 151.281187,87.7855759 151.962437,87.7855759 C152.131188,87.7855759 152.287438,87.8168586 152.437437,87.8794241 C152.443688,87.8794241 152.449938,87.8856806 152.456187,87.8856806 C152.462437,87.8919372 152.468687,87.8919372 152.481188,87.8981937 C153.718687,88.4487696 155.087437,88.7553403 156.518687,88.7553403 C162.031187,88.7553403 166.518687,84.2631414 166.518687,78.7448691 C166.518687,73.2265969 162.031187,68.7343979 156.518687,68.7343979 C155.087437,68.7343979 153.718687,69.0409686 152.481188,69.5915445 C152.331187,69.6541099 152.137438,69.6979058 151.968687,69.6979058 C151.281187,69.6979058 150.712437,69.1348168 150.712437,68.4403403 L150.712437,34.4860733 C150.337438,34.6675131 149.949938,34.8239267 149.556188,34.9615707 C148.412437,35.3619895 147.174937,35.574712 145.893687,35.574712 C139.687437,35.574712 134.643687,30.5256806 134.643687,24.3129319 C134.643687,18.1064398 139.687437,13.0511518 145.893687,13.0511518 C147.174937,13.0511518 148.412437,13.2638743 149.556188,13.6642932 C149.949938,13.8019372 150.337438,13.9583508 150.712437,14.1397906 L150.712437,0 L151.962437,0 L151.962437,14.064712 L151.962437,14.1397906 C151.962437,14.8280105 151.399937,15.3910995 150.712437,15.3910995 C150.512437,15.3910995 150.331187,15.3473037 150.168687,15.2659686 C150.149937,15.259712 150.131188,15.247199 150.112437,15.2409424 C148.824938,14.6340576 147.399937,14.3024607 145.893687,14.3024607 C140.381188,14.3024607 135.893687,18.7946597 135.893687,24.3129319 C135.893687,29.8312042 140.381188,34.3234031 145.893687,34.3234031 C147.399937,34.3234031 148.831187,33.9855497 150.112437,33.3849215 C150.287438,33.2848168 150.499937,33.2347644 150.712437,33.2347644 C151.406187,33.2347644 151.962437,33.7915969 151.962437,34.4860733 L151.962437,34.4985864 L151.962437,68.4528534 C152.343687,68.2839267 152.737437,68.1337696 153.137437,68.0086387 C154.206187,67.6645288 155.337438,67.483089 156.518687,67.483089 C162.724938,67.483089 167.768687,72.538377 167.768687,78.7448691 C167.768687,84.9576178 162.724938,90.0066492 156.518687,90.0066492 C155.337438,90.0066492 154.199938,89.8252094 153.131187,89.4810995 C152.731188,89.3559686 152.343687,89.2058115 151.962437,89.0368848 L151.962437,122.959869 C151.968687,122.991152 151.968687,123.028691 151.968687,123.06623 C151.968687,123.760707 151.406187,124.317539 150.718687,124.317539 C150.468687,124.317539 150.237437,124.248717 150.043687,124.11733 C150.031187,124.11733 150.018687,124.11733 150.012437,124.111073 C148.456187,123.116283 146.618687,122.540681 144.643687,122.540681 C139.131188,122.540681 134.643687,127.03288 134.643687,132.551152 C134.643687,138.069424 139.131188,142.561623 144.643687,142.561623 C146.568687,142.561623 148.356188,142.017304 149.881188,141.072565 C150.099938,140.878613 150.387438,140.759738 150.712437,140.759738 C151.399937,140.759738 151.962437,141.316571 151.962437,142.011047 L151.962437,142.02356 L151.962437,175.740079 C152.331187,175.546126 152.712438,175.377199 153.099938,175.220785 C154.356188,174.732775 155.718687,174.47 157.143687,174.47 C163.349938,174.47 168.393687,179.525288 168.393687,185.73178 C168.393687,191.944529 163.349938,196.99356 157.143687,196.99356 Z" id="path-9"></path>
-					<filter x="-1.4%" y="-0.2%" width="105.7%" height="100.5%" filterUnits="objectBoundingBox" id="filter-10">
-						<feOffset dx="1" dy="0" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.6 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-					<path d="M93.3936875,196.99356 C94.8186875,196.99356 96.1811875,196.730785 97.4311875,196.242775 C97.8249375,196.092618 98.2061875,195.917435 98.5749375,195.723482 L98.5724375,211.471204 L99.8249375,211.471204 L99.8249375,195.710969 L99.8186875,195.704712 C99.8061875,195.029005 99.2561875,194.478429 98.5686875,194.478429 C98.3686875,194.478429 98.1874375,194.522225 98.0186875,194.609817 C98.0061875,194.616073 97.9874375,194.62233 97.9749375,194.628586 C96.5999375,195.335576 95.0436875,195.742251 93.3936875,195.742251 C87.8811875,195.742251 83.3936875,191.250052 83.3936875,185.73178 C83.3936875,180.213508 87.8811875,175.721309 93.3936875,175.721309 C95.0311875,175.721309 96.5811875,176.115471 97.9436875,176.822461 C97.9624375,176.834974 97.9749375,176.84123 97.9936875,176.847487 C98.1624375,176.941335 98.3624375,176.991387 98.5686875,176.991387 C99.2436875,176.991387 99.7936875,176.459581 99.8186875,175.790131 L99.8249375,175.783874 L99.8249375,142.029817 C100.174938,142.255052 100.531188,142.461518 100.906188,142.642958 C102.406188,143.393743 104.099938,143.812932 105.893688,143.812932 C112.099938,143.812932 117.143688,138.763901 117.143688,132.551152 C117.143688,126.34466 112.099938,121.289372 105.893688,121.289372 C104.106188,121.289372 102.412438,121.70856 100.912438,122.459346 C100.537438,122.640785 100.174938,122.847251 99.8249375,123.072487 L99.8249375,88.9993455 C99.8061875,88.3236387 99.2561875,87.7855759 98.5749375,87.7855759 C98.4061875,87.7855759 98.2499375,87.8168586 98.0999375,87.8794241 C98.0936875,87.8794241 98.0874375,87.8856806 98.0811875,87.8856806 C98.0749375,87.8919372 98.0686875,87.8919372 98.0561875,87.8981937 C96.8186875,88.4487696 95.4499375,88.7553403 94.0186875,88.7553403 C88.5061875,88.7553403 84.0186875,84.2631414 84.0186875,78.7448691 C84.0186875,73.2265969 88.5061875,68.7343979 94.0186875,68.7343979 C95.4499375,68.7343979 96.8186875,69.0409686 98.0561875,69.5915445 C98.2061875,69.6541099 98.3999375,69.6979058 98.5686875,69.6979058 C99.2561875,69.6979058 99.8249375,69.1348168 99.8249375,68.4403403 L99.8249375,34.4860733 C100.199938,34.6675131 100.587438,34.8239267 100.981188,34.9615707 C102.124938,35.3619895 103.362438,35.574712 104.643688,35.574712 C110.849938,35.574712 115.893688,30.5256806 115.893688,24.3129319 C115.893688,18.1064398 110.849938,13.0511518 104.643688,13.0511518 C103.362438,13.0511518 102.124938,13.2638743 100.981188,13.6642932 C100.587438,13.8019372 100.199938,13.9583508 99.8249375,14.1397906 L99.8249375,0 L98.5749375,0 L98.5749375,14.064712 L98.5749375,14.1397906 C98.5749375,14.8280105 99.1374375,15.3910995 99.8249375,15.3910995 C100.024938,15.3910995 100.206188,15.3473037 100.368688,15.2659686 C100.387438,15.259712 100.406188,15.247199 100.424938,15.2409424 C101.712438,14.6340576 103.137438,14.3024607 104.643688,14.3024607 C110.156188,14.3024607 114.643688,18.7946597 114.643688,24.3129319 C114.643688,29.8312042 110.156188,34.3234031 104.643688,34.3234031 C103.137438,34.3234031 101.706188,33.9855497 100.424938,33.3849215 C100.249938,33.2848168 100.037438,33.2347644 99.8249375,33.2347644 C99.1311875,33.2347644 98.5749375,33.7915969 98.5749375,34.4860733 L98.5749375,34.4985864 L98.5749375,68.4528534 C98.1936875,68.2839267 97.7999375,68.1337696 97.3999375,68.0086387 C96.3311875,67.6645288 95.1999375,67.483089 94.0186875,67.483089 C87.8124375,67.483089 82.7686875,72.538377 82.7686875,78.7448691 C82.7686875,84.9576178 87.8124375,90.0066492 94.0186875,90.0066492 C95.1999375,90.0066492 96.3374375,89.8252094 97.4061875,89.4810995 C97.8061875,89.3559686 98.1936875,89.2058115 98.5749375,89.0368848 L98.5749375,122.959869 C98.5686875,122.991152 98.5686875,123.028691 98.5686875,123.06623 C98.5686875,123.760707 99.1311875,124.317539 99.8186875,124.317539 C100.068688,124.317539 100.299938,124.248717 100.493688,124.11733 C100.506188,124.11733 100.518688,124.11733 100.524938,124.111073 C102.081188,123.116283 103.918688,122.540681 105.893688,122.540681 C111.406188,122.540681 115.893688,127.03288 115.893688,132.551152 C115.893688,138.069424 111.406188,142.561623 105.893688,142.561623 C103.968688,142.561623 102.181188,142.017304 100.656188,141.072565 C100.437438,140.878613 100.149938,140.759738 99.8249375,140.759738 C99.1374375,140.759738 98.5749375,141.316571 98.5749375,142.011047 L98.5749375,142.02356 L98.5749375,175.740079 C98.2061875,175.546126 97.8249375,175.377199 97.4374375,175.220785 C96.1811875,174.732775 94.8186875,174.47 93.3936875,174.47 C87.1874375,174.47 82.1436875,179.525288 82.1436875,185.73178 C82.1436875,191.944529 87.1874375,196.99356 93.3936875,196.99356 Z" id="path-11"></path>
-					<filter x="-1.4%" y="-0.2%" width="105.7%" height="100.5%" filterUnits="objectBoundingBox" id="filter-12">
-						<feOffset dx="1" dy="0" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.6 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-					<path d="M55.2686875,196.99356 C53.8436875,196.99356 52.4811875,196.730785 51.2311875,196.242775 C50.8374375,196.092618 50.4561875,195.917435 50.0874375,195.723482 L50.0874375,211.471204 L48.8374375,211.471204 L48.8374375,195.710969 L48.8436875,195.704712 C48.8561875,195.029005 49.4061875,194.478429 50.0936875,194.478429 C50.2936875,194.478429 50.4749375,194.522225 50.6436875,194.609817 C50.6561875,194.616073 50.6749375,194.62233 50.6874375,194.628586 C52.0624375,195.335576 53.6186875,195.742251 55.2686875,195.742251 C60.7811875,195.742251 65.2686875,191.250052 65.2686875,185.73178 C65.2686875,180.213508 60.7811875,175.721309 55.2686875,175.721309 C53.6311875,175.721309 52.0811875,176.115471 50.7186875,176.822461 C50.6999375,176.834974 50.6874375,176.84123 50.6686875,176.847487 C50.4999375,176.941335 50.2999375,176.991387 50.0936875,176.991387 C49.4186875,176.991387 48.8686875,176.459581 48.8436875,175.790131 L48.8374375,175.783874 L48.8374375,142.029817 C48.4874375,142.255052 48.1311875,142.461518 47.7561875,142.642958 C46.2561875,143.393743 44.5624375,143.812932 42.7686875,143.812932 C36.5624375,143.812932 31.5186875,138.763901 31.5186875,132.551152 C31.5186875,126.34466 36.5624375,121.289372 42.7686875,121.289372 C44.5561875,121.289372 46.2499375,121.70856 47.7499375,122.459346 C48.1249375,122.640785 48.4874375,122.847251 48.8374375,123.072487 L48.8374375,88.9993455 C48.8561875,88.3236387 49.4061875,87.7855759 50.0874375,87.7855759 C50.2561875,87.7855759 50.4124375,87.8168586 50.5624375,87.8794241 C50.5686875,87.8794241 50.5749375,87.8856806 50.5811875,87.8856806 C50.5874375,87.8919372 50.5936875,87.8919372 50.6061875,87.8981937 C51.8436875,88.4487696 53.2124375,88.7553403 54.6436875,88.7553403 C60.1561875,88.7553403 64.6436875,84.2631414 64.6436875,78.7448691 C64.6436875,73.2265969 60.1561875,68.7343979 54.6436875,68.7343979 C53.2124375,68.7343979 51.8436875,69.0409686 50.6061875,69.5915445 C50.4561875,69.6541099 50.2624375,69.6979058 50.0936875,69.6979058 C49.4061875,69.6979058 48.8374375,69.1348168 48.8374375,68.4403403 L48.8374375,34.4860733 C48.4624375,34.6675131 48.0749375,34.8239267 47.6811875,34.9615707 C46.5374375,35.3619895 45.2999375,35.574712 44.0186875,35.574712 C37.8124375,35.574712 32.7686875,30.5256806 32.7686875,24.3129319 C32.7686875,18.1064398 37.8124375,13.0511518 44.0186875,13.0511518 C45.2999375,13.0511518 46.5374375,13.2638743 47.6811875,13.6642932 C48.0749375,13.8019372 48.4624375,13.9583508 48.8374375,14.1397906 L48.8374375,0 L50.0874375,0 L50.0874375,14.064712 L50.0874375,14.1397906 C50.0874375,14.8280105 49.5249375,15.3910995 48.8374375,15.3910995 C48.6374375,15.3910995 48.4561875,15.3473037 48.2936875,15.2659686 C48.2749375,15.259712 48.2561875,15.247199 48.2374375,15.2409424 C46.9499375,14.6340576 45.5249375,14.3024607 44.0186875,14.3024607 C38.5061875,14.3024607 34.0186875,18.7946597 34.0186875,24.3129319 C34.0186875,29.8312042 38.5061875,34.3234031 44.0186875,34.3234031 C45.5249375,34.3234031 46.9561875,33.9855497 48.2374375,33.3849215 C48.4124375,33.2848168 48.6249375,33.2347644 48.8374375,33.2347644 C49.5311875,33.2347644 50.0874375,33.7915969 50.0874375,34.4860733 L50.0874375,34.4985864 L50.0874375,68.4528534 C50.4686875,68.2839267 50.8624375,68.1337696 51.2624375,68.0086387 C52.3311875,67.6645288 53.4624375,67.483089 54.6436875,67.483089 C60.8499375,67.483089 65.8936875,72.538377 65.8936875,78.7448691 C65.8936875,84.9576178 60.8499375,90.0066492 54.6436875,90.0066492 C53.4624375,90.0066492 52.3249375,89.8252094 51.2561875,89.4810995 C50.8561875,89.3559686 50.4686875,89.2058115 50.0874375,89.0368848 L50.0874375,122.959869 C50.0936875,122.991152 50.0936875,123.028691 50.0936875,123.06623 C50.0936875,123.760707 49.5311875,124.317539 48.8436875,124.317539 C48.5936875,124.317539 48.3624375,124.248717 48.1686875,124.11733 C48.1561875,124.11733 48.1436875,124.11733 48.1374375,124.111073 C46.5811875,123.116283 44.7436875,122.540681 42.7686875,122.540681 C37.2561875,122.540681 32.7686875,127.03288 32.7686875,132.551152 C32.7686875,138.069424 37.2561875,142.561623 42.7686875,142.561623 C44.6936875,142.561623 46.4811875,142.017304 48.0061875,141.072565 C48.2249375,140.878613 48.5124375,140.759738 48.8374375,140.759738 C49.5249375,140.759738 50.0874375,141.316571 50.0874375,142.011047 L50.0874375,142.02356 L50.0874375,175.740079 C50.4561875,175.546126 50.8374375,175.377199 51.2249375,175.220785 C52.4811875,174.732775 53.8436875,174.47 55.2686875,174.47 C61.4749375,174.47 66.5186875,179.525288 66.5186875,185.73178 C66.5186875,191.944529 61.4749375,196.99356 55.2686875,196.99356 Z" id="path-13"></path>
-					<filter x="-1.4%" y="-0.2%" width="105.7%" height="100.5%" filterUnits="objectBoundingBox" id="filter-14">
-						<feOffset dx="1" dy="0" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.6 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-					<path d="M358.9375,159.378839 L340.7375,159.378839 C340.9625,159.729206 341.16875,160.085829 341.35,160.461221 C342.1,161.962792 342.51875,163.658316 342.51875,165.453944 C342.51875,171.666693 337.475,176.715724 331.26875,176.715724 C325.06875,176.715724 320.01875,171.666693 320.01875,165.453944 C320.01875,163.664572 320.4375,161.969049 321.1875,160.467478 C321.36875,160.092085 321.575,159.729206 321.8,159.378839 L289.3625,159.378839 C288.68125,159.36007 288.13125,158.809494 288.13125,158.121274 C288.13125,157.921064 288.175,157.739625 288.2625,157.570698 C288.26875,157.558185 288.275,157.539415 288.28125,157.526902 C288.9875,156.150462 289.39375,154.592583 289.39375,152.940855 C289.39375,147.422583 284.90625,142.930384 279.39375,142.930384 C273.88125,142.930384 269.39375,147.422583 269.39375,152.940855 C269.39375,154.58007 269.7875,156.131693 270.49375,157.495619 C270.50625,157.514389 270.5125,157.526902 270.51875,157.545672 C270.6125,157.714598 270.6625,157.914808 270.6625,158.121274 C270.6625,158.796981 270.13125,159.347557 269.4625,159.372583 L237.6125,159.378839 C237.8375,159.729206 238.04375,160.085829 238.225,160.461221 C238.975,161.962792 239.39375,163.658316 239.39375,165.453944 C239.39375,171.666693 234.35,176.715724 228.14375,176.715724 C221.94375,176.715724 216.89375,171.666693 216.89375,165.453944 C216.89375,163.664572 217.3125,161.969049 218.0625,160.467478 C218.24375,160.092085 218.45,159.729206 218.675,159.378839 L186.8625,159.378839 C186.18125,159.36007 185.63125,158.809494 185.63125,158.121274 C185.63125,157.921064 185.675,157.739625 185.7625,157.570698 C185.76875,157.558185 185.775,157.539415 185.78125,157.526902 C186.4875,156.150462 186.89375,154.592583 186.89375,152.940855 C186.89375,147.422583 182.40625,142.930384 176.89375,142.930384 C171.38125,142.930384 166.89375,147.422583 166.89375,152.940855 C166.89375,154.58007 167.2875,156.131693 167.99375,157.495619 C168.00625,157.514389 168.0125,157.526902 168.01875,157.545672 C168.1125,157.714598 168.1625,157.914808 168.1625,158.121274 C168.1625,158.796981 167.63125,159.347557 166.9625,159.372583 L135.7375,159.378839 C135.9625,159.729206 136.16875,160.085829 136.35,160.461221 C137.1,161.962792 137.51875,163.658316 137.51875,165.453944 C137.51875,171.666693 132.475,176.715724 126.26875,176.715724 C120.06875,176.715724 115.01875,171.666693 115.01875,165.453944 C115.01875,163.664572 115.4375,161.969049 116.1875,160.467478 C116.36875,160.092085 116.575,159.729206 116.8,159.378839 L84.0125,159.378839 C83.3375,159.36007 82.8,158.809494 82.8,158.12753 C82.8,157.958604 82.83125,157.80219 82.89375,157.652033 C82.89375,157.645776 82.9,157.63952 82.9,157.633263 C82.90625,157.627007 82.90625,157.62075 82.9125,157.608237 C83.4625,156.369441 83.76875,154.999258 83.76875,153.566509 C83.76875,148.048237 79.28125,143.556038 73.76875,143.556038 C68.25625,143.556038 63.76875,148.048237 63.76875,153.566509 C63.76875,154.999258 64.075,156.369441 64.625,157.608237 C64.6875,157.758394 64.73125,157.952347 64.73125,158.121274 C64.73125,158.809494 64.16875,159.378839 63.475,159.378839 L33.93125,159.378839 C34.1125,159.754232 34.26875,160.142138 34.40625,160.5363 C34.80625,161.681248 35.01875,162.920043 35.01875,164.202635 C35.01875,170.415384 29.975,175.464415 23.76875,175.464415 C17.56875,175.464415 12.51875,170.415384 12.51875,164.202635 C12.51875,162.920043 12.73125,161.681248 13.13125,160.5363 C13.26875,160.142138 13.425,159.754232 13.60625,159.378839 L0.206875,159.378839 L0.206875,158.12753 L13.53125,158.12753 L13.60625,158.12753 C14.29375,158.12753 14.85625,158.690619 14.85625,159.378839 C14.85625,159.579049 14.8125,159.760488 14.73125,159.923159 C14.725,159.941928 14.7125,159.960698 14.70625,159.979468 C14.1,161.268316 13.76875,162.694808 13.76875,164.202635 C13.76875,169.720907 18.25625,174.213106 23.76875,174.213106 C29.28125,174.213106 33.76875,169.720907 33.76875,164.202635 C33.76875,162.694808 33.43125,161.262059 32.83125,159.979468 C32.73125,159.804284 32.68125,159.591562 32.68125,159.378839 C32.68125,158.684363 33.2375,158.12753 33.93125,158.12753 L33.94375,158.12753 L63.4875,158.12753 C63.31875,157.745881 63.16875,157.351719 63.04375,156.9513 C62.7,155.881431 62.51875,154.748996 62.51875,153.566509 C62.51875,147.353761 67.56875,142.304729 73.76875,142.304729 C79.975,142.304729 85.01875,147.353761 85.01875,153.566509 C85.01875,154.748996 84.8375,155.887687 84.49375,156.957557 C84.36875,157.357975 84.21875,157.745881 84.05,158.12753 L116.6875,158.12753 C116.71875,158.121274 116.75625,158.121274 116.79375,158.121274 C117.4875,158.121274 118.04375,158.684363 118.04375,159.372583 C118.04375,159.622845 117.975,159.854337 117.84375,160.04829 C117.84375,160.060803 117.84375,160.073316 117.8375,160.079572 C116.84375,161.637452 116.26875,163.476876 116.26875,165.453944 C116.26875,170.972216 120.75625,175.464415 126.26875,175.464415 C131.78125,175.464415 136.26875,170.972216 136.26875,165.453944 C136.26875,163.526928 135.725,161.737557 134.78125,160.21096 C134.5875,159.991981 134.46875,159.70418 134.46875,159.378839 C134.46875,158.690619 135.025,158.12753 135.71875,158.12753 L135.73125,158.12753 L166.9125,158.12753 C166.71875,157.758394 166.55,157.376745 166.39375,156.988839 C165.90625,155.731274 165.64375,154.367347 165.64375,152.940855 C165.64375,146.728106 170.69375,141.679075 176.89375,141.679075 C183.1,141.679075 188.14375,146.728106 188.14375,152.940855 C188.14375,154.367347 187.88125,155.731274 187.39375,156.982583 C187.24375,157.376745 187.06875,157.758394 186.875,158.12753 L218.5625,158.12753 C218.59375,158.121274 218.63125,158.121274 218.66875,158.121274 C219.3625,158.121274 219.91875,158.684363 219.91875,159.372583 C219.91875,159.622845 219.85,159.854337 219.71875,160.04829 C219.71875,160.060803 219.71875,160.073316 219.7125,160.079572 C218.71875,161.637452 218.14375,163.476876 218.14375,165.453944 C218.14375,170.972216 222.63125,175.464415 228.14375,175.464415 C233.65625,175.464415 238.14375,170.972216 238.14375,165.453944 C238.14375,163.526928 237.6,161.737557 236.65625,160.21096 C236.4625,159.991981 236.34375,159.70418 236.34375,159.378839 C236.34375,158.690619 236.9,158.12753 237.59375,158.12753 L237.60625,158.12753 L269.4125,158.12753 C269.21875,157.758394 269.05,157.376745 268.89375,156.988839 C268.40625,155.731274 268.14375,154.367347 268.14375,152.940855 C268.14375,146.728106 273.19375,141.679075 279.39375,141.679075 C285.6,141.679075 290.64375,146.728106 290.64375,152.940855 C290.64375,154.367347 290.38125,155.731274 289.89375,156.982583 C289.74375,157.376745 289.56875,157.758394 289.375,158.12753 L321.6875,158.12753 C321.71875,158.121274 321.75625,158.121274 321.79375,158.121274 C322.4875,158.121274 323.04375,158.684363 323.04375,159.372583 C323.04375,159.622845 322.975,159.854337 322.84375,160.04829 C322.84375,160.060803 322.84375,160.073316 322.8375,160.079572 C321.84375,161.637452 321.26875,163.476876 321.26875,165.453944 C321.26875,170.972216 325.75625,175.464415 331.26875,175.464415 C336.78125,175.464415 341.26875,170.972216 341.26875,165.453944 C341.26875,163.526928 340.725,161.737557 339.78125,160.21096 C339.5875,159.991981 339.46875,159.70418 339.46875,159.378839 C339.46875,158.690619 340.025,158.12753 340.71875,158.12753 L340.73125,158.12753 L358.9375,158.12753 L358.9375,159.378839 Z" id="path-15"></path>
-					<filter x="-0.1%" y="-4.3%" width="100.6%" height="105.7%" filterUnits="objectBoundingBox" id="filter-16">
-						<feOffset dx="1" dy="-1" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.6 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-					<path d="M358.9375,105.891453 L341.503125,105.891453 C341.728125,105.541086 341.934375,105.184463 342.115625,104.80907 C342.865625,103.3075 343.284375,101.611976 343.284375,99.8163479 C343.284375,93.6035992 338.240625,88.5545678 332.034375,88.5545678 C325.834375,88.5545678 320.784375,93.6035992 320.784375,99.8163479 C320.784375,101.60572 321.203125,103.301243 321.953125,104.802814 C322.134375,105.178207 322.340625,105.541086 322.565625,105.891453 L290.128125,105.891453 C289.446875,105.910222 288.896875,106.460798 288.896875,107.149018 C288.896875,107.349227 288.940625,107.530667 289.028125,107.699594 C289.034375,107.712107 289.040625,107.730877 289.046875,107.74339 C289.753125,109.11983 290.159375,110.677709 290.159375,112.329437 C290.159375,117.847709 285.671875,122.339908 280.159375,122.339908 C274.646875,122.339908 270.159375,117.847709 270.159375,112.329437 C270.159375,110.690222 270.553125,109.138599 271.259375,107.774673 C271.271875,107.755903 271.278125,107.74339 271.284375,107.72462 C271.378125,107.555693 271.428125,107.355484 271.428125,107.149018 C271.428125,106.473311 270.896875,105.922735 270.228125,105.897709 L238.378125,105.891453 C238.603125,105.541086 238.809375,105.184463 238.990625,104.80907 C239.740625,103.3075 240.159375,101.611976 240.159375,99.8163479 C240.159375,93.6035992 235.115625,88.5545678 228.909375,88.5545678 C222.709375,88.5545678 217.659375,93.6035992 217.659375,99.8163479 C217.659375,101.60572 218.078125,103.301243 218.828125,104.802814 C219.009375,105.178207 219.215625,105.541086 219.440625,105.891453 L187.628125,105.891453 C186.946875,105.910222 186.396875,106.460798 186.396875,107.149018 C186.396875,107.349227 186.440625,107.530667 186.528125,107.699594 C186.534375,107.712107 186.540625,107.730877 186.546875,107.74339 C187.253125,109.11983 187.659375,110.677709 187.659375,112.329437 C187.659375,117.847709 183.171875,122.339908 177.659375,122.339908 C172.146875,122.339908 167.659375,117.847709 167.659375,112.329437 C167.659375,110.690222 168.053125,109.138599 168.759375,107.774673 C168.771875,107.755903 168.778125,107.74339 168.784375,107.72462 C168.878125,107.555693 168.928125,107.355484 168.928125,107.149018 C168.928125,106.473311 168.396875,105.922735 167.728125,105.897709 L136.503125,105.891453 C136.728125,105.541086 136.934375,105.184463 137.115625,104.80907 C137.865625,103.3075 138.284375,101.611976 138.284375,99.8163479 C138.284375,93.6035992 133.240625,88.5545678 127.034375,88.5545678 C120.834375,88.5545678 115.784375,93.6035992 115.784375,99.8163479 C115.784375,101.60572 116.203125,103.301243 116.953125,104.802814 C117.134375,105.178207 117.340625,105.541086 117.565625,105.891453 L84.778125,105.891453 C84.103125,105.910222 83.565625,106.460798 83.565625,107.142762 C83.565625,107.311688 83.596875,107.468102 83.659375,107.618259 C83.659375,107.624515 83.665625,107.630772 83.665625,107.637029 C83.671875,107.643285 83.671875,107.649542 83.678125,107.662055 C84.228125,108.900851 84.534375,110.271034 84.534375,111.703782 C84.534375,117.222055 80.046875,121.714254 74.534375,121.714254 C69.021875,121.714254 64.534375,117.222055 64.534375,111.703782 C64.534375,110.271034 64.840625,108.900851 65.390625,107.662055 C65.453125,107.511898 65.496875,107.317945 65.496875,107.149018 C65.496875,106.460798 64.934375,105.891453 64.240625,105.891453 L34.696875,105.891453 C34.878125,105.51606 35.034375,105.128154 35.171875,104.733992 C35.571875,103.589044 35.784375,102.350248 35.784375,101.067657 C35.784375,94.8549081 30.740625,89.8058767 24.534375,89.8058767 C18.334375,89.8058767 13.284375,94.8549081 13.284375,101.067657 C13.284375,102.350248 13.496875,103.589044 13.896875,104.733992 C14.034375,105.128154 14.190625,105.51606 14.371875,105.891453 L0.189925641,105.891453 L0.189925641,107.142762 L14.296875,107.142762 L14.371875,107.142762 C15.059375,107.142762 15.621875,106.579673 15.621875,105.891453 C15.621875,105.691243 15.578125,105.509803 15.496875,105.347133 C15.490625,105.328364 15.478125,105.309594 15.471875,105.290824 C14.865625,104.001976 14.534375,102.575484 14.534375,101.067657 C14.534375,95.5493846 19.021875,91.0571856 24.534375,91.0571856 C30.046875,91.0571856 34.534375,95.5493846 34.534375,101.067657 C34.534375,102.575484 34.196875,104.008233 33.596875,105.290824 C33.496875,105.466008 33.446875,105.67873 33.446875,105.891453 C33.446875,106.585929 34.003125,107.142762 34.696875,107.142762 L34.709375,107.142762 L64.253125,107.142762 C64.084375,107.524411 63.934375,107.918573 63.809375,108.318992 C63.465625,109.388861 63.284375,110.521296 63.284375,111.703782 C63.284375,117.916531 68.334375,122.965563 74.534375,122.965563 C80.740625,122.965563 85.784375,117.916531 85.784375,111.703782 C85.784375,110.521296 85.603125,109.382604 85.259375,108.312735 C85.134375,107.912316 84.984375,107.524411 84.815625,107.142762 L117.453125,107.142762 C117.484375,107.149018 117.521875,107.149018 117.559375,107.149018 C118.253125,107.149018 118.809375,106.585929 118.809375,105.897709 C118.809375,105.647447 118.740625,105.415955 118.609375,105.222002 C118.609375,105.209489 118.609375,105.196976 118.603125,105.19072 C117.609375,103.63284 117.034375,101.793416 117.034375,99.8163479 C117.034375,94.2980757 121.521875,89.8058767 127.034375,89.8058767 C132.546875,89.8058767 137.034375,94.2980757 137.034375,99.8163479 C137.034375,101.743364 136.490625,103.532735 135.546875,105.059332 C135.353125,105.278311 135.234375,105.566112 135.234375,105.891453 C135.234375,106.579673 135.790625,107.142762 136.484375,107.142762 L136.496875,107.142762 L167.678125,107.142762 C167.484375,107.511898 167.315625,107.893547 167.159375,108.281453 C166.671875,109.539018 166.409375,110.902945 166.409375,112.329437 C166.409375,118.542186 171.459375,123.591217 177.659375,123.591217 C183.865625,123.591217 188.909375,118.542186 188.909375,112.329437 C188.909375,110.902945 188.646875,109.539018 188.159375,108.287709 C188.009375,107.893547 187.834375,107.511898 187.640625,107.142762 L219.328125,107.142762 C219.359375,107.149018 219.396875,107.149018 219.434375,107.149018 C220.128125,107.149018 220.684375,106.585929 220.684375,105.897709 C220.684375,105.647447 220.615625,105.415955 220.484375,105.222002 C220.484375,105.209489 220.484375,105.196976 220.478125,105.19072 C219.484375,103.63284 218.909375,101.793416 218.909375,99.8163479 C218.909375,94.2980757 223.396875,89.8058767 228.909375,89.8058767 C234.421875,89.8058767 238.909375,94.2980757 238.909375,99.8163479 C238.909375,101.743364 238.365625,103.532735 237.421875,105.059332 C237.228125,105.278311 237.109375,105.566112 237.109375,105.891453 C237.109375,106.579673 237.665625,107.142762 238.359375,107.142762 L238.371875,107.142762 L270.178125,107.142762 C269.984375,107.511898 269.815625,107.893547 269.659375,108.281453 C269.171875,109.539018 268.909375,110.902945 268.909375,112.329437 C268.909375,118.542186 273.959375,123.591217 280.159375,123.591217 C286.365625,123.591217 291.409375,118.542186 291.409375,112.329437 C291.409375,110.902945 291.146875,109.539018 290.659375,108.287709 C290.509375,107.893547 290.334375,107.511898 290.140625,107.142762 L322.453125,107.142762 C322.484375,107.149018 322.521875,107.149018 322.559375,107.149018 C323.253125,107.149018 323.809375,106.585929 323.809375,105.897709 C323.809375,105.647447 323.740625,105.415955 323.609375,105.222002 C323.609375,105.209489 323.609375,105.196976 323.603125,105.19072 C322.609375,103.63284 322.034375,101.793416 322.034375,99.8163479 C322.034375,94.2980757 326.521875,89.8058767 332.034375,89.8058767 C337.546875,89.8058767 342.034375,94.2980757 342.034375,99.8163479 C342.034375,101.743364 341.490625,103.532735 340.546875,105.059332 C340.353125,105.278311 340.234375,105.566112 340.234375,105.891453 C340.234375,106.579673 340.790625,107.142762 341.484375,107.142762 L341.496875,107.142762 L358.9375,107.142762 L358.9375,105.891453 Z" id="path-17"></path>
-					<filter x="-0.1%" y="-4.3%" width="100.6%" height="105.7%" filterUnits="objectBoundingBox" id="filter-18">
-						<feOffset dx="1" dy="-1" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.6 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-					<path d="M358.9375,53.0175827 L341.3625,53.0175827 C341.5875,53.3679492 341.79375,53.7245723 341.975,54.0999649 C342.725,55.6015356 343.14375,57.2970592 343.14375,59.0926874 C343.14375,65.3054361 338.1,70.3544675 331.89375,70.3544675 C325.69375,70.3544675 320.64375,65.3054361 320.64375,59.0926874 C320.64375,57.3033157 321.0625,55.6077921 321.8125,54.1062215 C321.99375,53.7308288 322.2,53.3679492 322.425,53.0175827 L289.9875,53.0175827 L289.98125,53.0113262 C289.30625,52.9988131 288.75625,52.4482372 288.75625,51.7600173 C288.75625,51.5598079 288.8,51.3783681 288.8875,51.2094414 C288.89375,51.1969283 288.9,51.1781586 288.90625,51.1656455 C289.6125,49.7892058 290.01875,48.2313262 290.01875,46.5795984 C290.01875,41.0613262 285.53125,36.5691272 280.01875,36.5691272 C274.50625,36.5691272 270.01875,41.0613262 270.01875,46.5795984 C270.01875,48.2188131 270.4125,49.7704361 271.11875,51.1343628 C271.13125,51.1531325 271.1375,51.1656455 271.14375,51.1844152 C271.2375,51.3533419 271.2875,51.5535513 271.2875,51.7600173 C271.2875,52.4357241 270.75625,52.9863 270.0875,53.0113262 L270.08125,53.0175827 L238.2375,53.0175827 C238.4625,53.3679492 238.66875,53.7245723 238.85,54.0999649 C239.6,55.6015356 240.01875,57.2970592 240.01875,59.0926874 C240.01875,65.3054361 234.975,70.3544675 228.76875,70.3544675 C222.56875,70.3544675 217.51875,65.3054361 217.51875,59.0926874 C217.51875,57.3033157 217.9375,55.6077921 218.6875,54.1062215 C218.86875,53.7308288 219.075,53.3679492 219.3,53.0175827 L187.4875,53.0175827 L187.48125,53.0113262 C186.80625,52.9988131 186.25625,52.4482372 186.25625,51.7600173 C186.25625,51.5598079 186.3,51.3783681 186.3875,51.2094414 C186.39375,51.1969283 186.4,51.1781586 186.40625,51.1656455 C187.1125,49.7892058 187.51875,48.2313262 187.51875,46.5795984 C187.51875,41.0613262 183.03125,36.5691272 177.51875,36.5691272 C172.00625,36.5691272 167.51875,41.0613262 167.51875,46.5795984 C167.51875,48.2188131 167.9125,49.7704361 168.61875,51.1343628 C168.63125,51.1531325 168.6375,51.1656455 168.64375,51.1844152 C168.7375,51.3533419 168.7875,51.5535513 168.7875,51.7600173 C168.7875,52.4357241 168.25625,52.9863 167.5875,53.0113262 L167.58125,53.0175827 L136.3625,53.0175827 C136.5875,53.3679492 136.79375,53.7245723 136.975,54.0999649 C137.725,55.6015356 138.14375,57.2970592 138.14375,59.0926874 C138.14375,65.3054361 133.1,70.3544675 126.89375,70.3544675 C120.69375,70.3544675 115.64375,65.3054361 115.64375,59.0926874 C115.64375,57.3033157 116.0625,55.6077921 116.8125,54.1062215 C116.99375,53.7308288 117.2,53.3679492 117.425,53.0175827 L84.6375,53.0175827 C83.9625,52.9988131 83.425,52.4482372 83.425,51.7662738 C83.425,51.5973471 83.45625,51.4409335 83.51875,51.2907764 C83.51875,51.2845199 83.525,51.2782634 83.525,51.2720068 C83.53125,51.2657503 83.53125,51.2594937 83.5375,51.2469806 C84.0875,50.0081848 84.39375,48.6380016 84.39375,47.2052529 C84.39375,41.6869806 79.90625,37.1947817 74.39375,37.1947817 C68.88125,37.1947817 64.39375,41.6869806 64.39375,47.2052529 C64.39375,48.6380016 64.7,50.0081848 65.25,51.2469806 C65.3125,51.3971377 65.35625,51.5910906 65.35625,51.7600173 C65.35625,52.4482372 64.79375,53.0175827 64.1,53.0175827 L34.55625,53.0175827 C34.7375,53.3929754 34.89375,53.7808812 35.03125,54.1750435 C35.43125,55.3199911 35.64375,56.5587869 35.64375,57.8413785 C35.64375,64.0541272 30.6,69.1031586 24.39375,69.1031586 C18.19375,69.1031586 13.14375,64.0541272 13.14375,57.8413785 C13.14375,56.5587869 13.35625,55.3199911 13.75625,54.1750435 C13.89375,53.7808812 14.05,53.3929754 14.23125,53.0175827 L0.1875,53.0175827 L0.1875,51.7662738 L14.15625,51.7662738 L14.23125,51.7662738 C14.91875,51.7662738 15.48125,52.3293628 15.48125,53.0175827 C15.48125,53.2177921 15.4375,53.3992319 15.35625,53.5619021 C15.35,53.5806717 15.3375,53.5994414 15.33125,53.618211 C14.725,54.9070592 14.39375,56.3335513 14.39375,57.8413785 C14.39375,63.3596508 18.88125,67.8518497 24.39375,67.8518497 C29.90625,67.8518497 34.39375,63.3596508 34.39375,57.8413785 C34.39375,56.3335513 34.05625,54.9008026 33.45625,53.618211 C33.35625,53.4430277 33.30625,53.2303052 33.30625,53.0175827 C33.30625,52.3231063 33.8625,51.7662738 34.55625,51.7662738 L34.56875,51.7662738 L64.1125,51.7662738 C63.94375,51.3846246 63.79375,50.9904623 63.66875,50.5900435 C63.325,49.5201743 63.14375,48.3877398 63.14375,47.2052529 C63.14375,40.9925042 68.19375,35.9434728 74.39375,35.9434728 C80.6,35.9434728 85.64375,40.9925042 85.64375,47.2052529 C85.64375,48.3877398 85.4625,49.5264309 85.11875,50.5963 C84.99375,50.9967188 84.84375,51.3846246 84.675,51.7662738 L117.3125,51.7662738 C117.34375,51.7600173 117.38125,51.7600173 117.41875,51.7600173 C118.1125,51.7600173 118.66875,52.3231063 118.66875,53.0113262 C118.66875,53.261588 118.6,53.4930801 118.46875,53.687033 C118.46875,53.6995461 118.46875,53.7120592 118.4625,53.7183157 C117.46875,55.2761953 116.89375,57.1156194 116.89375,59.0926874 C116.89375,64.6109597 121.38125,69.1031586 126.89375,69.1031586 C132.40625,69.1031586 136.89375,64.6109597 136.89375,59.0926874 C136.89375,57.1656717 136.35,55.3763 135.40625,53.8497031 C135.2125,53.6307241 135.09375,53.342923 135.09375,53.0175827 C135.09375,52.3293628 135.65,51.7662738 136.34375,51.7662738 L136.35625,51.7662738 L167.5375,51.7662738 C167.34375,51.3971377 167.175,51.0154885 167.01875,50.6275827 C166.53125,49.3700173 166.26875,48.0060906 166.26875,46.5795984 C166.26875,40.3668497 171.31875,35.3178183 177.51875,35.3178183 C183.725,35.3178183 188.76875,40.3668497 188.76875,46.5795984 C188.76875,48.0060906 188.50625,49.3700173 188.01875,50.6213262 C187.86875,51.0154885 187.69375,51.3971377 187.5,51.7662738 L219.1875,51.7662738 C219.21875,51.7600173 219.25625,51.7600173 219.29375,51.7600173 C219.9875,51.7600173 220.54375,52.3231063 220.54375,53.0113262 C220.54375,53.261588 220.475,53.4930801 220.34375,53.687033 C220.34375,53.6995461 220.34375,53.7120592 220.3375,53.7183157 C219.34375,55.2761953 218.76875,57.1156194 218.76875,59.0926874 C218.76875,64.6109597 223.25625,69.1031586 228.76875,69.1031586 C234.28125,69.1031586 238.76875,64.6109597 238.76875,59.0926874 C238.76875,57.1656717 238.225,55.3763 237.28125,53.8497031 C237.0875,53.6307241 236.96875,53.342923 236.96875,53.0175827 C236.96875,52.3293628 237.525,51.7662738 238.21875,51.7662738 L238.23125,51.7662738 L270.0375,51.7662738 C269.84375,51.3971377 269.675,51.0154885 269.51875,50.6275827 C269.03125,49.3700173 268.76875,48.0060906 268.76875,46.5795984 C268.76875,40.3668497 273.81875,35.3178183 280.01875,35.3178183 C286.225,35.3178183 291.26875,40.3668497 291.26875,46.5795984 C291.26875,48.0060906 291.00625,49.3700173 290.51875,50.6213262 C290.36875,51.0154885 290.19375,51.3971377 290,51.7662738 L322.3125,51.7662738 C322.34375,51.7600173 322.38125,51.7600173 322.41875,51.7600173 C323.1125,51.7600173 323.66875,52.3231063 323.66875,53.0113262 C323.66875,53.261588 323.6,53.4930801 323.46875,53.687033 C323.46875,53.6995461 323.46875,53.7120592 323.4625,53.7183157 C322.46875,55.2761953 321.89375,57.1156194 321.89375,59.0926874 C321.89375,64.6109597 326.38125,69.1031586 331.89375,69.1031586 C337.40625,69.1031586 341.89375,64.6109597 341.89375,59.0926874 C341.89375,57.1656717 341.35,55.3763 340.40625,53.8497031 C340.2125,53.6307241 340.09375,53.342923 340.09375,53.0175827 C340.09375,52.3293628 340.65,51.7662738 341.34375,51.7662738 L341.35625,51.7662738 L358.9375,51.7662738 L358.9375,53.0175827 Z" id="path-19"></path>
-					<filter x="-0.1%" y="-4.3%" width="100.6%" height="105.7%" filterUnits="objectBoundingBox" id="filter-20">
-						<feOffset dx="1" dy="-1" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.6 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-					<path d="M38.4908512,3.3508804 C38.4173404,3.62471732 38.3399363,3.8975746 38.2585,4.16936126 C38.2129376,4.32138781 38.1658012,4.47077516 38.117157,4.61767193 C36.9208274,3.80961789 35.8872081,2.77754239 35.0769886,1.58190671 C35.405465,1.36436281 35.7634163,1.14992006 36.1534376,0.940361886 C36.7924082,1.86876559 37.5837067,2.68444502 38.4908512,3.3508804 Z M47.8377825,25.2325066 C48.4218197,25.5788073 48.800276,25.6981002 49.0378962,25.6787616 C53.0068522,27.4196958 55.784375,31.3892235 55.784375,35.999594 C55.784375,37.7952223 55.365625,39.4907458 54.615625,40.9923165 C54.434375,41.3677092 54.228125,41.7243322 54.003125,42.0746987 L68.2124375,42.0775458 L68.2124375,25.1825916 C68.2311875,24.5068848 68.7811875,23.968822 69.4624375,23.968822 C69.6311875,23.968822 69.7874375,24.0001047 69.9374375,24.0626702 C69.9436875,24.0626702 69.9499375,24.0689267 69.9561875,24.0689267 C69.9624375,24.0751832 69.9686875,24.0751832 69.9811875,24.0814398 C71.2186875,24.6320157 72.5874375,24.9385864 74.0186875,24.9385864 C78.6583595,24.9385864 82.5719235,21.7563214 83.6949227,17.4584101 C84.0725526,17.654765 84.4523163,17.8467201 84.8341176,18.0342223 C83.4848916,22.7397932 79.1489691,26.1898953 74.0186875,26.1898953 C72.8374375,26.1898953 71.6999375,26.0084555 70.6311875,25.6643455 C70.2311875,25.5392147 69.8436875,25.3890576 69.4624375,25.2201309 L69.4624375,42.0777963 L85.228125,42.0809552 C85.896875,42.1059814 86.428125,42.6565573 86.428125,43.3322641 C86.428125,43.5387301 86.378125,43.7389395 86.284375,43.9078662 C86.278125,43.9266359 86.271875,43.939149 86.259375,43.9579186 C85.553125,45.3218453 85.159375,46.8734683 85.159375,48.512683 C85.159375,54.0309552 89.646875,58.5231542 95.159375,58.5231542 C100.671875,58.5231542 105.159375,54.0309552 105.159375,48.512683 C105.159375,46.8609552 104.753125,45.3030757 104.046875,43.9266359 C104.040625,43.9141228 104.034375,43.8953531 104.028125,43.8828401 C103.940625,43.7139134 103.896875,43.5324736 103.896875,43.3322641 C103.896875,42.6440442 104.446875,42.0934683 105.128125,42.0746987 L118.574938,42.0746987 L118.574938,25.2201309 C118.193688,25.3890576 117.806188,25.5392147 117.406188,25.6643455 C116.337438,26.0084555 115.199938,26.1898953 114.018688,26.1898953 C111.248282,26.1898953 108.70952,25.1838066 106.747637,23.5172249 C107.465339,23.535806 108.183942,23.5388882 108.90298,23.5262128 C110.400746,24.422851 112.150878,24.9385864 114.018688,24.9385864 C115.449938,24.9385864 116.818688,24.6320157 118.056188,24.0814398 C118.068688,24.0751832 118.074938,24.0751832 118.081188,24.0689267 C118.087438,24.0689267 118.093688,24.0626702 118.099938,24.0626702 C118.249938,24.0001047 118.406188,23.968822 118.574938,23.968822 C119.256188,23.968822 119.806188,24.5068848 119.824938,25.1825916 L119.824938,42.0746987 L136.940625,42.0746987 C136.715625,41.7243322 136.509375,41.3614526 136.328125,40.9860599 C135.578125,39.4844893 135.159375,37.7889657 135.159375,35.999594 C135.159375,29.9966294 139.874089,25.0801192 145.785398,24.7549349 C146.823412,25.2276851 147.844219,25.7399889 148.84546,26.2897186 C148.065392,26.093408 147.249281,25.9891228 146.409375,25.9891228 C140.896875,25.9891228 136.409375,30.4813217 136.409375,35.999594 C136.409375,37.976662 136.984375,39.8160861 137.978125,41.3739657 C137.984375,41.3802223 137.984375,41.3927353 137.984375,41.4052484 C138.115625,41.5992013 138.184375,41.8306935 138.184375,42.0809552 C138.184375,42.7691751 137.628125,43.3322641 136.934375,43.3322641 C136.896875,43.3322641 136.859375,43.3322641 136.828125,43.3260076 L119.824938,43.3260076 L119.824938,59.255733 C120.174938,59.0304974 120.537438,58.8240314 120.912438,58.6425916 C122.412438,57.8918063 124.106188,57.4726178 125.893688,57.4726178 C132.099938,57.4726178 137.143688,62.5279058 137.143688,68.7343979 C137.143688,74.7551156 132.406915,79.6829433 126.465731,79.9818215 C125.992581,79.5893866 125.527593,79.1622472 125.074089,78.7116884 C125.344213,78.7336673 125.617535,78.7448691 125.893688,78.7448691 C131.406188,78.7448691 135.893688,74.2526702 135.893688,68.7343979 C135.893688,63.2161257 131.406188,58.7239267 125.893688,58.7239267 C123.918688,58.7239267 122.081188,59.2995288 120.524938,60.2943194 C120.518688,60.3005759 120.506188,60.3005759 120.493688,60.3005759 C120.299938,60.4319634 120.068688,60.5007853 119.818688,60.5007853 C119.131188,60.5007853 118.568688,59.9439529 118.568688,59.2494764 C118.568688,59.2119372 118.568688,59.1743979 118.574938,59.1431152 L118.574938,43.3260076 L105.140625,43.3260076 C105.334375,43.6951437 105.509375,44.0767929 105.659375,44.4709552 C106.146875,45.7222641 106.409375,47.0861908 106.409375,48.512683 C106.409375,54.7254317 101.365625,59.7744631 95.159375,59.7744631 C88.959375,59.7744631 83.909375,54.7254317 83.909375,48.512683 C83.909375,47.0861908 84.171875,45.7222641 84.659375,44.4646987 C84.815625,44.0767929 84.984375,43.6951437 85.178125,43.3260076 L69.4624375,43.3260076 L69.4624375,59.1431152 C69.4686875,59.1743979 69.4686875,59.2119372 69.4686875,59.2494764 C69.4686875,59.9439529 68.9061875,60.5007853 68.2186875,60.5007853 C67.9686875,60.5007853 67.7374375,60.4319634 67.5436875,60.3005759 C67.5311875,60.3005759 67.5186875,60.3005759 67.5124375,60.2943194 C65.9561875,59.2995288 64.1186875,58.7239267 62.1436875,58.7239267 C56.6311875,58.7239267 52.1436875,63.2161257 52.1436875,68.7343979 C52.1436875,74.2526702 56.6311875,78.7448691 62.1436875,78.7448691 C64.0686875,78.7448691 65.8561875,78.2005497 67.3811875,77.2558115 C67.5999375,77.0618586 67.8874375,76.9429843 68.2124375,76.9429843 C68.8999375,76.9429843 69.4624375,77.4998168 69.4624375,78.1942932 L69.4624375,78.2068063 L69.4624375,94.3107764 L74.8576419,94.3107764 C74.8660776,94.7091883 74.8728046,95.126509 74.8770043,95.5577494 L69.4624375,95.5588344 L69.4624375,111.923325 C69.8311875,111.729372 70.2124375,111.560445 70.5999375,111.404031 C71.5188776,111.047054 72.4946705,110.810597 73.5101918,110.709354 C73.4940747,111.114331 73.4808977,111.536432 73.4703999,111.9728 C72.2661775,112.113897 71.1261045,112.469995 70.0936875,113.005707 C70.0749375,113.01822 70.0624375,113.024476 70.0436875,113.030733 C69.8749375,113.124581 69.6749375,113.174634 69.4686875,113.174634 C68.7936875,113.174634 68.2436875,112.642827 68.2186875,111.973377 L68.2124375,111.96712 L68.2124375,95.5590848 L63.8433162,95.5599603 C63.8253333,95.1092969 63.8253175,94.6942326 63.8386213,94.3107764 L68.2124375,94.3107764 L68.2124375,78.2130628 C67.8624375,78.4382984 67.5061875,78.6447644 67.1311875,78.8262042 C65.6311875,79.5769895 63.9374375,79.996178 62.1436875,79.996178 C55.9374375,79.996178 50.8936875,74.9471466 50.8936875,68.7343979 C50.8936875,62.5279058 55.9374375,57.4726178 62.1436875,57.4726178 C63.9311875,57.4726178 65.6249375,57.8918063 67.1249375,58.6425916 C67.4999375,58.8240314 67.8624375,59.0304974 68.2124375,59.255733 L68.2124375,43.3260076 L53.996875,43.3260076 L53.984375,43.3260076 C53.290625,43.3260076 52.734375,42.7629186 52.734375,42.0746987 C52.734375,41.7493584 52.853125,41.4615573 53.046875,41.2425783 C53.990625,39.7159814 54.534375,37.9266097 54.534375,35.999594 C54.534375,30.4813217 50.046875,25.9891228 44.534375,25.9891228 C39.021875,25.9891228 34.534375,30.4813217 34.534375,35.999594 C34.534375,37.976662 35.109375,39.8160861 36.103125,41.3739657 C36.109375,41.3802223 36.109375,41.3927353 36.109375,41.4052484 C36.240625,41.5992013 36.309375,41.8306935 36.309375,42.0809552 C36.309375,42.7691751 35.753125,43.3322641 35.059375,43.3322641 C35.021875,43.3322641 34.984375,43.3322641 34.953125,43.3260076 L34.7910412,43.3260076 C34.7265208,42.909263 34.6649728,42.4921593 34.6063373,42.0746987 L35.065625,42.0746987 C34.8445405,41.7304293 34.641559,41.3740785 34.4626101,41.0056464 C34.1990981,38.954542 34.0044606,36.8948975 33.871625,34.8269924 L33.875375,34.9602568 C33.8685,34.9077018 33.859125,34.8538955 33.846625,34.8007149 C33.810375,34.6417987 33.751,34.4847594 33.65975,34.3577516 C33.454125,34.5585866 33.46475,34.8901835 33.4335,35.1629688 C33.3944029,35.501244 33.3447088,35.8287046 33.2853063,36.1456791 C33.284686,36.0970551 33.284375,36.0483594 33.284375,35.999594 C33.284375,29.7868453 38.334375,24.7378139 44.534375,24.7378139 C45.6835845,24.7378139 46.7929345,24.9109342 47.8377825,25.2325066 Z M154.306886,29.8666105 C155.299948,30.6506393 156.204834,31.4560689 157.028686,32.277371 C157.437183,33.4428973 157.659375,34.695613 157.659375,35.999594 C157.659375,37.7952223 157.240625,39.4907458 156.490625,40.9923165 C156.309375,41.3677092 156.103125,41.7243322 155.878125,42.0746987 L163.320542,42.0761607 C163.47039,42.4942207 163.607591,42.9110466 163.732958,43.3260076 L155.871875,43.3260076 L155.859375,43.3260076 C155.165625,43.3260076 154.609375,42.7629186 154.609375,42.0746987 C154.609375,41.7493584 154.728125,41.4615573 154.921875,41.2425783 C155.865625,39.7159814 156.409375,37.9266097 156.409375,35.999594 C156.409375,33.6911158 155.624049,31.5622034 154.306886,29.8666105 Z M164.360443,57.4761343 C164.278277,57.9317301 164.193087,58.3536077 164.109692,58.7380367 C163.22578,58.7848575 162.371874,58.9472085 161.562811,59.2102393 C161.3875,58.7309308 161.245841,58.2950151 161.161072,58.0247029 C162.172483,57.6943335 163.246553,57.5038196 164.360443,57.4761343 Z M160.993507,58.080911 C160.869167,58.5822513 160.720086,59.0906008 160.54764,59.6031005 C157.070745,61.1728084 154.643688,64.676128 154.643688,68.7343979 C154.643688,69.1126245 154.664769,69.4860307 154.705817,69.8535006 C154.348276,70.3009235 153.993515,70.723469 153.646564,71.1163485 C153.480922,70.3483839 153.393688,69.5514545 153.393688,68.7343979 C153.393688,63.8051944 156.575061,59.6021193 160.993507,58.080911 Z M142.910788,111.267433 C142.785218,111.674709 142.650707,112.070836 142.506655,112.453398 C138.906132,111.405114 136.043733,108.599669 134.914836,105.028692 C135.571749,102.150177 135.682706,99.3851596 135.469527,96.8414301 C135.499926,96.7775225 135.530919,96.7139514 135.5625,96.6507241 C135.74375,96.2753314 135.95,95.9124518 136.175,95.5620853 L135.333566,95.5620853 C135.278977,95.1376975 135.2159,94.7204112 135.145469,94.3107764 L136.0625,94.3107764 C136.09375,94.3045199 136.13125,94.3045199 136.16875,94.3045199 C136.8625,94.3045199 137.41875,94.8676089 137.41875,95.5558288 C137.41875,95.8060906 137.35,96.0375827 137.21875,96.2315356 C137.21875,96.2440487 137.21875,96.2565618 137.2125,96.2628183 C136.21875,97.8206979 135.64375,99.660122 135.64375,101.63719 C135.64375,106.207374 138.721728,110.073773 142.910788,111.267433 Z M119.824938,137.028172 C119.340777,136.94529 118.926615,136.83634 118.574938,136.705456 L118.574938,131.906728 C118.28596,132.058722 117.989307,132.19919 117.684976,132.325119 C118.059751,131.727191 118.508157,131.166852 118.911935,130.709149 C119.430814,130.856438 119.808356,131.329454 119.818688,131.887958 L119.824938,131.894215 L119.824938,137.028172 Z M73.9403707,131.900756 C73.9688262,132.458784 73.986125,132.771382 73.986125,132.771382 C73.8137543,132.889733 73.6433535,133.004956 73.4749011,133.117115 C72.4724594,133.013938 71.509921,132.778846 70.6061875,132.426021 C70.2124375,132.275864 69.8311875,132.100681 69.4624375,131.906728 L69.4624375,135.411183 C69.0280047,135.613121 68.6115755,135.789445 68.2124375,135.942248 L68.2124375,131.894215 L68.2186875,131.887958 C68.2311875,131.212251 68.7811875,130.661675 69.4686875,130.661675 C69.6686875,130.661675 69.8499375,130.705471 70.0186875,130.793063 C70.0311875,130.799319 70.0499375,130.805576 70.0624375,130.811832 C71.2403958,131.417508 72.5513789,131.802775 73.9403707,131.900756 Z M30.2926012,42.0746987 C29.8287905,42.5546229 29.3454748,42.9690334 28.8644943,43.3260076 L17.3249375,43.3260076 L17.3249375,56.2360745 C16.8969861,56.6509119 16.479627,57.0089709 16.0749375,57.317418 L16.0749375,43.3260076 L2.315625,43.3260076 C2.484375,43.7076568 2.634375,44.0955626 2.759375,44.4959814 C3.103125,45.5658505 3.284375,46.7045416 3.284375,47.8870285 C3.284375,50.7083062 2.24426893,53.2896057 0.527517733,55.2670857 C0.356014636,54.7741436 0.207777285,54.292805 0.0802835261,53.824095 C1.30782047,52.1617395 2.034375,50.1074479 2.034375,47.8870285 C2.034375,46.4542798 1.728125,45.0840966 1.178125,43.8453008 C1.171875,43.8327877 1.171875,43.8265312 1.165625,43.8202746 C1.165625,43.8140181 1.159375,43.8077615 1.159375,43.801505 C1.096875,43.6513479 1.065625,43.4949343 1.065625,43.3260076 C1.065625,42.6440442 1.603125,42.0934683 2.278125,42.0746987 L16.0749375,42.0746987 L16.0749375,25.2201309 C15.6936875,25.3890576 15.3061875,25.5392147 14.9061875,25.6643455 C13.8374375,26.0084555 12.6999375,26.1898953 11.5186875,26.1898953 C9.91572499,26.1898953 8.39031237,25.8530765 7.00932311,25.2463825 C6.913741,24.7242862 6.86372721,24.2362378 6.85299765,23.7800651 C8.24682251,24.5194891 9.83494159,24.9385864 11.5186875,24.9385864 C12.9499375,24.9385864 14.3186875,24.6320157 15.5561875,24.0814398 C15.5686875,24.0751832 15.5749375,24.0751832 15.5811875,24.0689267 C15.5874375,24.0689267 15.5936875,24.0626702 15.5999375,24.0626702 C15.7499375,24.0001047 15.9061875,23.968822 16.0749375,23.968822 C16.7561875,23.968822 17.3061875,24.5068848 17.3249375,25.1825916 L17.3249375,42.0746987 L30.2926012,42.0746987 Z M6.96057894,6.019929 C6.69712957,5.65189123 6.47023782,5.29913154 6.27494448,4.96614282 C7.84134301,4.13651723 9.62595436,3.66633508 11.5186875,3.66633508 C12.6999375,3.66633508 13.8311875,3.84777487 14.8999375,4.19188482 C15.2999375,4.31701571 15.6936875,4.46717277 16.0749375,4.63609948 L16.0749375,2.26409421 C16.2243303,2.29862406 16.3725061,2.33291784 16.51925,2.36685079 C16.8035969,2.43261179 17.0718582,2.50911792 17.3249375,2.59480405 L17.3249375,4.62358639 C17.3249375,5.31806283 16.7561875,5.88115183 16.0686875,5.88115183 C15.8999375,5.88115183 15.7061875,5.83735602 15.5561875,5.77479058 C14.3186875,5.22421466 12.9499375,4.91764398 11.5186875,4.91764398 C9.87810754,4.91764398 8.32831411,5.31552808 6.96057894,6.019929 Z" id="path-21"></path>
-					<filter x="-0.3%" y="-1.1%" width="101.2%" height="101.5%" filterUnits="objectBoundingBox" id="filter-22">
-						<feOffset dx="1" dy="-1" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feComposite in="shadowOffsetOuter1" in2="SourceAlpha" operator="out" result="shadowOffsetOuter1"></feComposite>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.2 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-					<path d="M30.5031553,57.4213507 C30.1484354,57.9423573 29.934875,58.2456139 29.934875,58.2456139 C29.934875,58.2456139 29.8837922,57.8535091 29.7258291,57.3113461 C29.9817143,57.357935 30.2409573,57.3947376 30.5031553,57.4213507 Z M24.5192743,54.6039415 C24.164071,54.7667517 23.7861778,54.975501 23.3841988,55.2362436 C21.4547151,53.2138804 20.2686875,50.4748364 20.2686875,47.4621466 C20.2686875,42.1154073 24.011854,37.6230222 29.0167685,36.4810082 C29.14345,36.7077141 29.2748617,36.9327541 29.411125,37.1560537 C29.445152,37.2117654 29.5775774,37.3702818 29.7848375,37.6023876 C25.0953153,38.426964 21.5186875,42.5356548 21.5186875,47.4621466 C21.5186875,50.2555415 22.6685949,52.7860084 24.5192743,54.6039415 Z M2.62939916,21.1955283 C2.27834623,20.7135996 1.84365535,20.2802841 1.35704673,19.8567529 C1.28937195,19.675349 1.22640197,19.4920004 1.16875,19.3073209 C0.825,18.2374518 0.64375,17.1050173 0.64375,15.9225304 C0.64375,13.6711926 1.30688983,11.5726682 2.44758777,9.81183996 C2.98496379,9.91183157 3.47257581,9.88063119 3.879875,9.65416099 C4.39970463,9.36501936 4.64296589,8.47118058 4.71971894,7.25408476 C6.05642572,6.14367127 7.65296584,5.33658358 9.40271703,4.9390425 C9.61224139,5.37027452 9.80683736,5.754789 9.98797276,6.09456646 C5.38324024,6.98818811 1.89375,11.0564208 1.89375,15.9225304 C1.89375,17.3552791 2.2,18.7254623 2.75,19.9642581 C2.8125,20.1144152 2.85625,20.3083681 2.85625,20.4772948 C2.85625,20.7430855 2.77235276,20.991146 2.62939916,21.1955283 Z M12.658679,5.94107326 C12.7305705,5.5691191 12.8018254,5.1552123 12.8752662,4.70316124 C15.45648,4.92721044 17.7901952,6.02645153 19.5781579,7.70231817 C19.8445351,8.57986508 20.2326986,9.27554425 20.7835901,9.02570926 C22.2625522,10.932182 23.14375,13.3256991 23.14375,15.9225304 C23.14375,17.1050173 22.9625,18.2437084 22.61875,19.3135775 C22.49375,19.7139963 22.34375,20.1019021 22.175,20.4835513 L26.1857695,20.4835513 C26.1547107,20.9009513 26.1290423,21.3181528 26.1092444,21.7348602 L22.1375,21.7348602 C21.4625,21.7160906 20.925,21.1655147 20.925,20.4835513 C20.925,20.3146246 20.95625,20.158211 21.01875,20.0080539 C21.01875,20.0017974 21.025,19.9955408 21.025,19.9892843 C21.03125,19.9830277 21.03125,19.9767712 21.0375,19.9642581 C21.5875,18.7254623 21.89375,17.3552791 21.89375,15.9225304 C21.89375,10.6618549 17.8154305,6.33369296 12.658679,5.94107326 Z M145.074089,111.24572 C145.527593,111.696279 145.992581,112.123418 146.465731,112.515853 C146.276237,112.525386 146.085518,112.530209 145.893688,112.530209 C144.099938,112.530209 142.406188,112.111021 140.906188,111.360236 C140.531188,111.178796 140.174938,110.97233 139.824938,110.747094 L139.824938,111.101578 C139.357379,110.940164 138.932903,110.792406 138.576372,110.667717 C138.607968,110.001734 139.157557,109.477016 139.824938,109.477016 C140.149938,109.477016 140.437438,109.59589 140.656188,109.789843 C141.962418,110.599053 143.461235,111.114489 145.074089,111.24572 Z M119.271012,111.5095 C118.77259,111.774165 118.27299,112.015947 117.773405,112.236402 C116.717209,111.855398 115.579179,111.647661 114.39375,111.647661 C111.80269,111.647661 109.438084,112.64013 107.659498,114.265127 C107.079643,114.262086 106.514307,114.241061 105.966207,114.205603 C108.029514,111.870398 111.043763,110.396352 114.39375,110.396352 C116.140422,110.396352 117.795016,110.796272 119.271012,111.5095 Z M54.4626101,73.5396778 C54.4594408,73.5331528 54.4562791,73.526624 54.453125,73.5200914 C53.7235646,72.0594427 53.3074453,70.4152687 53.2853063,68.6797105 C53.3447088,68.362736 53.3944029,68.0352754 53.4335,67.6970003 C53.46475,67.4242149 53.454125,67.0926181 53.65975,66.891783 C53.751,67.0187908 53.810375,67.1758301 53.846625,67.3347463 C53.859125,67.387927 53.8685,67.4417332 53.875375,67.4942882 C53.874125,67.4498668 53.872875,67.4054453 53.871625,67.3610238 C54.0044606,69.4289289 54.1990981,71.4885734 54.4626101,73.5396778 Z M54.6063373,74.6087301 C54.6649728,75.0261907 54.7265208,75.4432945 54.7910411,75.860039 L48.8644943,75.860039 C49.3454748,75.5030648 49.8287905,75.0886543 50.2926012,74.6087301 L54.6063375,74.6087301 Z M51.5790861,93.5504758 C51.5582912,94.1599692 51.5089518,94.7523002 51.4230792,95.3090598 C49.5987538,92.8527103 46.6781606,91.2579581 43.3936875,91.2579581 C42.8176577,91.2579581 42.2533244,91.3069223 41.7042542,91.4009698 L40.9531649,90.2737139 C41.7392341,90.098674 42.5561703,90.0066492 43.3936875,90.0066492 C46.6169587,90.0066492 49.5266645,91.3702306 51.5790861,93.5504758 Z M58.117157,37.1517033 C58.1658012,37.0048066 58.2129376,36.8554192 58.2585,36.7033927 C58.3399363,36.431606 58.4173404,36.1587487 58.4908512,35.8849118 C60.146518,37.1012504 62.1880845,37.8204361 64.39375,37.8204361 C69.90625,37.8204361 74.39375,33.3282372 74.39375,27.8099649 C74.39375,25.8829492 73.85,24.0935775 72.90625,22.5669806 C72.7125,22.3480016 72.59375,22.0602005 72.59375,21.7348602 C72.59375,21.0466403 73.15,20.4835513 73.84375,20.4835513 L73.85625,20.4835513 L84.1771759,20.4835513 C84.1399295,20.6955441 84.1111714,20.918096 84.091625,21.1527513 C84.0739507,21.365096 84.0803256,21.5586867 84.1084108,21.7348602 L73.8625,21.7348602 C74.0875,22.0852267 74.29375,22.4418497 74.475,22.8172424 C75.225,24.3188131 75.64375,26.0143366 75.64375,27.8099649 C75.64375,34.0227136 70.6,39.071745 64.39375,39.071745 C62.072023,39.071745 59.9115597,38.3637226 58.117157,37.1517033 Z M76.1869172,1.67351483 C76.3584325,1.30434739 76.5387389,0.908671604 76.7284371,0.485677119 C78.4170275,2.00012584 80.6209115,2.94984752 83.0400634,3.0345022 C83.1217616,3.50869303 83.2231923,3.93414877 83.3505169,4.29190807 C80.6270463,4.28162909 78.1295653,3.29910214 76.1869172,1.67351483 Z M88.2124375,4.48956904 C88.6585663,4.23883943 89.081677,3.99416498 89.4624375,3.81258243 L89.4624375,12.2176118 C89.0679297,13.0566784 88.64361,13.756358 88.2124375,14.36565 L88.2124375,4.48956904 Z M88.2124375,23.2604124 C88.6290796,23.2141682 89.0504566,23.1486039 89.4624375,23.0717966 L89.4624375,27.6407276 C89.0842897,27.8949044 88.6688245,28.1524283 88.2124375,28.4120571 L88.2124375,23.2604124 Z M69.0378962,58.212793 C68.6488103,58.0421252 68.2482743,57.892875 67.8377825,57.7665381 C68.4218197,58.1128387 68.800276,58.2321317 69.0378962,58.212793 Z" id="path-23"></path>
-					<filter x="-0.3%" y="-1.3%" width="101.4%" height="101.8%" filterUnits="objectBoundingBox" id="filter-24">
-						<feOffset dx="1" dy="-1" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feComposite in="shadowOffsetOuter1" in2="SourceAlpha" operator="out" result="shadowOffsetOuter1"></feComposite>
-						<feColorMatrix values="0 0 0 0 1   0 0 0 0 1   0 0 0 0 1  0 0 0 0.1 0" type="matrix" in="shadowOffsetOuter1"></feColorMatrix>
-					</filter>
-					<path d="M26.7108906,173.681675 L26.54375,173.768892 C26.3625,174.144284 26.20625,174.53219 26.06875,174.926352 C25.66875,176.0713 25.45625,177.310096 25.45625,178.592687 C25.45625,184.805436 30.50625,189.854468 36.70625,189.854468 C42.9125,189.854468 47.95625,184.805436 47.95625,178.592687 C47.95625,177.310096 47.74375,176.0713 47.34375,174.926352 C47.20625,174.53219 47.05,174.144284 46.86875,173.768892 L46.7261011,173.681675 L61.875,173.681675 L61.875,190.77241 C62.0932632,191.13765 62.4929804,191.38144 62.95,191.38144 C63.15625,191.38144 63.35625,191.331387 63.525,191.237539 C63.54375,191.231283 63.55625,191.225026 63.575,191.212513 C64.9375,190.505524 66.4875,190.111361 68.125,190.111361 C73.6375,190.111361 78.125,194.60356 78.125,200.121832 C78.125,205.640105 73.6375,210.132304 68.125,210.132304 C66.475,210.132304 64.91875,209.725628 63.54375,209.018639 C63.53125,209.012382 63.5125,209.006126 63.5,208.999869 C63.33125,208.912277 63.15,208.868482 62.95,208.868482 C62.4903634,208.868482 62.0921859,209.114576 61.875,209.479628 L61.875,227.487958 L10.625,227.487958 L10.625,173.681675 L26.7108906,173.681675 Z" id="path-25"></path>
-					<filter x="-4.4%" y="-5.6%" width="108.9%" height="111.2%" filterUnits="objectBoundingBox" id="filter-26">
-						<feGaussianBlur stdDeviation="1.5" in="SourceAlpha" result="shadowBlurInner1"></feGaussianBlur>
-						<feOffset dx="-3" dy="3" in="shadowBlurInner1" result="shadowOffsetInner1"></feOffset>
-						<feComposite in="shadowOffsetInner1" in2="SourceAlpha" operator="arithmetic" k2="-1" k3="1" result="shadowInnerInner1"></feComposite>
-						<feColorMatrix values="0 0 0 0 0.803921569   0 0 0 0 0.835294118   0 0 0 0 0.862745098  0 0 0 1 0" type="matrix" in="shadowInnerInner1"></feColorMatrix>
-					</filter>
-					<path d="M164.9,82.8428781 L164.9,67.4080105 L180.519354,67.4080105 L180.525,67.4013785 C181.19375,67.3763524 181.725,66.8257764 181.725,66.1500696 C181.725,65.9436037 181.675,65.7433942 181.58125,65.5744675 C181.575,65.5556979 181.56875,65.5431848 181.55625,65.5244152 C180.85,64.1604885 180.45625,62.6088654 180.45625,60.9696508 C180.45625,55.4513785 184.94375,50.9591796 190.45625,50.9591796 C195.96875,50.9591796 200.45625,55.4513785 200.45625,60.9696508 C200.45625,62.6213785 200.05,64.1792581 199.34375,65.5556979 C199.3375,65.568211 199.33125,65.5869806 199.325,65.5994937 C199.2375,65.7684204 199.19375,65.9498602 199.19375,66.1500696 C199.19375,66.8382895 199.74375,67.3888654 200.41875,67.4013785 L200.425,67.4076351 L200.424439,67.4080105 L214.0125,67.4080105 L214.0125,82.8429051 C213.631187,82.6739791 213.237437,82.523822 212.837437,82.3986911 C211.768687,82.0545812 210.637437,81.8731414 209.456187,81.8731414 C203.249937,81.8731414 198.206187,86.9284293 198.206187,93.1349215 C198.206187,99.3476702 203.249937,104.396702 209.456187,104.396702 C210.637437,104.396702 211.774937,104.215262 212.843687,103.871152 C213.243687,103.746021 213.631187,103.595864 214.012437,103.426937 L214.0125,120.275812 L200.556598,120.275812 L200.565625,120.281505 C199.884375,120.300275 199.334375,120.850851 199.334375,121.53907 C199.334375,121.73928 199.378125,121.92072 199.465625,122.089646 C199.471875,122.102159 199.478125,122.120929 199.484375,122.133442 C200.190625,123.509882 200.596875,125.067762 200.596875,126.719489 C200.596875,132.237762 196.109375,136.72996 190.596875,136.72996 C185.084375,136.72996 180.596875,132.237762 180.596875,126.719489 C180.596875,125.080275 180.990625,123.528652 181.696875,122.164725 C181.709375,122.145955 181.715625,122.133442 181.721875,122.114673 C181.815625,121.945746 181.865625,121.745536 181.865625,121.53907 C181.865625,120.863364 181.334375,120.312788 180.665625,120.287762 C180.672436,120.283776 180.679243,120.279793 180.686047,120.275812 L164.9,120.275812 L164.9,103.426965 C165.28123,103.59588 165.668709,103.746028 166.068687,103.871152 C167.137437,104.215262 168.274937,104.396702 169.456187,104.396702 C175.662437,104.396702 180.706187,99.3476702 180.706187,93.1349215 C180.706187,86.9284293 175.662437,81.8731414 169.456187,81.8731414 C168.274937,81.8731414 167.143687,82.0545812 166.074937,82.3986911 C165.674959,82.5238152 165.281231,82.6739626 164.9,82.8428781 Z" id="path-27"></path>
-					<filter x="-6.1%" y="-3.5%" width="112.2%" height="107.0%" filterUnits="objectBoundingBox" id="filter-28">
-						<feGaussianBlur stdDeviation="1.5" in="SourceAlpha" result="shadowBlurInner1"></feGaussianBlur>
-						<feOffset dx="-3" dy="3" in="shadowBlurInner1" result="shadowOffsetInner1"></feOffset>
-						<feComposite in="shadowOffsetInner1" in2="SourceAlpha" operator="arithmetic" k2="-1" k3="1" result="shadowInnerInner1"></feComposite>
-						<feColorMatrix values="0 0 0 0 0.803921569   0 0 0 0 0.835294118   0 0 0 0 0.862745098  0 0 0 1 0" type="matrix" in="shadowInnerInner1"></feColorMatrix>
-					</filter>
-					<path d="M282.90506,121.502094 C282.766794,122.114895 282.664065,122.504699 282.596875,122.671505 C282.109375,123.92907 281.846875,125.292997 281.846875,126.719489 C281.846875,132.932238 286.896875,137.981269 293.096875,137.981269 C299.303125,137.981269 304.346875,132.932238 304.346875,126.719489 C304.346875,125.292997 304.084375,123.92907 303.596875,122.677762 C303.446875,122.283599 303.271875,121.90195 303.078125,121.532814 L303.077818,121.502094 L317.75,121.502094 L317.75,137.368558 C317.754144,137.362349 317.75829,137.356137 317.762437,137.349921 C317.756188,137.381204 317.756188,137.418743 317.756188,137.456283 C317.756188,138.150759 318.318688,138.707592 319.006188,138.707592 C319.256188,138.707592 319.487437,138.63877 319.681187,138.507382 C319.693688,138.507382 319.706187,138.507382 319.712438,138.501126 C321.268688,137.506335 323.106187,136.930733 325.081188,136.930733 C330.593687,136.930733 335.081187,141.422932 335.081187,146.941204 C335.081187,152.459476 330.593687,156.951675 325.081188,156.951675 C323.156187,156.951675 321.368687,156.407356 319.843687,155.462618 C319.624937,155.268665 319.337438,155.149791 319.012437,155.149791 C318.324937,155.149791 317.762437,155.706623 317.762437,156.401099 L317.762437,156.413613 L317.75,156.393357 L317.75,172.530471 L302.312371,172.530471 L302.3125,172.517583 C302.50625,172.148447 302.68125,171.766797 302.83125,171.372635 C303.31875,170.121326 303.58125,168.757399 303.58125,167.330907 C303.58125,161.118159 298.5375,156.069127 292.33125,156.069127 C286.13125,156.069127 281.08125,161.118159 281.08125,167.330907 C281.08125,168.757399 281.34375,170.121326 281.83125,171.378892 C281.9875,171.766797 282.15625,172.148447 282.35,172.517583 L282.350518,172.530471 L267.375,172.530471 L267.375,156.150768 C267.259025,155.577521 266.752491,155.149791 266.149937,155.149791 C265.824937,155.149791 265.537437,155.268665 265.318687,155.462618 C263.793687,156.407356 262.006187,156.951675 260.081187,156.951675 C254.568687,156.951675 250.081187,152.459476 250.081187,146.941204 C250.081187,141.422932 254.568687,136.930733 260.081187,136.930733 C262.056187,136.930733 263.893687,137.506335 265.449937,138.501126 C265.456187,138.507382 265.468687,138.507382 265.481187,138.507382 C265.674937,138.63877 265.906187,138.707592 266.156187,138.707592 C266.7485,138.707592 267.24803,138.294277 267.375,137.735714 L267.375,121.502094 L282.90506,121.502094 Z" id="path-29"></path>
-					<filter x="-3.5%" y="-5.9%" width="107.1%" height="111.8%" filterUnits="objectBoundingBox" id="filter-30">
-						<feGaussianBlur stdDeviation="1.5" in="SourceAlpha" result="shadowBlurInner1"></feGaussianBlur>
-						<feOffset dx="-3" dy="3" in="shadowBlurInner1" result="shadowOffsetInner1"></feOffset>
-						<feComposite in="shadowOffsetInner1" in2="SourceAlpha" operator="arithmetic" k2="-1" k3="1" result="shadowInnerInner1"></feComposite>
-						<feColorMatrix values="0 0 0 0 0.803921569   0 0 0 0 0.835294118   0 0 0 0 0.862745098  0 0 0 1 0" type="matrix" in="shadowInnerInner1"></feColorMatrix>
-					</filter>
-					<filter x="-21.8%" y="-16.9%" width="143.6%" height="133.8%" filterUnits="objectBoundingBox" id="filter-31">
-						<feOffset dx="-3" dy="3" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feGaussianBlur stdDeviation="1.5" in="shadowOffsetOuter1" result="shadowBlurOuter1"></feGaussianBlur>
-						<feColorMatrix values="0 0 0 0 0.803921569   0 0 0 0 0.835294118   0 0 0 0 0.862745098  0 0 0 1 0" type="matrix" in="shadowBlurOuter1" result="shadowMatrixOuter1"></feColorMatrix>
-						<feMerge>
-							<feMergeNode in="shadowMatrixOuter1"></feMergeNode>
-							<feMergeNode in="SourceGraphic"></feMergeNode>
-						</feMerge>
-					</filter>
-					<filter x="-26.0%" y="-15.1%" width="150.0%" height="130.2%" filterUnits="objectBoundingBox" id="filter-32">
-						<feOffset dx="-3" dy="3" in="SourceAlpha" result="shadowOffsetOuter1"></feOffset>
-						<feGaussianBlur stdDeviation="1.5" in="shadowOffsetOuter1" result="shadowBlurOuter1"></feGaussianBlur>
-						<feColorMatrix values="0 0 0 0 0.803921569   0 0 0 0 0.835294118   0 0 0 0 0.862745098  0 0 0 1 0" type="matrix" in="shadowBlurOuter1" result="shadowMatrixOuter1"></feColorMatrix>
-						<feMerge>
-							<feMergeNode in="shadowMatrixOuter1"></feMergeNode>
-							<feMergeNode in="SourceGraphic"></feMergeNode>
-						</feMerge>
-					</filter>
-				</defs>
-				<g id="Illustration---Moose-Puzzle---Incomplete" stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
-					<rect fill="#FFFFFF" x="0" y="0" width="450" height="320"></rect>
-					<g id="Group" transform="translate(-3.000000, 0.000000)">
-						<g id="Puzzle" transform="translate(204.257789, 147.127856) rotate(-6.000000) translate(-204.257789, -147.127856) translate(11.757789, 27.627856)">
-							<g id="Combined-Shape" fill="black" fill-opacity="1">
-								<use filter="url(#filter-2)" xlink:href="#path-1"></use>
-							</g>
-							<rect id="Border" stroke="#B5BDC2" stroke-width="1.25" fill="#F1F5FB" x="12.5" y="13.7637435" width="360" height="212.723822" rx="13.75"></rect>
-							<g id="Moose">
-								<g id="Background" transform="translate(20.220000, 23.723941)">
-									<g id="Far-Trees" transform="translate(3.994677, 107.790189)" fill="#8CCD5A" fill-opacity="0.4">
-										<path d="M108.156604,75.9316269 C86.6218246,76.307201 60.3907211,76.7646728 60.3885104,76.7646728 C60.3803854,76.7646728 18.2347604,75.0910471 18.2347604,75.0910471 C18.0753854,75.0847906 17.9241354,75.0178455 17.8128854,74.9039764 L0.17851044,56.8707382 C0.0247604398,56.713699 -0.0352395602,56.487212 0.0203854398,56.2738639 L11.7885104,11.4657435 C11.8603854,11.1904555 12.1091354,10.9990052 12.3928854,10.9990052 L12.4010104,10.9990052 C12.6878854,11.0027592 12.9353854,11.2010916 13.0010104,11.4807592 L13.9116354,15.3379188 L16.6091354,3.18896073 C16.6722604,2.90428796 16.9235104,2.70095026 17.2147604,2.69844764 L17.2191354,2.69844764 C17.5085104,2.69844764 17.7603854,2.89740576 17.8272604,3.17957592 L21.4497604,18.3723429 L25.3703854,0.491764398 C25.4328854,0.206465969 25.6835104,0.0025026178 25.9753854,-2.84217094e-14 L25.9810104,-2.84217094e-14 C26.2697604,-2.84217094e-14 26.5216354,0.198332461 26.5891354,0.480502618 L31.5641354,21.3485812 L34.8066354,6.56374084 C34.8660104,6.29158115 35.0985104,6.09137173 35.3766354,6.07322775 C35.3897604,6.07260209 35.4035104,6.07260209 35.4172604,6.07260209 C35.6791354,6.07260209 35.9153854,6.23652356 36.0047604,6.48615969 L43.5710104,27.5256675 L45.8703854,17.1704607 C45.9335104,16.885788 46.1847604,16.6824503 46.4760104,16.6805733 L46.4803854,16.6805733 C46.7697604,16.6805733 47.0216354,16.8789058 47.0885104,17.1610759 L50.7110104,32.3538429 L54.6316354,14.4732644 C54.6941354,14.187966 54.9447604,13.9846283 55.2366354,13.9815 L55.2422604,13.9815 C55.5310104,13.9815 55.7828854,14.1804581 55.8503854,14.4620026 L60.8260104,35.3300812 L64.0678854,20.5452408 C64.1297604,20.2605681 64.3803854,20.0566047 64.6722604,20.0541021 L64.6778854,20.0541021 C64.9666354,20.0541021 65.2185104,20.2524346 65.2860104,20.5346047 L69.2185104,37.0299843 L70.0022604,33.4143272 C70.0635104,33.1290288 70.3147604,32.9244398 70.6066354,32.9213115 L70.6128854,32.9213115 C70.9022604,32.9213115 71.1541354,33.1202696 71.2210104,33.4030654 L72.1310104,37.2602251 L74.8291354,25.111267 C74.8922604,24.8265942 75.1428854,24.6232565 75.4341354,24.6213796 L75.4391354,24.6213796 C75.7278854,24.6213796 75.9797604,24.819712 76.0472604,25.1018822 L79.6691354,40.2946492 L83.5903854,22.4140707 C83.6522604,22.1293979 83.9035104,21.9254346 84.1947604,21.9229319 L84.2003854,21.9229319 C84.4897604,21.9229319 84.7416354,22.1212644 84.8085104,22.4034346 L89.7841354,43.2708874 L93.0260104,28.4860471 C93.0885104,28.2013743 93.3391354,27.997411 93.6303854,27.9949084 L93.6366354,27.9949084 C93.9253854,27.9949084 94.1772604,28.1932408 94.2441354,28.475411 L97.0560104,40.2677461 L99.1435104,30.6301649 C99.2053854,30.3442408 99.4566354,30.1396518 99.7485104,30.1365236 L99.7541354,30.1365236 C100.044135,30.1365236 100.29601,30.3361073 100.36226,30.6182775 L101.27226,34.4754372 L103.970385,22.3264791 C104.03351,22.0418063 104.28476,21.8384686 104.57601,21.8365916 L104.580385,21.8365916 C104.86976,21.8365916 105.121635,22.0349241 105.18851,22.3170942 L108.81101,37.5104869 L112.731635,19.6292827 C112.794135,19.3446099 113.04476,19.1406466 113.336635,19.138144 L113.34226,19.138144 C113.63101,19.138144 113.882885,19.3364764 113.950385,19.6186466 L118.925385,40.4860995 L122.167885,25.7018848 C122.22976,25.4197147 122.47601,25.2170026 122.764135,25.2101204 L122.77851,25.2101204 C123.06101,25.2101204 123.309135,25.4003194 123.38226,25.674356 L127.803521,42.23195 L130.659323,30.2548974 C130.726198,29.9727272 130.978073,29.7743948 131.266823,29.7743948 L131.272448,29.7743948 C131.564323,29.7768974 131.814948,29.9808607 131.877448,30.2655335 L135.798073,48.146112 L139.420573,32.953345 C139.488073,32.6711749 139.739323,32.4728424 140.028698,32.4728424 L140.033698,32.4728424 C140.324323,32.4747194 140.575573,32.6780571 140.638698,32.9627298 L143.336823,45.1110623 L144.246823,41.2545283 C144.313073,40.9723581 144.565573,40.7727743 144.854948,40.7727743 L144.860573,40.7727743 C145.152448,40.7759026 145.403698,40.9804916 145.465573,41.2657901 L147.553073,50.9033712 L150.364323,39.1116618 C150.431823,38.8294916 150.683698,38.6305335 150.972448,38.6305335 L150.978073,38.6305335 C151.269323,38.6336618 151.520573,38.8376251 151.583073,39.1222979 L154.824948,53.9071382 L159.800573,33.0390597 C159.867448,32.7575152 160.119323,32.5585571 160.408698,32.5585571 L160.414323,32.5585571 C160.705573,32.5616853 160.956823,32.765023 161.018698,33.0503215 L164.939948,50.9309 L168.562448,35.738133 C168.629323,35.4559628 168.881198,35.2576304 169.169948,35.2576304 L169.174948,35.2576304 C169.466198,35.2595073 169.716823,35.462845 169.779948,35.7475178 L172.478073,47.8964759 L173.388073,44.0393162 C173.454948,43.7565204 173.706823,43.5575623 173.996198,43.5575623 L174.002448,43.5575623 C174.294323,43.5600649 174.545573,43.7652796 174.606823,44.050578 L175.389948,47.6656094 L179.323073,31.1708555 C179.389948,30.8886853 179.641823,30.6903529 179.931198,30.6903529 L179.936823,30.6903529 C180.228073,30.6928555 180.478698,30.8968188 180.541198,31.1814916 L183.783698,45.9663319 L188.759323,25.0982534 C188.826198,24.8160832 189.078073,24.6177508 189.366823,24.6177508 L189.372448,24.6177508 C189.664323,24.6202534 189.914948,24.8242168 189.977448,25.1095152 L193.898073,42.9900937 L197.520573,27.7973267 C197.587448,27.5151565 197.839323,27.3161984 198.128698,27.3161984 L198.133698,27.3161984 C198.424323,27.318701 198.675573,27.5220387 198.738698,27.8067115 L200.347771,35.0531515 L192.296885,71.7680712 C192.236885,72.0427335 192.303135,72.3292832 192.478135,72.5495136 C192.65251,72.7691183 192.91751,72.8986288 193.198135,72.9017571 L239.780702,73.406991 L236.778073,77.2133921 C236.670573,77.3497848 236.510573,77.4348738 236.338073,77.4492639 C236.338073,77.4492639 183.973073,81.6987089 183.956198,81.6987089 C183.948698,81.6987089 108.029948,78.8669969 108.029948,78.8669969 C107.833698,78.859489 107.652448,78.7600099 107.539948,78.5985911 C107.428073,78.4371723 107.398073,78.2319576 107.459323,78.0455126 L108.156604,75.9316269 L108.156604,75.9316269 Z" id="Path"></path>
-									</g>
-									<g id="Close-Trees" opacity="0.8" transform="translate(195.454397, -0.000000)" fill="#8CCD5A">
-										<path d="M118.150885,173.822636 C118.114647,173.891786 118.071032,173.957801 118.020291,174.019529 C117.793416,174.296068 117.450291,174.456861 117.087166,174.456861 L117.074666,174.456861 L64.3708209,173.901271 L65.879041,180.227085 C65.945291,180.506752 65.879666,180.801435 65.700291,181.026671 C65.523416,181.248778 65.254666,181.377663 64.971541,181.377663 L64.961541,181.377663 L1.738416,180.691946 C1.457791,180.688818 1.192791,180.559307 1.018416,180.339702 C0.843415998,180.119472 0.777165998,179.832922 0.837165998,179.55826 L31.048416,41.7841448 C31.142166,41.3586997 31.516541,41.0546317 31.951541,41.0508777 L31.959666,41.0508777 C32.390916,41.0508777 32.767166,41.3468123 32.867166,41.7678777 L49.0985844,109.846086 L73.682791,0.912204188 C73.802166,0.383526178 74.281541,0.0050052356 74.837791,0 L74.849041,0 C75.400916,0 75.881541,0.368510471 76.009666,0.891557592 L112.641876,150.17781 L123.086541,102.546454 C123.134041,102.328726 123.325916,102.172938 123.549041,102.171061 L123.553416,102.171061 C123.774041,102.171061 123.966541,102.322469 124.017791,102.53832 L140.930916,173.475022 C140.965291,173.618297 140.931541,173.769079 140.839666,173.8842 C140.749041,173.998069 140.611541,174.064388 140.465916,174.064388 L140.460916,174.064388 L118.150885,173.822636 L118.150885,173.822636 Z" id="Path"></path>
-									</g>
-									<path d="M2.2874375,175.237553 C26.2074375,199.09626 72.2366756,189.729069 97.046,189.488898 C121.915203,189.248146 119.448063,199.964042 150.599313,195.160268 C173.809938,191.581524 202.298688,193.702493 227.333063,194.324393 C297.141188,196.059959 345.607438,189.032391 345.607438,176.653409 C345.607438,174.326356 342.22861,169.284795 319.724995,169.284795 C213.386709,169.284795 -11.0543714,161.930102 2.2874375,175.237553 Z" id="Fill-16" fill="#9EA5A9"></path>
-									<path d="M329.903437,173.211678 C329.903437,177.244568 302.381079,177.641954 282.285726,177.083294 C286.909877,176.072964 292.429008,176.167303 292.429008,174.535596 C292.429008,169.318454 287.420308,164.632039 277.403437,161.820822 C282.352119,160.909575 287.971637,160.392024 293.927369,160.392024 C313.800704,160.392024 329.903437,166.130933 329.903437,173.211678" id="Fill-57" fill="#E3E9F1"></path>
-									<path d="M297.158937,176.43992 C297.158937,181.558869 268.296845,182.063481 247.223395,181.35393 C252.072822,180.071454 257.860687,180.191324 257.860687,178.120673 C257.860687,171.498842 252.607881,165.550473 242.103938,161.982096 C247.293402,160.825934 253.186279,160.169229 259.431416,160.169229 C280.272619,160.169229 297.158937,167.45292 297.158937,176.43992" id="Fill-45" fill="#F9FBFF"></path>
-									<path d="M132.753438,166.538453 C132.753438,171.325961 104.490313,171.797705 83.8540625,171.134511 C88.6028125,169.935131 94.2709375,170.047124 94.2709375,168.110097 C94.2709375,161.916744 89.1271875,156.353425 78.8409375,153.016184 C83.9228125,151.934427 89.6934375,151.320035 95.8090625,151.320035 C116.217188,151.320035 132.753438,158.132786 132.753438,166.538453" id="Fill-47" fill="#CDD5DC"></path>
-									<path d="M111.57025,171.743836 C111.57025,173.418713 104.758375,174.218925 100.62775,174.700679 C86.5265,176.344273 69.162125,171.658747 54.75525,169.495234 C39.73025,171.769488 14.957125,171.099412 14.957125,165.35215 C14.957125,157.121666 24.69025,161.932323 38.78775,161.932323 C44.154625,161.932323 51.020875,157.858061 55.135875,159.516045 C60.984,157.249925 68.482125,155.889752 76.65775,155.889752 C83.959,155.889752 88.5465,157.002791 94.204625,158.520629 C101.487125,160.474548 111.57025,166.7386 111.57025,171.743836" id="Fill-49" fill="#F9FBFF"></path>
-									<path d="M105.4875,178.237441 C105.4875,186.094409 59.999375,187.132996 47.97,181.3532 C43.4525,181.742983 38.6325,181.909407 33.860625,181.852472 C16.639375,181.637873 0,178.478318 0,172.369428 C0,159.71369 15.995625,149.454834 35.73,149.454834 C37.21,149.454834 38.6675,149.511768 40.103125,149.626263 C45.155,147.873179 50.938125,146.878388 57.084375,146.878388 C73.999375,146.878388 88.17,154.415648 91.86875,164.538737 C100.0725,167.466174 105.4875,172.509574 105.4875,178.237441" id="Fill-51" fill="#B5BDC2"></path>
-									<path d="M282.82025,182.730015 C282.82025,184.81845 276.008375,185.816369 271.87775,186.416997 C257.7765,188.467266 240.412125,182.62428 226.00525,179.925832 C210.98025,182.761924 186.207125,181.926675 186.207125,174.758552 C186.207125,164.494065 195.94025,170.494091 210.03775,170.494091 C215.404625,170.494091 222.270875,165.412526 226.385875,167.479688 C232.234,164.654232 239.732125,162.957458 247.90775,162.957458 C255.209,162.957458 259.7965,164.345785 265.454625,166.239015 C272.737125,168.675314 282.82025,176.487861 282.82025,182.730015" id="Fill-53" fill="#868C8F"></path>
-									<path d="M255.754187,187.331391 C242.746062,190.295742 220.102937,190.517223 205.716687,187.963302 C189.426687,190.873846 162.467312,188.655276 162.467312,181.315723 C162.467312,170.803477 174.856062,162.280812 190.144812,162.280812 C193.986687,162.280812 197.651062,162.818875 200.978562,163.806784 C207.232937,158.093933 217.182312,153.895791 228.359812,154.403823 C281.061062,156.797577 274.646687,182.595187 255.754187,187.331391" id="Fill-55" fill="#B5BDC2"></path>
-								</g>
-								<g id="Background-Puzzle-Lines" transform="translate(12.937500, 14.390052)">
-									<g id="Fill-433">
-										<use fill="black" fill-opacity="1" filter="url(#filter-4)" xlink:href="#path-3"></use>
-										<use fill="#B5BDC2" fill-rule="evenodd" xlink:href="#path-3"></use>
-									</g>
-									<g id="Fill-435">
-										<use fill="black" fill-opacity="1" filter="url(#filter-6)" xlink:href="#path-5"></use>
-										<use fill="#B5BDC2" fill-rule="evenodd" xlink:href="#path-5"></use>
-									</g>
-									<g id="Fill-437">
-										<use fill="black" fill-opacity="1" filter="url(#filter-8)" xlink:href="#path-7"></use>
-										<use fill="#B5BDC2" fill-rule="evenodd" xlink:href="#path-7"></use>
-									</g>
-									<g id="Fill-439">
-										<use fill="black" fill-opacity="1" filter="url(#filter-10)" xlink:href="#path-9"></use>
-										<use fill="#B5BDC2" fill-rule="evenodd" xlink:href="#path-9"></use>
-									</g>
-									<g id="Fill-441">
-										<use fill="black" fill-opacity="1" filter="url(#filter-12)" xlink:href="#path-11"></use>
-										<use fill="#B5BDC2" fill-rule="evenodd" xlink:href="#path-11"></use>
-									</g>
-									<g id="Fill-443">
-										<use fill="black" fill-opacity="1" filter="url(#filter-14)" xlink:href="#path-13"></use>
-										<use fill="#B5BDC2" fill-rule="evenodd" xlink:href="#path-13"></use>
-									</g>
-									<g id="Fill-445">
-										<use fill="black" fill-opacity="1" filter="url(#filter-16)" xlink:href="#path-15"></use>
-										<use fill="#B5BDC2" fill-rule="evenodd" xlink:href="#path-15"></use>
-									</g>
-									<g id="Fill-447">
-										<use fill="black" fill-opacity="1" filter="url(#filter-18)" xlink:href="#path-17"></use>
-										<use fill="#B5BDC2" fill-rule="evenodd" xlink:href="#path-17"></use>
-									</g>
-									<g id="Fill-449">
-										<use fill="black" fill-opacity="1" filter="url(#filter-20)" xlink:href="#path-19"></use>
-										<use fill="#B5BDC2" fill-rule="evenodd" xlink:href="#path-19"></use>
-									</g>
-								</g>
-								<g transform="translate(69.429627, 35.910622)">
-									<path d="M42.937123,55.8757015 C43.335873,54.7595339 42.917748,53.4900811 41.886498,52.9069711 C39.753373,51.7007093 35.633373,47.2691989 35.418998,46.9182067 C29.550873,37.3018978 32.680248,24.4578376 35.123373,14.3472617 C35.647748,12.1768664 35.430248,9.60480096 32.878998,10.4569423 C30.591498,11.2202407 28.709623,15.7725025 27.733998,17.6294449 C26.210873,20.5255994 25.502123,17.5318428 25.223373,15.8175496 C24.648373,12.280725 24.865248,8.62565174 24.878373,5.06067269 C24.885248,3.30821457 24.127123,1.65961509 22.257123,2.51050515 C20.607748,3.26066483 20.148998,6.89384023 19.881498,8.33972766 C18.097123,17.9866936 18.810873,22.6315523 13.813998,11.1983428 C5.75024801,-7.25158124 13.100248,17.6294449 9.88774801,19.416314 C6.67524801,21.2025575 -1.53350199,10.8410942 0.250873011,19.7729371 C2.03587301,28.7054057 9.11399801,28.3531622 9.53087301,33.3496387 C9.85399801,37.2286962 5.62462301,35.1840575 3.69712301,35.3166962 C1.31462301,35.481869 1.34649801,37.9725994 2.27837301,39.4234921 C3.68649801,41.6145339 7.62587301,42.6393559 9.88774801,43.354479 C15.920248,45.2614737 17.181498,49.1261413 13.853998,54.297801 C13.000248,55.624814 10.068373,57.9691413 11.575248,59.4982407 C12.947748,60.8903219 17.512748,58.836924 18.810873,58.360801 C29.518373,54.4304397 22.023373,68.7222643 22.023373,68.7222643 C22.023373,68.7222643 19.881498,73.3671229 27.377123,66.5781465 C34.872123,59.7897957 35.942748,68.0077669 35.942748,68.0077669 C35.942748,68.0077669 41.342748,60.339746 42.937123,55.8757015" id="Fill-59" fill="#9EA5A9"></path>
-									<path d="M189.576498,103.254323 C188.864623,106.111687 187.082123,100.037208 187.082123,100.037208 C186.877748,100.937525 186.591498,101.862868 186.231498,102.796344 C184.023373,108.54298 179.182123,114.576792 176.728373,116.115902 C175.191498,117.081912 175.690248,119.840423 174.863998,121.289438 C172.427123,125.570792 171.232748,129.918465 170.513373,134.813585 C170.513373,134.813585 171.608998,152.079771 166.702123,158.072289 C161.803373,164.072941 160.307123,174.003328 160.307123,174.003328 C160.307123,174.003328 153.167748,180.430051 146.036498,179.357054 C138.897123,178.284682 146.036498,171.857959 146.036498,171.857959 L155.310248,158.637255 C168.517748,142.198184 156.741498,125.047119 156.741498,125.047119 L152.455873,122.550132 C152.455873,122.550132 151.932748,122.460038 151.033373,122.321143 C146.862748,121.674216 133.850873,114.265842 130.310248,117.573051 C119.801498,127.372051 105.342748,121.117383 103.919623,121.47776 C102.488373,121.838137 100.705873,132.554347 100.705873,132.554347 C100.705873,132.554347 101.433373,143.794229 99.993998,148.63304 C98.758998,152.750472 99.993998,175.067567 99.993998,175.067567 C87.498373,183.647166 85.355248,175.787695 85.355248,175.787695 C85.355248,175.787695 92.183998,150.949839 90.335873,141.674512 C88.479623,132.399184 92.690873,132.570614 88.822748,130.041093 C84.946498,127.511572 84.946498,124.310098 84.946498,124.310098 L77.864623,120.405389 C72.867748,118.260645 70.724623,113.258538 70.724623,113.258538 C70.724623,113.258538 66.799623,111.113794 66.079623,104.687072 C65.368373,98.2522159 56.495248,93.978996 56.495248,93.978996 C56.495248,93.978996 57.182123,98.4242709 57.312748,101.396129 C57.410873,103.762355 57.787123,103.90125 56.960873,106.856216 C56.527748,108.412218 55.742748,106.545266 54.875873,108.019308 C53.510248,110.336106 53.018998,108.076242 51.857748,108.019308 C50.001498,107.937347 50.165248,109.009718 48.954623,106.856216 C48.341498,105.783844 47.049623,107.184059 46.632748,106.512732 C45.928998,105.391559 47.957123,104.622004 48.022748,101.633253 C48.063373,99.8651531 46.672748,100.438253 46.542123,99.1281321 C46.329623,97.0084148 45.692123,96.0173782 43.958373,97.8924646 C39.312748,102.893946 35.747748,101.10958 35.747748,101.10958 C35.747748,101.10958 29.170873,104.729616 26.668373,97.9350091 C24.173998,91.1485353 26.105248,86.4636347 26.105248,86.4636347 L33.604623,69.6729462 C30.390873,60.7411033 38.601498,59.6687316 38.601498,59.6687316 L40.743998,54.6666243 C31.822123,50.0167604 31.102123,44.3026583 31.102123,44.3026583 C31.822123,42.1579148 37.889623,43.5906635 42.527123,44.6630353 C47.163998,45.735407 47.523373,49.6651426 47.523373,49.6651426 L57.517123,50.0167604 C57.517123,50.0167604 56.330248,44.1731478 66.975248,41.433407 L64.159623,48.3518939 C70.432123,46.4930745 76.212123,44.9001583 82.664623,45.3093363 C90.867123,45.8336347 96.182748,50.6636871 102.496498,55.2810169 C114.215873,63.8524829 129.132748,67.7571923 143.477123,64.907962 C148.849623,63.8443494 153.298998,62.8126452 158.859623,63.4345457 C165.197748,64.1471661 171.290248,66.3494698 176.728373,69.6729462 C196.004623,81.4615274 190.287748,100.397585 189.576498,103.254323" id="Fill-61" fill="#A87B3B"></path>
-									<path d="M46.2763105,83.3754669 C46.7994355,84.3708832 46.2631855,86.4374198 45.2063105,88.6359695 C43.9944355,91.1586083 42.0975605,93.8558046 40.3319355,95.3110769 C39.9788105,95.6013805 39.6313105,95.8428832 39.2950605,96.022446" id="Smile" stroke="#B9C2D1" stroke-width="1.25" stroke-linecap="round"></path>
-									<path d="M144.07923,97.2479779 C140.01394,110.237585 145.716897,117.352869 150.821799,121.648501" id="Flank" stroke="#B5BDC2" stroke-width="1.25" stroke-linecap="round"></path>
-									<path d="M110.66003,109.027299 C111.160655,110.012177 110.647469,112.056836 109.636049,114.232111 C108.476296,116.728043 107.04497,119.449791 105.355283,120.889656 C105.017346,121.176887 104.920314,121.47084 104.598526,121.648501" id="Flank" stroke="#B5BDC2" stroke-width="1.25" stroke-linecap="round"></path>
-									<path d="M48.409498,65.936225 C48.409498,67.4202774 49.611373,68.6234109 51.093248,68.6234109 C52.576373,68.6234109 53.777623,67.4202774 53.777623,65.936225 C53.777623,64.4521727 52.576373,63.2496648 51.093248,63.2496648 C49.611373,63.2496648 48.409498,64.4521727 48.409498,65.936225" id="Fill-65" fill="#494C4E"></path>
-									<path d="M109.592297,113.66735 L103.337265,122.089109 C103.337265,122.089109 118.407615,128.609661 130.257796,117.733168 C130.257796,117.733168 133.212515,114.663046 143.457919,120.033357 C143.457919,120.033357 155.046376,124.155445 153.432207,123.014958 C148.419391,119.472721 144.047735,111.371963 144.040848,111.36167 C141.627107,107.729539 135.236174,109.242641 131.922682,110.166285 C130.774981,110.486061 129.646691,110.859362 128.530297,111.25119 C122.263338,113.452565 116.382703,116.25369 109.592297,113.66735 Z" id="Belly" fill="#B5BDC2"></path>
-									<path d="M59.854498,77.0968994 C59.866998,77.15008 59.876373,77.2038863 59.883248,77.2564413 C59.881998,77.2120198 59.880748,77.1675983 59.879498,77.1231769 C60.130123,81.024758 60.600748,84.8969334 61.338873,88.7378261 C62.077623,92.5824727 62.979498,96.8262868 64.404498,100.476355 C64.404498,100.476355 58.710123,94.6027109 56.563248,94.2861298 C56.563248,94.2861298 58.598248,102.860098 56.962623,106.859907 C56.962623,106.859907 55.350748,108.972742 54.259498,109.256164 C53.359498,109.489533 53.146373,109.54459 52.036998,108.123104 C52.036998,108.123104 49.005748,109.418208 48.520748,106.626538 C48.520748,106.626538 47.357623,106.996926 46.661373,106.707248 C45.965123,106.41757 46.297623,105.00797 46.984498,104.522462 C47.671373,104.036954 48.023248,101.630062 48.023248,101.630062 L46.338248,99.1011664 L45.336373,96.9107502 C49.833248,93.5121952 50.508873,92.6175093 50.508873,87.6886036 C50.508873,87.6886036 58.441373,86.1113287 59.441373,77.4591533 C59.472623,77.186368 59.461998,76.8547711 59.667623,76.653936 C59.758873,76.7809439 59.818248,76.9379832 59.854498,77.0968994" id="Beard" fill="#B5BDC2"></path>
-									<path d="M58.045748,56.5004175 C60.377623,52.9986295 62.964498,50.8094646 64.266373,46.4655457 C65.683248,41.7368494 65.879498,36.684064 65.588248,31.786441 C65.343873,27.6733887 64.408248,23.6591897 63.510123,19.6518729 C63.009498,17.4170353 60.351373,10.243907 62.215748,8.44139651 C64.952623,5.79425253 68.601373,13.7919934 69.411998,15.3336059 C70.972623,18.302962 71.673873,15.0877237 71.938873,13.3915745 C72.466998,10.0155431 72.416373,6.41365043 72.210123,3.02573159 C72.093248,1.10872635 73.090748,-0.771990927 75.028873,0.32415567 C76.572623,1.19694363 76.990748,4.56922112 77.263873,6.04388865 C79.048248,15.6908546 78.334498,20.3357133 83.331373,8.90250384 C91.394498,-9.5480459 86.886998,13.3421478 90.099498,15.1283913 C93.311998,16.9146347 98.678873,8.54525515 96.893873,17.4770981 C95.109498,26.4095667 90.515748,25.9178023 90.099498,30.9149044 C89.742623,35.2025143 99.190748,31.8440012 99.190748,31.8440012 C99.190748,31.8440012 100.693873,36.8429803 87.130748,41.1305902 C73.567623,45.4182002 89.171998,55.9892578 89.171998,55.9892578 C89.171998,55.9892578 89.041998,59.9946976 78.334498,56.064962 C67.626998,52.1346007 75.121998,66.4264253 75.121998,66.4264253 C75.121998,66.4264253 77.263873,71.0712839 69.768248,64.2823075 C62.273248,57.4939567 61.656373,63.7792813 61.656373,63.7792813 C61.656373,63.7792813 55.886373,59.7438101 58.045748,56.5004175" id="Fill-73" fill="#9EA5A9"></path>
-								</g>
-								<g id="Body-Puzzle-Lines" transform="translate(95.437500, 78.206806)">
-									<g id="Combined-Shape">
-										<use fill="black" fill-opacity="1" filter="url(#filter-22)" xlink:href="#path-21"></use>
-										<use fill-opacity="0.7" fill="#6E7376" fill-rule="evenodd" xlink:href="#path-21"></use>
-									</g>
-								</g>
-								<g id="Antler-Puzzle-Lines" transform="translate(75.437500, 45.672775)">
-									<g id="Combined-Shape">
-										<use fill="black" fill-opacity="1" filter="url(#filter-24)" xlink:href="#path-23"></use>
-										<use fill-opacity="0.4" fill="#494C4E" fill-rule="evenodd" xlink:href="#path-23"></use>
-									</g>
-								</g>
-							</g>
-							<path d="M318.75,48.8787528 L319.012438,48.8761257 C319.387438,49.0575654 319.774938,49.2139791 320.168688,49.351623 C321.312438,49.7520419 322.549938,49.9647644 323.831188,49.9647644 C330.037438,49.9647644 335.081188,44.915733 335.081188,38.7029843 C335.081188,32.4964921 330.037438,27.4412042 323.831188,27.4412042 C322.549938,27.4412042 321.312438,27.6539267 320.168688,28.0543455 C319.774938,28.1919895 319.387438,28.3484031 319.012438,28.5298429 L318.75,28.5298429 L318.75,10.0104712 L376.25,10.0104712 L376.25,66.3193717 L353.663401,66.3193717 C353.284965,66.5351115 353.03125,66.9428361 353.03125,67.4076351 C353.03125,67.7329754 353.15,68.0207764 353.34375,68.2397555 C354.2875,69.7663524 354.83125,71.5557241 354.83125,73.4827398 C354.83125,79.001012 350.34375,83.493211 344.83125,83.493211 C339.31875,83.493211 334.83125,79.001012 334.83125,73.4827398 C334.83125,71.5056717 335.40625,69.6662476 336.4,68.1083681 C336.40625,68.1021115 336.40625,68.0895984 336.40625,68.0770853 C336.5375,67.8831325 336.60625,67.6516403 336.60625,67.4013785 C336.60625,66.9410348 336.357376,66.5366764 335.984947,66.3193717 L318.75,66.3193717 L318.75,48.8787528 Z" id="Combined-Shape" fill="#FFFFFF"></path>
-							<g id="Combined-Shape">
-								<use fill="#FFFFFF" fill-rule="evenodd" xlink:href="#path-25"></use>
-								<use fill="black" fill-opacity="1" filter="url(#filter-26)" xlink:href="#path-25"></use>
-							</g>
-							<g id="Combined-Shape">
-								<use fill="#FFFFFF" fill-rule="evenodd" xlink:href="#path-27"></use>
-								<use fill="black" fill-opacity="1" filter="url(#filter-28)" xlink:href="#path-27"></use>
-							</g>
-							<g id="Combined-Shape">
-								<use fill="#FFFFFF" fill-rule="evenodd" xlink:href="#path-29"></use>
-								<use fill="black" fill-opacity="1" filter="url(#filter-30)" xlink:href="#path-29"></use>
-							</g>
-						</g>
-						<g id="Piece-2" filter="url(#filter-31)" transform="translate(393.493902, 275.565699) rotate(60.000000) translate(-393.493902, -275.565699) translate(365.993902, 240.065699)" fill="#F9FBFF" stroke="#B5BDC2" stroke-width="1.25">
-							<path d="M0.00565417948,36.7675679 L0.00529414739,36.7314379 L0.261177938,36.200159 L0.26246356,36.200159 L0.268665261,36.8225128 L0.125182373,36.8239426 C0.0850888262,36.8053496 0.0452479149,36.7865597 0.00565417914,36.7675679 Z M0.625,37.0404524 L0.625,52.9364035 L17.4032869,52.9364035 L17.5488541,53.0209525 C18.1211993,53.3533846 18.48125,53.9655945 18.48125,54.638486 C18.48125,54.9604901 18.4025838,55.2682501 18.2498977,55.5387898 C18.2298181,55.5921279 18.2007027,55.6491701 18.160316,55.7043823 C17.2141232,57.1911164 16.70625,58.9023018 16.70625,60.6921702 C16.70625,65.8396574 20.9134371,70.0320825 26.08125,70.0320825 C31.2490629,70.0320825 35.45625,65.8396574 35.45625,60.6921702 C35.45625,58.9538776 34.9796867,57.2955629 34.0876511,55.8430102 C33.8105034,55.510561 33.65625,55.0921555 33.65625,54.644714 C33.65625,53.9651368 34.0234886,53.3477923 34.60493,53.0178313 L34.748418,52.9364035 L54.375,52.9364035 L54.375,13.75 C54.375,6.50126266 48.4987373,0.625 41.25,0.625 L0.625,0.625 L0.625,15.1009487 C0.815295436,15.02397 1.01114048,14.9505851 1.21300751,14.8803393 C2.43460164,14.4546141 3.74252369,14.2351754 5.0811875,14.2351754 C11.6300213,14.2351754 16.9561875,19.5446892 16.9561875,26.0707018 C16.9561875,32.6008461 11.6321246,37.9062281 5.0811875,37.9062281 C3.74252369,37.9062281 2.43460164,37.6867894 1.2132805,37.2611593 C1.01127161,37.190864 0.815344794,37.1174499 0.625,37.0404533 Z M-2.93698474e-14,15.3989736 L-3.04323001e-14,15.3765499 C0.0404624396,15.3571102 0.081183032,15.3378819 0.122167293,15.3188596 L0.2624375,15.3188596 L0.2624375,15.9438596 L-2.94209102e-14,15.3989736 Z" id="Combined-Shape"></path>
-						</g>
-						<g id="Piece-1" filter="url(#filter-32)" transform="translate(397.298071, 47.853805) rotate(134.000000) translate(-397.298071, -47.853805) translate(372.298071, 4.853805)">
-							<path d="M-1.77635684e-14,31.7580275 L-1.77635684e-14,16.3839971 L15.6193537,16.3839971 L15.625,16.3773913 C16.29375,16.3524638 16.825,15.804058 16.825,15.1310145 C16.825,14.9253623 16.775,14.725942 16.68125,14.5576812 C16.675,14.5389855 16.66875,14.5265217 16.65625,14.5078261 C15.95,13.1492754 15.55625,11.6037681 15.55625,9.97101449 C15.55625,4.47449275 20.04375,-5.32907052e-15 25.55625,-5.32907052e-15 C31.06875,-5.32907052e-15 35.55625,4.47449275 35.55625,9.97101449 C35.55625,11.6162319 35.15,13.167971 34.44375,14.5389855 C34.4375,14.5514493 34.43125,14.5701449 34.425,14.5826087 C34.3375,14.7508696 34.29375,14.9315942 34.29375,15.1310145 C34.29375,15.8165217 34.84375,16.3649275 35.51875,16.3773913 L35.525,16.3836232 L35.5244389,16.3839971 L49.1125,16.3839971 L49.1125,31.7580544 C48.7311875,31.5897942 48.3374375,31.440229 47.9374375,31.3155913 C46.8686875,30.9728377 45.7374375,30.792113 44.5561875,30.792113 C38.3499375,30.792113 33.3061875,35.8274754 33.3061875,42.0095043 C33.3061875,48.1977652 38.3499375,53.2268957 44.5561875,53.2268957 C45.7374375,53.2268957 46.8749375,53.046171 47.9436875,52.7034174 C48.3436875,52.5787797 48.7311875,52.4292145 49.1124375,52.2609536 L49.1125,69.0434174 L35.6565977,69.0434174 L35.665625,69.0490884 C34.984375,69.0677841 34.434375,69.6161899 34.434375,70.3016971 C34.434375,70.5011174 34.478125,70.681842 34.565625,70.8501029 C34.571875,70.8625667 34.578125,70.8812623 34.584375,70.8937261 C35.290625,72.2647406 35.696875,73.8164797 35.696875,75.4616971 C35.696875,80.9582188 31.209375,85.4327116 25.696875,85.4327116 C20.184375,85.4327116 15.696875,80.9582188 15.696875,75.4616971 C15.696875,73.8289435 16.090625,72.2834362 16.796875,70.9248855 C16.809375,70.9061899 16.815625,70.8937261 16.821875,70.8750304 C16.915625,70.7067696 16.965625,70.5073493 16.965625,70.3016971 C16.965625,69.6286536 16.434375,69.0802478 15.765625,69.0553203 C15.772436,69.0513508 15.7792434,69.0473831 15.7860473,69.0434174 L-1.77635684e-14,69.0434174 L-1.77635684e-14,52.2609812 C0.38122985,52.4292308 0.768709358,52.5787865 1.1686875,52.7034174 C2.2374375,53.046171 3.3749375,53.2268957 4.5561875,53.2268957 C10.7624375,53.2268957 15.8061875,48.1977652 15.8061875,42.0095043 C15.8061875,35.8274754 10.7624375,30.792113 4.5561875,30.792113 C3.3749375,30.792113 2.2436875,30.9728377 1.1749375,31.3155913 C0.774959358,31.4402222 0.381230533,31.5897779 -8.6458618e-15,31.7580275 Z" id="Combined-Shape" stroke="#B5BDC2" stroke-width="1.25" fill="#F1F5FB"></path>
-							<path d="M-1.77635684e-14,22.2661705 L-1.77635684e-14,17.7152784 C1.96732686,17.3499411 3.720375,16.7292435 3.720375,16.7292435 C3.720375,16.7292435 4.4853074,19.2631705 2.54685162e-13,22.2661705 Z" id="Combined-Shape" fill="#9EA5A9"></path>
-							<path d="M15.3716176,45.1033686 C22.2071466,48.4470577 29.695761,50.3690473 37.2851369,50.5647597 C39.2470202,52.2247725 41.7857818,53.2268957 44.5561875,53.2268957 C45.7374375,53.2268957 46.8749375,53.046171 47.9436875,52.7034174 C48.3436875,52.5787797 48.7311875,52.4292145 49.1124375,52.2609536 L49.1125,69.0434174 L35.6565977,69.0434174 L35.665625,69.0490884 C34.984375,69.0677841 34.434375,69.6161899 34.434375,70.3016971 C34.434375,70.5011174 34.478125,70.681842 34.565625,70.8501029 C34.571875,70.8625667 34.578125,70.8812623 34.584375,70.8937261 C35.290625,72.2647406 35.696875,73.8164797 35.696875,75.4616971 C35.696875,80.9582188 31.209375,85.4327116 25.696875,85.4327116 C20.184375,85.4327116 15.696875,80.9582188 15.696875,75.4616971 C15.696875,73.8289435 16.090625,72.2834362 16.796875,70.9248855 C16.809375,70.9061899 16.815625,70.8937261 16.821875,70.8750304 C16.915625,70.7067696 16.965625,70.5073493 16.965625,70.3016971 C16.965625,69.6286536 16.434375,69.0802478 15.765625,69.0553203 C15.772436,69.0513508 15.7792434,69.0473831 15.7860473,69.0434174 L-1.77635684e-14,69.0434174 L-1.77635684e-14,52.2609812 C0.38122985,52.4292308 0.768709358,52.5787865 1.1686875,52.7034174 C2.2374375,53.046171 3.3749375,53.2268957 4.5561875,53.2268957 C9.68646907,53.2268957 14.0223916,49.7903923 15.3716176,45.1033686 Z" id="Combined-Shape" fill="#A87B3B"></path>
-						</g>
-					</g>
-				</g>
-			</svg>
-		</template>
-`;
-	}
-
-	static get is() {
-		return 'd2l-end-of-lesson-image';
-	}
-	static get properties() {
-		return {
-			completed: {
-				type: Boolean,
-				value: false
-			}
-		};
-	}
-}
-customElements.define(D2LEndOfLessonImage.is, D2LEndOfLessonImage);
