+++ components/d2l-sequences-content-eol-main.js
@@ -5,8 +5,7 @@ import './d2l-sequences-content-eol-missed.js';
 import '@polymer/polymer/polymer-legacy.js';
 import 'd2l-polymer-siren-behaviors/store/entity-behavior.js';
 import '../localize-behavior.js';
-import './d2l-end-of-lesson-image.js';
-import './d2l-end-of-lesson-progress.js';
+import './d2l-sequences-end-of-lesson-progress.js';
 import '../mixins/d2l-sequences-return-mixin.js';
 import 's-html/s-html.js';
 import { D2LPoller } from 'd2l-poller/d2l-poller.js';
