+++ demo/d2l-sequence-navigator/index.html
@@ -15,7 +15,7 @@
 		import '../../d2l-sequence-navigator/d2l-sequence-end.js';
 		import '../../d2l-sequence-navigator/d2l-lesson-header.js';
 		import '../../d2l-sequence-navigator/d2l-missed-activity.js';
-		import '../../d2l-sequence-launcher-unit/d2l-sequence-launcher-unit.js';
+		import '../../d2l-sequence-navigator/d2l-sequence-navigator.js';
 		window.D2L.Siren.WhitelistBehavior._testMode(true);
 	</script>
