+++ test/d2l-sequence-viewer/sequence-viewer-header.html
@@ -34,7 +34,7 @@
 		</test-fixture>
 
 		<script type="module">
-			import '../../d2l-sequence-viewer/sequence-viewer-header.js';
+			import '../../d2l-sequence-viewer/d2l-sequence-viewer-header.js';
 			import TelemetryHelper from '../../helpers/telemetry-helper';
 			/* global chai */
 			// TODO: This test suite is missing a lot of test cases.
